CHROMIUM: uapi/fuse.h: Add message definitions for CHROMEOS_TMPFILE
This is a chrome os extension for fuse implemented in a way that will
hopefully not conflict with rebases.  The upstream maintainer was not
happy with this approach because it uses 2 round-trips instead of one.
However, his proposed mechanism requires changes to the VFS layer but
the VFS maintainer was non-committal.

At this point, I think the patch is non-intrusive enough that we can
just maintain it ourselves and if upstream ever gains support for
O_TMPFILE we can drop it at that time.

BUG=b:160932094
TEST=Successfully open a file with  O_TMPFILE on virtio-fs

Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2520739
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>
Change-Id: I241c0379ac2aca9a895a6d373a30fe8ba0e07397
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2992592
Commit-Queue: Hikaru Nishida <hikalium@chromium.org>
Tested-by: Hikaru Nishida <hikalium@chromium.org>
Signed-off-by: Richard Fung <richardfung@google.com>