CHROMIUM: fuse: Implement CHROMEOS_TMPFILE
This is a chrome os extension for fuse implemented in a way that will
hopefully not conflict with rebases.  The upstream maintainer was not
happy with this approach because it uses 2 round-trips instead of one.
However, his proposed mechanism requires changes to the VFS layer but
the VFS maintainer was non-committal.

At this point, I think the patch is non-intrusive enough that we can
just maintain it ourselves and if upstream ever gains support for
O_TMPFILE we can drop it at that time.

BUG=b:160932094
TEST=Successfully open a file with  O_TMPFILE on virtio-fs

Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2520740
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>
Conflicts: fs/fuse/file.c
[rebase510(groeck):
    no init_security in v5.10
    1st argument to create_new_entry() changed
]
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Change-Id: I3f5b305e909a54169915a80d81ebe97e1874c26a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2992593
Commit-Queue: Hikaru Nishida <hikalium@chromium.org>
Tested-by: Hikaru Nishida <hikalium@chromium.org>
Signed-off-by: Richard Fung <richardfung@google.com>