Revert "build: introduce DISABLE_LTO_CFI_CONFIGS option"
This feature enabled work on the kernel to be done with more efficient
incremental builds (less time consuming). However, the way this version
is implemented trades off disabling CFI (which depends on LTO) which is
probably not a good trade-off in most cases.

Most of the incremental performance can be reclaimed with LTO_THIN,
which is available on newer kernels, but crucially it does not require
CFI to be disabled.

A follow-up change will modify this functionality to be equivalent to
"LTO=none" but favor the use of "LTO=thin" to allow CFI to stay enabled.

This reverts commit 51918f884da73464198de81a56833c6a6845d275.

Bug: 167259389
Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: I52a9e3795df777b12e994ec2425e9f38061c14e0