symtab reader: ignore 'public' property of exported kernel symbols
Kernel symbols might be exported through ksymtab entries. At the same
time it is ok for those symbols to have LOCAL binding and HIDDEN
visibility. Either of those properties flags the kernel symbol as not
public. When reading the ksymtab and when walking through a kernel
symbol table, we can entirely rely on the property of being exported by
the ksymtab. Hence, do that and ignore any ELF visibility or binding.

Reported-by: Elliot Berman <eberman@codeaurora.org>
Bug: 182194803
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: If311743bf0e9b9ff83e8a127e8019574ebfaaf74