build.sh: Avoid unnecessary touching of the kernel config by build.sh
Touching the kernel config (without any modification), triggers rules to
relink the kernel. That becomes relevant when running with
SKIP_MRPROPER=1 and in a mixed build scenarios where one of the kernels
is not actually modified.

In order to overcome this, we store a copy of .config right before any
modification to it is potentially be done. Right before the actual
kernle build we restore the old copy if there hasn't been any
modification to the .config. Along with that we restore the modification
timestamp that effectively allows the kernel build to skip relinking the
kernel if the sole reason for that was the configuration change itself.

Bug: 179648610
Reported-by: Elliot Berman <eberman@quicinc.com>
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Ieb794b99a66f116917cbf6eb28e732e0d43b309b