ANDROID: Make KMI_SYMBOL_LIST_STRICT_MODE overrideable
For development purposes, it can be useful to disable strict symbol list
checking, but this value is currently always being set to 1 in the arm64
config. Allow for this value to be overridden, with the default of being
enabled.

Bug: 184667897
Signed-off-by: J. Avila <elavila@google.com>
Change-Id: Ie6295a2cb9d89b30e2242e2d55639a2c4882a897
(cherry picked from commit 308e49bc5f4b789a5049a38065ad66e10c992d69)