extract_symbols: Add tracepoint dependencies to the symbol list
The core tracepoint API (on which relies GKI for vendor hooks) casts the
callback pointers to void * when a module tries to attach. This means
that changes to vendor hooks are not detected by our ABI monitoring,
which is likely to cause problems in the future.

To workaround the issue, make sure to always add to the symbol list the
__traceiter_*() function of each tracepoint as it has full type
definition.

The alternative to this would be to create artificial dependencies on
the kernel side, but this would require nasty changes to the core
tracepoint code. By implementing it in extract_symbols, this solves the
issue for 5.10 and later without any kernel change, and should simplify
the changes needed in 5.4 and prior.

Bug: 185777726
Signed-off-by: Quentin Perret <qperret@google.com>
Change-Id: Ib428394bb7eb5ad57318ff07625597e6a8977e70