Add mkbootfs to kernel build-tools
So that we can pack/repack mkbootfs data in the kernel tree with
`mkbootfs` instead of `cpio`.
Using `mkbootfs` in the kernel tree has many benefits, one is `mkbootfs`
produces stable output by re-ordering inode numbers and sorts the
archive entries, the other is that this makes the kernel tree behaves
more like the platform tree, so that's one less "platform and kernel
tree differences" to worry about (less mental burden).

Bug: 184490452
Test: Presubmit
Test: ./prebuilts/kernel-build-tools/build-prebuilts.sh
Change-Id: I87edb989d5a8666e626e75e690b6a6566dd5c8a5