ANDROID: Add export alloc_contig_range/free_contig_range
The alloc_contig_range/free_contig_range have been export in upstream:
(virtio-mem: Paravirtualized memory hotunplug part 2 "255f5985070839059")

But this commit is one of a series of function improvement for virtio
and we need these two symbols only.

Thus backport the export symbol only in this commit.

Because this is new export symbol in this kernel version so make it GPL.

(cherry picked from commit 255f598507083905995ecab96392770ae03aac7f)
Bug: 187366010
Signed-off-by: Lecopzer Chen <lecopzer.chen@mediatek.com>
Change-Id: Ic7b44e8ee56e4f93a7a639829eb067c1bfc4dd57