ANDROID: gki - CONFIG_NET_SCH_{CODEL,FQ_CODEL,SFQ}=y
This enables the codel, fq_codel and sfq qdiscs.
The netem qdisc is already recommended in Android R,
and appears to have been left out of the gki configs by mistake,
and was just recently enabled - this enables a few
more qdiscs in a similar vein.

These qdiscs are very useful for testing (netem) and/or
significantly better then the defaults (pfifo_fast)
for multi-flow configurations/benchmarks.

Test: built and booted on a gki using phone
Bug: 124467469
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I34acd453cfa164efb220c658868823f77d6ae8aa