build.sh: Add support for building vendor_boot images alone
For vendors developing with GKI using prebuilts, they can use the
downloaded boot images instead of needing to build them. However, in
order to test any changes to modules that would be loaded in the
vendor_boot partition, they'd need a mechanism to build vendor_boot
when BUILD_BOOT_IMG is unset.

Signed-off-by: J. Avila <elavila@google.com>
Change-Id: Ib150ef5e6f908353ebe8297c5ff85ec3750ec7fa