Kleaf: add initial bazel build environment
Initialize a minimum bazel build environment.

The bazel.WORKSPACE file will be symbolically linked to toplevel
WORKSPACE via repo. It marks the top of a bazel build environment and
has some initial settings.

bazel.sh is a wrapper to ensure the correct bazel version and the
correct runtime is used. It will enable the default configuration common
for all kernel bazel builds. In the full repo checkout, this script will
be made accessible via tools/bazel. It can either be called directly,
relative or absolute, or it can be invoked by simply calling the host
bazel. The builtin logic will use tools/bazel as long as we are within a
subdirecty of WORKSPACE.

common.bazelrc defines some common configuration options and also allows
to load user specific configuration.

Bug: 189451175
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I563c28ff09c483d9db79155dd28e49af0dd50005