Kleaf: migrate kernel_env to starlark rule
This introduces some nice properties: All labels (such as the script
locations and the source locations) are properly (type) checked for
existence and error messages about that are more obvious. Since callers
can use labels, usage is more obvious as well. E.g. the build_config can
now be referred to relatively from where this is called from and it
still gets resolved correctly.

Lastly, we spare some escaping in the command part.

Bug: 192656402
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I15ff9a5de13c0a406879de088fabad26b959b75d