Revert "ANDROID: ABI: update symbols to unisoc whitelist for the fifth time"
This reverts commit 52f8b40ff6a51e69f9c32687fbb53a518e8be34d.

It turns out that the unisoc file was never even hooked up to the gki
symbol build script so this file is doing nothing at all.

I tried to add it, but it breaks the build as it includes a lot of
symbols that are not even part of this kernel tree, so it must have been
created from a different kernel version.

Please resubmit this file properly, and hook it up to the build process
(by adding it to the list in the build.config.gki.aarch64 file.)

Bug: 186088840
Signed-off-by: Jian Gong <jian.gong@unisoc.com>
Fixes: 52f8b40ff6a5 ("ANDROID: ABI: update symbols to unisoc whitelist for the fifth time")
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: Ibdb7d57f0901aa7273c0c29e0a7e0eaf26909e26