Revert "usb: dwc3: gadget: synchronize_irq dwc irq in suspend"
This reverts commit 887a873e0ce2e5464a5d62f701900a07651e7a47.

dwc3_gadget_suspend() is called in a few places holding a
spinlock, and thus the synchronize_irq() added in the change
above causes scheduling-while-atomic deadlocks when the usb-c
cable is unplugged from the hikey960 board.

[   81.307243] BUG: scheduling while atomic: kworker/u16:1/49/0x00000002
[   81.313680] Modules linked in:
[   81.316732] Preemption disabled at:
[   81.316751] [<0000000065d6e7c2>] dwc3_gadget_suspend+0x54/0x84
[   81.326067] CPU: 4 PID: 49 Comm: kworker/u16:1 Tainted: G S              4.14.242-01115-ge4e640edffc0 #1
[   81.335544] Hardware name: HiKey960 (DT)
[   81.339470] Workqueue: events_power_efficient event_work
[   81.344780] Call trace:
[   81.347228]  dump_backtrace+0x0/0x15c
[   81.350887]  show_stack+0x14/0x1c
[   81.354202]  dump_stack+0xc0/0xfc
[   81.357516]  __schedule_bug+0x88/0xac
[   81.361177]  __schedule+0x90/0xaf8
[   81.364575]  schedule+0x6c/0x88
[   81.367713]  synchronize_irq+0x98/0xd4
[   81.371459]  dwc3_gadget_suspend+0x74/0x84
[   81.375552]  dwc3_otg_work+0x134/0x1d0
[   81.379297]  event_work+0x490/0x6b0
[   81.382784]  process_one_work+0x250/0x44c
[   81.386789]  worker_thread+0x244/0x450
[   81.390537]  kthread+0x17c/0x18c
[   81.393763]  ret_from_fork+0x10/0x18

So I don't think this backport makes total sense against 4.14,
at least not with the hisi usb changes in this tree.

As we were not seeing the problems the change describes, go
ahead and just revert this for now.

With this patch, I no longer see crashes when unplugging the
usb-c cable from the board.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I9497990919009370697c88856f152214e5ba70bb