ANDROID: pvmfw: Add Android Boot Partition header
Use mkbootimg to pass the pvmfw size to the system ABL; note that,
although pvmfw.img will have version 3+ of the header which requires the
presence of a corresponding vendor_boot image, no such image (and
therefore, image header) will be present for pvmfw.

AVB signing is not added by this patch.

This change requires aosp/1798576 (and parent) and aosp/1798147.

Bug: 194265608
Change-Id: I4c96e2a1d766a1d0fa76e7e3297e85e4af2aadc0