tools-utils: Do not skip symbolic links when analyzing a linux tree
For various reasons, (parts of) a linux kernel tree containing vmlinux
and kernel modules might be organized using symbolic links. One reason
is that those binaries physically reside on a different storage and are
linked into place. To support such scenarios by allowing to traverse
symbolic links when discovering linux kernel binaries.

I could not see a reason why this was disabled in the first place.
Ignoring symbolic links had been the default since the inception of the
kernel mode.

Hence, allow valid symbolic links and that is what this patch does.

	* src/abg-tools-utils.cc (find_vmlinux_and_module_paths): Do not
	skip valid symbolic links.

Link: https://sourceware.org/pipermail/libabigail/2021q3/003589.html
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I8b58853a36d613e0fb95707e5e438301ee1acfab