abg-reader: tidy logic around iterating over corpus nodes
A previous commit replaced bespoke XML iteration logic with loops that
iterate using xmlNextElementSibling. A couple of these loops have
bodies that can be executed at most once, resulting in a Clang
warning. In fact, inspecion reveals that the bodies are executed
exactly once and that the surrounding node logic can be simplified.

There are no behavioural changes.

	* src/abg-reader.cc (read_translation_unit_from_input):
	Simplify corpus node update logic.
	(read_elf_needed_from_input): Likewise.

Bug: 191235788
Change-Id: I96015f74e55a281ad9c9865e26a9c3fca262ec0a
Signed-off-by: Giuliano Procida <gprocida@google.com>