ANDROID: GKI: update arm64 gki_defconfig due to upstream changes
Commit 424953cf3c66 ("qcom_scm: hide Kconfig symbol") caused the
ordering of the config options to be moved around a bit.  This trips the
Android build system saying that the configuration files are not
correct, so fix this up by moving the sections into the place that they
are expected to be now.

This is only an issue with the arm64 GKI configuration, x86 is fine.

Fixes: 424953cf3c66 ("qcom_scm: hide Kconfig symbol")
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: I2ba470f71727078b4f70185d76408304f8b8d66e