aosp/android-mainline: change fscrypt direct I/O patches back to ANDROID
By necessity, we've been maintaining this feature out-of-tree ever since
it was required to be added back in January 2020
(https://android-review.googlesource.com/c/kernel/common/+/1212270).
These patches are also still not expected upstream in the near feature,
for various reasons.  Use the ANDROID tag, as was done originally,
rather than FROMLIST to avoid confusion with patches that are expected
to land upstream soon and don't have a maintenance commitment.

Change-Id: I0e6f6bbd2789b5ed5abff34b07777607a73d4b9b
Signed-off-by: Eric Biggers <ebiggers@google.com>