Interceptor: relativate execution based on `cwd`
`ROOT_DIR` - if set, otherwise assume `cwd` - is assumed to be the
directory all execution is done relatively to. Hence, when rewriting
commands, do it with that directory as a base.

The interceptor wrapper is determining the INTERCEPTOR_root_dir based on
either ROOT_DIR or falls back to `cwd`.

Bug: 205577427
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Ifef91ce7ae446d1a6cea7a744f36d0ff27cb4b3e