Interceptor: add command analysis
Add first version of the command analysis utilities. We match the
program executed against a set of (command line) analyzers and delegate
the detection of any inputs/outputs that can be derived from the command
line.

For the moment, this only implements analyzers for clang/gcc
compilers/linkers as well as archivers (ar, llvm-ar).

Bug: 205577427
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I72791d05a946a4210be6322358003a7250912b80