kleaf: filter build.config files within _kernel_env
... instead of in the macro.

This ensures that filegroups in `srcs` will be filtered properly,
because filegroups are not expanded in macros, but in rules.

The side effect is that the sources target, e.g.
kernel_aarch64_sources, contains both the kernel sources and the
build configs. This is okay because if build configs has changed,
the _env target is invalidated anyways, so everything will need
to be rebuilt.

Test: build slider
Bug: 205759526
Change-Id: I650838760514769f63eea6d7f19e17dfdea5ecd8