Autoupdate of 5 submodules
*prebuilts/kernel-build-tools updated to 8f0be9e88f61da123e52b3e081355d9604c5326e
add btfinfo and stgdiff to the prebuilt binaries

Bug: 188176217
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Ibe0f2faf65eb125088f1f6fdad1934fd31157e59

*external/elfutils updated to eed13d52bc4865ee18a257970be91b1402c7833b
Merge "Extend visibility to //external/stg"
*external/libabigail updated to 42c7a4e0f3de01e85a2fe2bb513c77d8316a64e0
Extend (header) visibility to //external/stg

This extends the general visibility to //external/stg.
Further, `libabigail` does now also export headers to targets that link
`libabigail` as static lib. This was not needed before because of the
lack of external library users.
stg is currently using libabigail headers that are not part of the
public interface (mostly due to unneccessary exposure), but this is
generally fine. Hence add `libabigail_src_headers` to expose those
internal headers specifically to //external/stg only. This will link as
we do not apply any symbol visibility restrictions to the built
libabigail (as opposed to the upstream libabigail build).

Bug: 188176217
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I6022906a4d6cf9ce315b0ebe51ea1b7489f69ca6

*external/libbpf updated to 7a43309a815342500d22e97d539a7e00c33c77e9
Extend visibility to //external/stg

Bug: 188176217
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Ib6faaa35b1b9c0c0796626aae77e5d3b432465ae

*external/stg updated to d3e19ed2744ee526040593f558efaff200e13977
Initial Android.bp for STG

This builds stgdiff and btfinfo and all dependent libraries.

Bug: 188176217
Change-Id: I151e46c84eb2167bc0273bb9ca180937a4c0ab2d
Signed-off-by: Matthias Maennich <maennich@google.com>