scripts/update_android_mainline: Skip non-Torvalds merges
During the merge-window, when `git merge`ing Mainline merge commits
into android-mainline, it is not ideal, but sometimes necessary (when
build failures occur between Torvald's merges) to merge in
non-Torvalds merges commits.  Because these can be based on arbitrary
merge-bases, these can put the Tracking repo into an unknown state
when rebasing on top of them.

For that reason it's probably best to skip them altogether and only
rebase on known-good (centralised) merge commits created by Linus.

Unfortunately, this means that when we go back to cherry-picking
single ANDROID commits, it is possible to do so on a different base
than they were applied to in android-mainline.  Which in turn means
that we cannot conduct our usual `git diff` checks prior to committing
them to Common Patches.

Here we are adding functionality for a transient state by providing a
Status: tag inside the series file header.  If set to 'Untested' the
tooling will not conduct `git diff` checks when a) committing the
changes to Common Patches nor when b) firing up the tooling for a
subsequent run.  However, once the Tracking repo is rebased onto its
next Torvalds merge commit, it will continue the checks from that
point and change the Status: back to Tested.

Signed-off-by: Lee Jones <lee.jones@linaro.org>
Change-Id: I66137f0e0b6c34116a145ea503f81366e20c3d05