ANDROID: allmodconfig: Re-enable TEST_KMOD
This has been disabled since the introduction of the configuration in
commit 884416f5176a ("ANDROID: create build.configs for allmodconfig"),
likely due to the fact that TEST_KMOD force selects XFS_FS, which was
also disabled due to a clang bug.

Commit 8c8c5eea927b ("ANDROID: build.config.allmodconfig: Re-enable
XFS_FS") re-enabled XFS_FS because the compiler was fixed but did not
re-enable TEST_KMOD for the same reason. Do so now so that allmodconfig
coverage is increased.

Fixes: 8c8c5eea927b ("ANDROID: build.config.allmodconfig: Re-enable XFS_FS")
Test: build.config.allmodconfig.{aarch64,arm,x86_64}
Change-Id: I4f6e70c268e214ccb054f57452386c509f1b2882
Signed-off-by: Nathan Chancellor <nathan@kernel.org>