interceptor: add an initial implementation for fake execution
Add an option to allow fake execution. This is enabled by running the
interceptor with --fake | -f, which it will enable analyzers during
execution to rewrite the command before execution (but after logging it
away). The default_fake implementation replaces the entire command by an
equivalent of `sh -c "truncate --size 0 <output> <output> ..."`. That
effectively touches all outputs leaving empty files behind. This is
enabled for the existing analyzers and possibly a good default for many
more. Yet there is no guarantee that this will actually make the build
succeed if downstream rules that are not captured by analyzers stumble
when working with the empty files.

The reason for wrapping the `truncate` in /bin/sh is to avoid having to
do the PATH lookup - $(which truncate) - ourselves. Hermetic toolchains
might provide a different `truncate` than /usr/bin/truncate, e.g. the
toybox variant.

Bug: 205735739
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Ib8478f52940071d07f6544d8ac4804048404ca64