kleaf: include modules.order.* for kernel_modules
In preparation for adding support to include the modules.order for
external modules, let's add a check to see if the files exist in the
staging directory for kernel_modules. If it's there, then include them
in the modules_staging_archive. The reason we can't add the
modules.order.* files as kernel_module outs in BUILD.bazel files is
because the name of the file depends on the variable ext_mod_rel and we
can't use a regex for kernel_module outs. Since all external modules
will either have this file or not, adding support in kernel_module
similiar to Module.symvers should be fine.

Bug: 210713925
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ic04e3f7bad71d0f58455c377dd82107a59862493