Autoupdate of 2 submodules
*build updated to 5751a77ef3cfc10e3edb6783cc797ce6cc7a75dc
kleaf: include modules.order.* for kernel_modules

In preparation for adding support to include the modules.order for
external modules, let's add a check to see if the files exist in the
staging directory for kernel_modules. If it's there, then include them
in the modules_staging_archive. The reason we can't add the
modules.order.* files as kernel_module outs in BUILD.bazel files is
because the name of the file depends on the variable ext_mod_rel and we
can't use a regex for kernel_module outs. Since all external modules
will either have this file or not, adding support in kernel_module
similiar to Module.symvers should be fine.

Bug: 210713925
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ic04e3f7bad71d0f58455c377dd82107a59862493

*common updated to 77cbbcd16fab0db6c1251fed13fd1f12a08e3325
FROMLIST: kbuild: install the modules.order for external modules

Add support to install the modules.order file for external modules
during module_install in order to retain the Makefile ordering
of external modules. This helps reduce the extra steps necessary to
properly order loading of external modules when there are multiple
kernel modules compiled within a given KBUILD_EXTMOD directory.

To handle compiling multiple external modules within the same
INSTALL_MOD_DIR, kbuild will append a suffix to the installed
modules.order file defined like so:

  echo ${KBUILD_EXTMOD} | sed 's:[./_]:_:g'

Ex:
  KBUILD_EXTMOD=/mnt/a.b/c-d/my_driver results in:
  modules.order._mnt_a_b_c_d_my_driver

The installed module.order.$(extmod_suffix) files can then be cat'd
together to create a single modules.order file which would define the
order to load all of the modules during boot.

Link: https://lore.kernel.org/all/20211215172349.388497-1-willmcvicker@google.com/
Bug: 210713925
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I3c23098d551ef17427436feee78bfddc9f9d36c6