kleaf: pass INSTALL_HDR_PATH to external kernel modules and archive headers.
With https://r.android.com/1825055:

"""
Makefiles of external modules that install custom UAPI headers using
"make headers_install" need to have access to INSTALL_HDR_PATH for them
to install their UAPI headers in the correct directory.
"""

Do this in kleaf as well. Archive the UAPI headers directory.

Note: This change does not archive the UAPI headers to the final
$DIST_DIR/kernel-uapi-headers.tar.gz. Subsequent changes are needed.

Bug: 211695803

Change-Id: I372b71cf18b85f1ce06849c48ba609f8d8810ca5
Signed-off-by: Yifan Hong <elsk@google.com>