kleaf: //build:kernel-build-scripts -> //build/kernel:kernel-build-scripts
This is part of the build/ -> build/kernel migration.

The //build/kernel:kernel-build-scripts module already exists,
so we migrate all usages to it.

Remove the //build:kernel-build-scripts module to prevent further
usage with a conditional. Right now, build/BUILD.bazel
is still a symlink to build/kernel/BUILD.bazel. Once
the symlink is removed, the conditional can be deleted.

Remove //build/kernel:internal-build-tools that were used as a
workaround during the transition. Now that all references to
kernel-build-scripts are pointing to
//build/kernel:kernel-build-scripts, the workaround is no longer
necessary.

Bug: 217630659
Test: TH
Change-Id: I005707c1d804875c1bcf8148c0427aa7daf2bf51