kleaf: //build:kernel-abi-scripts -> //build/kernel:kernel-abi-scripts.
This is part of the build/ -> build/kernel migration.

There are no current usages of //build:kernel-abi-scripts.

//build/kernel:kernel-abi-scripts already exists.

Remove the //build:kernel-abi-scripts module to prevent further
usage with a conditional. Right now, build/BUILD.bazel
is still a symlink to build/kernel/BUILD.bazel. Once
the symlink is removed, the conditional can be deleted.

Bug: 217630659
Test: TH

Change-Id: I9639109f93efea33b2fe2d85297679420ab20074