kleaf: Add stub for GKI modules.
When GKI modules are added to kernel_*_modules_install,
they are added to the dist target. Hence they must be
added to the downloaded target as well.

This ensures that, when a mixed device build adds
kernel_aarch64_additional_artifacts to their dist rule,
the GKI modules are also added to the DIST_DIR.

Currently, there are no GKI modules, so this CL has no
effect.

Test: none
Bug: 206079661
Change-Id: Ie653b9b64125864b70cf4f43fd7fc4787c16b6d9