Improve the SilentDeathTest docs a bit.
The test folks want to start chasing down *all* crashes seen while
running tests (which is a great idea), so we're likely to have more
new users. Let's make the docs more friendly to them, in particular by
adding some copy & paste for the *common* case, not just the *hard* case
as before!

Bug: http://b/211967298
Test: N/A
Change-Id: Ic0e902eb87544b50e2c9a32b8cdaa59c8e8922ed