Revert "ANDROID: Enable trimmed CONFIG_DISTRO_DEFAULTS"
This reverts commit 9a9f933b3acb1799d9291c90b9c60222a1de0386.

Updated to revert only bits of the original change. We won't use the
default distro_bootcmd any more because OtherOS can be implemented with
env injection via uenvcmd.

This eliminates a non-upstream diff to Kconfig and allows some unused
commands such as "fdt" and "pci" to be disabled. This also disables ISO
partition support and the default bootcmd, which were never used.

Clean up the pvmfw.fragment after some of the disabled features have
become the default, or were moved to the qemu minimal fragment (GUID
partition table support, no network support, no virtio_net support).
The only change for pvmfw is that the "fdt" command is disabled.

Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: Ifd90626e33e47ea055023a026fc0a88e53cc82eb