kleaf: source_date_epoch use workspace_status
Now that we have workspace_status properly set up,
SOURCE_DATE_EPOCH can also be stored in stable-status.txt.
Only when the content of the file changes (not the timestamp),
a rebuild will be triggered.

Delete the source_date_epoch.sh script and the hack in the
bazel Python wrapper.

Test: build twice, does not rebuild
Test: SOURCE_DATE_EPOCH=100 bazel build ...
      correctly passes the value to _kernel_env and triggers
      rebuild
Test: bazel build ... &&
      SOURCE_DATE_EPOCH=$(git -C .source_date_epoch_dir log -1 --pretty=%ct) \
      bazel build ...
      does not trigger rebuild
Bug: 202077908

Change-Id: I1c7b3a646769a4915c2089a1f9d01fbe5497f639