ANDROID: Avoid checking code_data if no BCC
The 'code_data' will be NULL if AVB verification failed, but libavb
allowed the failure through because the bootloader was unlocked. In such
a case, we shouldn't check the code_data BCC.

To unblock CI, skip accessing the code_data if no BCC is detected, which
should be the case for CF for now.

Bug: 217161634
Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: Id98a2860504f26390f67afbcda625d12dab080d6