Autoupdate of 4 submodules
*gs/google-modules/soc-modules updated to 3fef1d646c90ac46590fe55f232e44fc790a7fb5
Add support for using the ext modules touch Kconfigs

This adds supports to use the Kconfigs for the touch external modules:
common, sec, and fts.

Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ib11e0808f034e98dc02b001eb661ff094cad8ba9

*gs/google-modules/touch/common updated to cb103aa87480bad8d9674e6df14ff13ec8125c1f
Add support for using the Kconfig

With Kconfig support for external modules, We can now enable/disable the
configs using the defconfig. This change removes the hardcoded
KBUILD_OPTIONS and adds kleaf support for visibility to the Kconfig
file.

Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I0a111c0621799f15db82f4449c2051df016a7275

*gs/google-modules/touch/fts updated to b3f5e71ca55fab3dd25170d5c1fdca5db752cd5a
Add support for using the Kconfig

With Kconfig support for external modules, we can now enable/disable the
configs using the defconfig. This change:

* Removes the hardcoded KBUILD_OPTIONS.
* Adds SPDX license to the Makefile and Kbuild.
* Update BUILD.bazel to allow visibility to the Kconfig.

Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I44a27fae2d14fccb7fa40ec9d44e08d782a76a9f

*gs/google-modules/touch/sec updated to 1cf0e389666ceffde25ca2877b484aca1bca28c8
Add support for using the Kconfig

With Kconfig support for external modules, we can now enable/disable the
configs using the defconfig. This change:

* Removes the hardcoded KBUILD_OPTIONS.
* Adds SPDX license to the Makefile and Kbuild.
* Update BUILD.bazel to allow visibility to the Kconfig.

Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I4a1c791182be6ee3cf0f48821a98221ceb714f0b