ANDROID: align constness of extcon_get_state parameter
The implementation of extcon_get_state() defines `id` as const unsigned
int, while the declarations do not qualify. This leads to inconsistent
type information recorded when comparing Dwarf produced by Clang full
LTO and Clang thin LTO.

This is still valid code, but for now address this by aligning the
declarations with the definitions until it can be addressed in the
compiler.

Bug: 221022839
Bug: 220181989
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Iad555f551e998bec62a27cce63b3c033a46baf7f