Autoupdate of 2 submodules
*common updated to 3162b2f66c9fc4f00899ebda2fd86603f75968f8
ANDROID: GKI: update the pixel symbol list

Adds vb2_dma_sg_memops to the symbol list now that VIDEOBUF2_DMA_SG is
built-in to the GKI kernel.

Bug: 219998156
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I59af06d1da835e21751636dd758ac25d9d00c8b1

*gs/google-modules/soc-modules updated to 9eecba606815bcae2296b11b6ce179f2a1df573e
kleaf: remove videobuf2-dma-sg.ko as an output

The patch [1] builds this driver into the GKI kernel now. So we need to
remove it as a module output from the BUILD.bazel target.

[1] https://android-review.googlesource.com/c/kernel/common/+/1990854

Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I9a117787cdc9c3343978dc77b349a3e5435011e9