compare_mixed_trees: Add script to compare mixed tree sources
In a mixed build environment, it is important to keep the files used to
compile GKI kernel the same between the GKI kernel and the device
kernel. Otherwise, unexpected issues may arise. For instance, a change
made to a core kernel file on the device kernel would not have any
effect on the generated boot.img since that file isn't used to compile
vmlinux, the one from GKI kernel tree is.

compare_mixed_trees is a host-processing python script to compare the
mixed build source trees. After compiling a mixed-build workspace,
compare_mixed_trees can be run. This script will determine all the files
used to compile the GKI kernel and check whether any differ on the
device kernel.

The "core kernel files" list is generated by inspecting all the .cmd
files that Kbuild generates. These .cmd files contain list of source
files compiled and the files they depend on.

Bug: 203187541
Change-Id: I7ee5646c77f95aae732eec81177801a997f9c676