Revert "ANDROID: fips140: support "evaluation testing" builds via build.sh"
This reverts commit b51d71e4aa5f6caecd42f2428869674119b11f74 because
this script didn't end up being used.  Remove it for now.

This script was planned to be used originally; however, in fact making a
custom build of fips140.ko alone isn't sufficient to support the testing
requirements, some of which were unknown at the time.  The current
procedure is now to use a device-specific script to build custom boot
images; see http://go/fips140-for-android-gki#heading=h.bcuh1mqjhpc4.

Bug: 188620248
Change-Id: Ia6e41e3885fad7716d41ab466a656f097f3de2af
Signed-off-by: Eric Biggers <ebiggers@google.com>