Kleaf: add experimental flag that filters sandbox paths from the output
Passing --experimental_strip_sandbox_path will enable a filter on the
build output (stdout and stderr) to remove any occurrences of the fully
qualified sandbox path.

This obviously does not play well with the interactive mode of the bazel
output and colors are also lost. But it serves as a discussion base.

Bug: 222357601
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Ie56f13b80a9a8625e9e02cec43462c8f6896fe52