_setup_env: support Bionic sysroot for USERCFLAGS
This is necessary to support CONFIG_UAPI_HEADER_TEST=y in a build
environment which does not have a cross compiled version of glibc.
Instead, we will use Bionic.

This requires setting the correct path to the prebuilt Bionic headers
and libraries via --sysroot, and using the correct --target= triple,
which differs from what we generally use to cross compile the kernel
itself (which doesn't require a libc).

Bug: 190019968
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>
Change-Id: Iac98c377f516ea642cc605ab6d6d3fd022cb20d4