abi: remove close_fd from deny list
Some unknown out-of-tree IPC implementations want to throw around
internal file descriptors so in order for them to work in a GKI system,
the ability to close a file descriptor needs to be present.  Now that
this symbol is in an android-only-vfs-namespace, it should be ok to
remove it from the deny list.

Bug: 227656251
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: I265859ca56d8538ca2bea98a5cf089bd5107e1ee