Autoupdate of 2 submodules
*common updated to 094f5702db8686e9736e799cab89a2e5778a3d86
ANDROID: debug_kinfo driver, move to drivers/android

drivers/staging/android is going away, and systems seem to be relying on
this driver now, so move it into drivers/android/ to make everything
self-contained and prevent merge issues later on.

Bug: 170851792
Bug: 169101608
Cc: Jone Chou <jonechou@google.com>
Cc: Will McVicker <willmcvicker@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: Ib969a7544ea74881da61d7cba1f529a269921049

*gs/google-modules/soc-modules updated to 7947aecaf44ca183daae97878a8eb2b1a2117597
google/debug: update path to debug_kinfo.h

This header is moving out of the staging folder. So update the include
accordinly.

Bug: 229102238
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I630f640a4959e50986bd56cf1d5c22b9bc046b82