kleaf: Add intermediates_dir function.
This generally ensures that different targets have their own intermediates
dir. This is similar to

```
ctx.actions.declare_directory(ctx.attr.name + "_intermediates")
```

... but not actually declaring the directory, so there's no `File` object
and no need to add it to the list of outputs of an action. It also won't
conflict with any other actions that generates outputs of
`declare_file(ctx.attr.name + "_intermediates/" + file_name)`.

For sandboxed actions, this means the intermediates dir does not need to be
cleaned up. However, for local actions, the result of intermediates dir from
a previous build may remain and affect a later build. Use with caution.

Bug: 222767611
Test: none

Change-Id: I8aebd2495c7e89facf87fdc0a92c6749194e62a9
Signed-off-by: Yifan Hong <elsk@google.com>