kleaf: Add directory_with_structure.
This is a "class" that aims to replace tarballs with
declare_directory. However, the problem with declare_directory
is that, empty directories and subdirectories are not symlinked
in the sandbox.

To fix this problem, the "class" declares another text file
that stores the directory structure, and recovers it upon request.

Test: TH
Bug: 226673718

Change-Id: I26868d60bd311a4195c11b313f322845bcf9dcda