ANDROID: pvmfw: Prevent memory corruption from FDT
As the FDT location is received from the host, it can't be trusted. In
particular, the host could place it in a location leading to memory
corruption (such as right below the top the stack) when modifying it.
The FDT could also be placed in the expected location (start of RAM, as
done by crosvm) but advertise a size larger than its maximum size,
potentially overflowing into other regions of RAM (e.g. data, code,
stack, heap).

To prevent this, check that the FDT location and size are exactly as
provided by crosvm's BIOS mode.

Bug: 229964496
Bug: 229966941
Signed-off-by: Pierre-Clément Tosi <ptosi@google.com>
Change-Id: I61917352efc2664790ba3e132af6cbcac3f6e10b