Fix kfifo_out return value check
This fixes commit 24ba4efd42fe ("Check return value of kfifo_out") which
exposes a bug in the bluetooth HAL that causes a SIGABRT due to an
uncaught exception from trying to parse a negative number using stoul.
This SIGABRT crashes the bluetooth HAL. So to workaround this, instead
of returning an error, return 0 on error which is what happens if this
procfs file doesn't exist.

Fixes: 24ba4efd42fe ("Check return value of kfifo_out")
Bug: 230397443
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I8058a32985b6e7d5d38e216ee1032c8178edd4b3