kleaf: Fix abi_symbollist.raw logic for local builds again.
Before the change,
for local builds (--strategy KernelBuild/KernelConfig=local),
we simply hardcode the value
$(relpath abi_symbollist.raw ${KERNEL_DIR}) to UNUSED_KSYMS_WHITELIST
in .config. This does not work because the .config may be used in
sandboxed actions as well, e.g. when building modules_prepare.

Hence, set it to a consistent absolute path. In the case of
KERNEL_DIR=common for kernel_aarch64, for example,
the config is set to:

  /<workspace_root>/out/bazel/output_user_root/<hash>/execroot/__main__/bazel-out/k8-fastbuild/bin/common/kernel_aarch64_raw_kmi_symbol_list/abi_symbollist.raw

This is done by using a local action.

Dependant actions also depend on the File object that
points to the aforementioned file.

- For local builds, the config sets to the right path to
  access this file, even though the path is absolute.
- For sandbox builds, the config sets to the real path that
  exists outside of the sandbox.

The disadvantage is that using absolute paths may hurt
remote caching for developer builds in the future. We'll need
to revisit this in the future.

Test: TH
Test: manually build //common:kernel_aarch64_dist with \
   --strategy KernelConfig=local --strategy KernelBuild=local
Fixes: 228528064
Fixes: c263c4b0ecf788952bca4cda3f5364cd42d4f426
  (kleaf: Do not restore abi_symbollist.raw to KERNEL_DIR for non-sandbox builds.)

BYPASS_INCLUSIVE_LANGUAGE_REASON=The name CONFIG_UNUSED_KSYMS_WHITELIST
  is set upstream. This cannot be fixed without fixing upstream.

Change-Id: I6f5ba8cd6af1b0b3954b13cfe909c0981cd1b868
Signed-off-by: Yifan Hong <elsk@google.com>