Add rel_path2 to build_utils.
This is like rel_path, but creates a simpler path when
<from> is directly under <to>.

For example:
  $ rel_path /tmp/foo /tmp/foo/bar
  ../../foo
  $ rel_path_py /tmp/foo /tmp/foo/bar
  ..

This is useful in Kleaf. When a device calculates
  rel_path $ROOT_DIR $KERNEL_DIR
in build.configs, because build.configs are evaluated
in a KernelEnv action which is in a sandbox, this will
give (e.g. if KERNEL_DIR=common):
  ../../__main__
... where __main__ is the name of the directory of the
execroot. However, if building with --config=local,
the KernelConfig action is executed outside of the
sandbox, so ../../__main__ no longer works (unless
the workspace root directory happens to be named
__main__).

Creating this function so a device build config can
use it.

Test: TH
Bug: 229309039

Change-Id: I74764073d17674fa1e2e5a5e16b5fcfb7a6ed761