Autoupdate of 2 submodules
*build/kernel updated to 76087e638a0cb2105bdbbf5daf9734bf988988a4
Add rel_path2 to build_utils.

This is like rel_path, but creates a simpler path when
<from> is directly under <to>.

For example:
  $ rel_path /tmp/foo /tmp/foo/bar
  ../../foo
  $ rel_path_py /tmp/foo /tmp/foo/bar
  ..

This is useful in Kleaf. When a device calculates
  rel_path $ROOT_DIR $KERNEL_DIR
in build.configs, because build.configs are evaluated
in a KernelEnv action which is in a sandbox, this will
give (e.g. if KERNEL_DIR=common):
  ../../__main__
... where __main__ is the name of the directory of the
execroot. However, if building with --config=local,
the KernelConfig action is executed outside of the
sandbox, so ../../__main__ no longer works (unless
the workspace root directory happens to be named
__main__).

Creating this function so a device build config can
use it.

Test: TH
Bug: 229309039

Change-Id: I74764073d17674fa1e2e5a5e16b5fcfb7a6ed761

*gs/google-modules/soc-modules updated to 7a3253660cbfa8705d0448fc2744a9e71564ea7f
kleaf: Fix KCONFIG_EXT_MODULES_PREFIX for kleaf --config=local builds.

When it calculates
  rel_path $ROOT_DIR $KERNEL_DIR
in build.configs, because build.configs are evaluated
in a KernelEnv action which is in a sandbox, this will
give (e.g. if KERNEL_DIR=common):
  ../../__main__
... where __main__ is the name of the directory of the
execroot. However, if building with --config=local,
the KernelConfig action is executed outside of the
sandbox, so ../../__main__ no longer works (unless
the workspace root directory happens to be named
__main__).

This CL changes rel_path to rel_path_py so that,
though it is slower than rel_path by less than a fraction
of a second, KCONFIG_EXT_MODULES_PREFIX is set to
../ instead of ../../__main__ for Kleaf builds, which
works in --config=local builds too.

Bug: 229309039
Test: manual
Change-Id: I8743c1120b5a978d484108480205f40b853a300e
Signed-off-by: Yifan Hong <elsk@google.com>