kleaf: Add ability to control -j in Kbuild make
When --make_jobs is passed to the Bazel command line,
it sets the KLEAF_MAKE_JOBS, which is then fed into
the KernelEnv set up script via workspace_status_command.

The variable is placed in volatile-status.txt because
the number of jobs should not affect the final build output.

The variable does not impact MAKEFLAGS in _env.sh from
KernelEnv to ensure its value is updated for new actions.
For example:

  bazel build kernel_aarch64_env --make_jobs=10
  # 10 is not embedded in kernel_aarch64_env.sh
  bazel build kernel_aarch64 --make_jobs=60
  # kernel_aarch64 will use the value 60 if it
  # needs to rebuild.

This allows us to control number of jobs on a virtual
machine.

Test: TH
Bug: 226669832

Change-Id: I18c1b51241827523755d666c03411ead25b67861