abi/extract_symbols: update to avoid extra symbol list churn
If a symbol list is shared by multiple devices on a single GKI kernel
branch, you may experience shuffling of the symbols in and out of the
"preserved by --additions-only" section depending on which device tree
you run extract_symbols for. To avoid this extra symbol list churn, mix
all the symbols together if the flags "--additions-only" and
"--skip-module-grouping" are used together.

Bug: 232105104
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I69a83272223f2b4b5e25ded88abb896453df9dbb