checkpatch: Fix regex that detects KERNEL_DIR project
When a project path ends with the same string as KERNEL_DIR, it gets
picked up in the GIT_SHA1 variable which leads to an incorrect failure.
Fix the regex to only match KERNEL_DIR at the start of the line.

For example, we are incorrectly matching gs/google-modules/touch/common
when KERNEL_DIR=common:

  fatal: ambiguous argument
  'gs/google-modules/touch/a5e22817ff702d29a62b3ba505ce02f1afec6448':
  unknown revision or path not in the working tree.

Bug: 232841076
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: If693410014a9a3b25251c6f9768cdd6c6040ece7