extract_symbols: include gki_modules when collecting undefined symbols
If symbols don't have a "user", then they will get trimmed. This is
determined by an intersection of the generic_symbols and undefined
symbols. To support GKI modules, we need to include the symbols
referenced between all the GKI modules. For example, if GKI_module_A
references a symbol from GKI_module_B, we need to include that symbol in
the symbol list. This can be done by including all of the extracted GKI
module symbols in both the "generic_exports" list and the
"undefined_symbols" list.

Bug: 232988816
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ia9d8df863941b481923fe6ea62d3dd87cfb81e89