Autoupdate of 2 submodules
*build/kernel updated to 4d7edd707848292905073fee93ee11331bac402d
extract_symbols: include gki_modules when collecting undefined symbols

If symbols don't have a "user", then they will get trimmed. This is
determined by an intersection of the generic_symbols and undefined
symbols. To support GKI modules, we need to include the symbols
referenced between all the GKI modules. For example, if GKI_module_A
references a symbol from GKI_module_B, we need to include that symbol in
the symbol list. This can be done by including all of the extracted GKI
module symbols in both the "generic_exports" list and the
"undefined_symbols" list.

Bug: 232988816
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ia9d8df863941b481923fe6ea62d3dd87cfb81e89

*common updated to 158384f20d1eb3a2b360e5eebedc51ce7ed5cdd4
ANDROID: include GKI_MODULES_LIST

We need to include the GKI_MODULES_LIST so that we can properly extract
the GKI module's symbols when generating the symbol lists.

Bug: 232988816
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ifb459809ee311531c8cb3e04c8b5210d469e78cb