Allow KeyMint/Rust to depend on bssl_ffi
Some OpenSSL functionality is not present in BoringSSL (for example,
`EVP_PKEY_HMAC` and `EVP_PKEY_CMAC`) and so the corresponding safe
wrappers in the rust-openssl crate are configured out with
`#[cfg(not(boringssl))]`.

However, KeyMint/Rust needs access to some of this functionality, and
the only remaining way to get to it is directly via the `unsafe`
bindgen-created FFI entrypoints.

Bug: 195310053
Test: build
Change-Id: If1865dcce16365f55cad7802b0d1b63d8f5c2962