pcie: exynos: reset DMA mask to 36-bits after host init
The upstream commit 660c486590aa ("PCI: dwc: Set 32-bit DMA mask for MSI
target address allocation") forces the MSI target address to be a 32-bit
target address allocation so that the MSI target address can be used on
both 32-bit and 64-bit MSI capable devices. However, the wlan driver
uses a 64-bit allocation for the msi_msg. Due to this, we need to reset
the DMA mask back to 36-bit in order to not run into issues where the
target address is greating than the DMA mask.

This fixes the below PCIe failure which leads to the wi-fi driver
failing to initialize:

 DMA addr 0x00000008b1234010+8192 overflow (mask ffffffff, bus limit 0).
 WARNING: CPU: 4 PID: 857 at kernel/dma/direct.h:99 dma_map_page_attrs+0x254/0x278
 ...
 Call trace:
         : dma_map_page_attrs+0x254/0x278
         : gs101_pcie_dma_map_page+0x30/0x68 [pcie_exynos_gs]
         : dma_map_page_attrs+0x64/0x278
         : osl_dma_map+0xc8/0x1e4 [bcmdhd4389]
         : dhd_msgbuf_rxbuf_post_ctrlpath+0x1dc/0x1508 [bcmdhd4389]
         : dhd_msgbuf_rxbuf_post_ioctlresp_bufs+0x168/0x40c [bcmdhd4389]
         : dhd_prot_init+0x197c/0x2344 [bcmdhd4389]
         : dhd_bus_start+0x8ec/0xcdc [bcmdhd4389]
         : dhd_bus_devreset+0x148c/0x1740 [bcmdhd4389]
         : dhd_net_bus_devreset+0xb0/0x7fc [bcmdhd4389]
         : wl_android_wifi_on+0x5fc/0xbc0 [bcmdhd4389]
         : wl_android_wifi_accel_on+0x3c0/0x854 [bcmdhd4389]
         : dhd_open+0xda0/0x1a60 [bcmdhd4389]
         : wl_cfgvendor_set_hal_started+0xa8/0xae4 [bcmdhd4389]
         : nl80211_vendor_cmd+0x218/0x220
         : genl_rcv_msg+0x494/0x5c4

Bug: 232293973
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I719cee617e7ebe528d4e9e7490fc1955bc61b67b