kleaf: Add --verbose_failures and --noshow_progress to Android CI builds.
--verbose_failures are already configured. Moving
the flag here so static configurations stays in
AOSP.

--noshow_progress: The progress messages create unnecessary verbose build log files as bazel
does not distinguish between interactive and non-interactive terminals. Drop
the progress to make build logs even more readable.

There is a caveat that it will get difficult to triage stuck builds, but unless
that becomes a frequent issue, the tradeoff should be worth it.

Bug: none
Test: TH
Change-Id: Ieb145d32185f83c07bf20efcff2bc15e036efad1