kleaf: Enable building boot.img for GKI x86_64.
This is the Kleaf change for
5f16f3e97495fd476113be0712c8ebf2be906832
("Support building a boot.img for x86_64 GKI")

This allows building boot.img for x86_64 GKI when
the configurations are provided, e.g. with:

"kernel_x86_64": {
    "build_gki_artifacts": True,
    "gki_boot_img_sizes": {
        "": "67108864",
    },
}

This change also skips building boot-img.tar.gz
for x86_64 because build_gki_artifacts_x86_64
does not build it.

Bug: 233692305
Test: TH

Change-Id: Iec1978399d76726cdb36057db59112009d57b760