Revert "ANDROID: modules: introduce the MODULE_SCMVERSION config"
This reverts the MODULE_SCMVERSION config addition as it causes lots of
merge problems with 5.19-rc1.  It can be added back after 5.19-rc1 is
out if it is still needed.

Also, if it is still needed, perhaps it can be made a bit
less-intrusive to make it easier with future merges...

Bug: 180027765
Cc: Will McVicker <willmcvicker@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: I42e4ccbc4d2291523fd8460d530528d0c19c3e70