Dry running certify_bootimg
The certify_bootimg script will be executed on a server over a GKI
boot.img during the official certification process, which embeds
a GKI certificate into the boot.img. The certificate is for Android
VTS to verify that a GKI boot.img is authentic.

Dry running the process here so we can catch related issues early.

Bug: 230426945
Test: BUILD_CONFIG=common/build.config.gki.aarch64 build/build.sh
Change-Id: Ibd3ceddd007af62d2c74b73f9ae7bdbff5d3f95c