UPSTREAM: iomap: Support partial direct I/O on user copy failures
commit 97308f8b0d867e9ef59528cd97f0db55ffdf5651 upstream

In iomap_dio_rw, when iomap_apply returns an -EFAULT error and the
IOMAP_DIO_PARTIAL flag is set, complete the request synchronously and
return a partial result.  This allows the caller to deal with the page
fault and retry the remainder of the request.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Reviewed-by: Darrick J. Wong <djwong@kernel.org>
Signed-off-by: Anand Jain <anand.jain@oracle.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit ea7a57858875256e233d29b9c01b9f558f3bd12a)
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: I5de5ec3e55f18305f76e78ab4c72ce3ee1272522