kleaf: Add define_kleaf_workspace_epilog().
Because @bazel_skylib is defined in define_kleaf_workspace(),
we can't load symbols from @bazel_skylib in workspace.bzl.
Hence, create a separate epilog extension for WORKSPACE
file to load.

The epilog is currently optional. It is only needed to run
analysis test, defined in the next CL. A device kernel
repo may skip this if it doesn't run Bazel analysis tests.

Test: TH
Bug: 233286182
Change-Id: I79dba4e96e4a93887c26cc50278c9f9eac4846af