kleaf: Add kasan support.
The --kasan flag enables kasan build. This automatically
implies --lto=none; if --lto is specified otherwise, build
fails.

Also add analysis test to ensure that the above is enforced.

Also add artifacts test to ensure that the CONFIG_KASAN
is set properly, even when it is used in combination with
other flags.

Test: TH
Bug: 236012429
Change-Id: I170d3cc8b672f51a24192e4df39704924fc38771