ANDROID: nl80211: Fix reserved UAPI attributes order
Order of the reserved UAPI attributes merged through b/233387627 was
changed due to change in merge order of the commits to
wireless-next.git main branch.

Fix the reserved UAPI attributes order to be in sync with
wireless-next.git. These attributes are still reserved since the changes
not into linux.git yet. So, these attributes must not be used in
production until the UAPI change lands into linux.git tree.

Fixes: af34adaa6ab1 ("BACKPORT: FROMLIST: cfg80211: Increase akm_suites array size in cfg80211_crypto_settings")
Fixes: d975927513c7 ("BACKPORT: FROMGIT: wifi: cfg80211: do some rework towards MLO link APIs")
Fixes: 3f048cd911ff ("BACKPORT: FROMGIT: wifi: nl80211: support MLO in auth/assoc")
Fixes: 6abf3c5ef229 ("BACKPORT: FROMLIST: cfg80211: Indicate MLO connection info in connect and roam callbacks")

Bug: 238103864
Change-Id: If9038ba6ce953933faf827a6a4102223dcc436b3
Signed-off-by: Veerendranath Jakkam <quic_vjakkam@quicinc.com>