build_rootfs.sh: Mount by label and not path
In future changes, the block devices for the esp and rootfs can no
longer by assumed to be the first and second partitions, so change
stage1 to mount by LABEL. This causes some new problems, because
the label is the same for the ramdisk and rootfs on disk, so delete
the ramdisk device node so mount/blkid can't find or scan it. The
proc filesystem must also be mounted for labels to work.

Change-Id: Iae217e8d4b43e5ded4ce7c006e1fa1fbbb448a25