build_rootfs.sh: Add esp support to rockpi
Support building partitioned disk images with GRUB EFI embedded in a
special ESP partition. This recycles the ESP/GRUB work done for GCP
and Cuttlefish, but does not switch rockpi over to using the ESP boot
flow just yet.

When building rockpi with -g, the output image file will be partitioned
to match the original wrapping script's method, but with an additional
128MB ESP partition.

Change-Id: I887b8bfbd8eca3d4cd7b1c81753952f1d1debd4c