build_rootfs.sh: Fix arm64 with old builds
Newer cuttlefish-common packages don't depend on libc6:amd64 any more,
which is correct because newer builds don't need it, but we still need
to run these builds in automation. Manually preserve this package.

The code that was trying to do this was broken in three ways; firstly,
"uname -m" doesn't return arm64, secondly it was installing the wrong
package (-dev is for building, but we don't need it specially) and
finally it was installed inside the package dependency detection block
which caused it to be removed after the build was completed.

Change-Id: Iafe49910e99c556be859744e13364c6cb92129ab