abi: process_symbols: be silent by default
Teach `process_symbols` a `--verbose` flag in which it tells what it
does and where it puts files. Use that flag in build.sh and build_abi.sh
to preserve the current output, but do not enable the mode in kleaf
builds. This silents the bazel build for non-error cases.

Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I8a23a28347690d6d96e41860e05ec9111dec6031