build.sh: Set FRAME_WARN to 0 in LTO=none builds
In development builds such as sanitizer and LTO=none builds we may
end up with frame sizes that exceed the size limit due to compiler
implementation details. It is useful to enforce the limit in shipping
builds but not for these development builds.

We are already setting FRAME_WARN to 0 in KASan builds. Do likewise for
LTO=none builds.

Change-Id: Iefe991d5e6e231c3ee22ed313d7568d99a655619