kleaf: Use mnemonic matching for --config=local.
Change the mnemonic in the case of --config=local so that we
can tweak --//build/kernel/kleaf:config_local with transitions.

That is, when we use transitions to replace the value of
--//build/kernel/kleaf:config_local for a certain dependency "foo",
foo's action will have mnemonic "KernelBuild" or "KernelModule"
when config_local is false and "KernelBuildLocal" or "KernelModuleLocal"
when config_local is true.
Only *Local actions are set to --strategy local.

Test: manual
Bug: 248351908
Bug: 244649269
Change-Id: I90c1f8358161688b724d662616eef850274aefd3