Autoupdate of 2 submodules
*external/boringssl updated to f1228020c406e33689ef62ddd83d749bf313c01a
Add libcrypto_baremetal

BoringSSL is already used in bare-metal environments, eg. in
external/u-boot, but so far hasn't been compiled that way in Soong.
Now that we are building pvmfw in AOSP, add a libcrypto_baremetal
target that will allow us to link it against BoringSSL.

The target defines:
 - OPENSSL_NO_THREADS_CORRUPT_MEMORY_AND_LEAK_SECRETS_IF_THREADED
   - no threading support in this environment
 - OPENSSL_SMALL
   - reduce size by removing uncommon functionality
 - DOPENSSL_NO_ASM
   - remove dependency on getauxval() and __system_property_get

Bug: 237372981
Test: atest vmbase_example.integration_test
Change-Id: Ie412d2f9178b183162c9be111a3b4496ac6243b0

*external/open-dice updated to a7ec6851bde9f93fe44af682a00b5fa63a9708dc
Change bindgen targets to only use core

So as to use libopen_dice in baremetal Rust environments, modify the
bindgen targets to use core instead of std.

Bug: 237372981
Test: atest vmbase_example.integration_test
Change-Id: I89373fb37f3ab25e2291ca8667ea27375acf07b4