Language stable parcelables are VINTF stable.
Previously, we've only used these for NDK Parcelables.

The underlying mechanism here - relying on the framework
or NDK libraries - is VINTF stability.

There is a risk here - if these are configured incorrectly,
but that's nothing new - stability is a product of the
interface as well as the application, and since the
annotation does say 'Stable' it shouldn't be used
haphazardly.

Bug: 253506817
Test: aidl_unittests
Change-Id: Ib55eb0a872051e336dcbffcecba1be9c327d54e0