kleaf: Grab cmds from kernel_build/module
When the preserve_cmd setting is set (via transitions),
the kernel_build and kernel_module rules also preserve
the .cmd files from the output directory (similar to
how we keep .symtypes).

kernel_modules_install also provides all cmds from
its kernel_build and list of kernel_modules, so that
we can use that one label in label_settings to refer
to all targets that we need to analyze DDK headers on.

Test: TH
Bug: 248351908
Change-Id: I363d351ddfa101d4658ed2eb76ef1721b493da7c