ANDROID: mm: Introduce `__GFP_NO_INIT_ON_ALLOC`
`init_on_alloc` is a hardening measure which will zero out all kernel
allocations, whether using page based APIs (alloc_pages(), etc.) or
using kmalloc and friends. This is to prevent issues originating from
forgetting to initialize allocated memory and reading from it. This
feature can be turned on using 2 ways:
* compile time: `CONFIG_INIT_ON_ALLOC_DEFAULT_ON=y`
* boot time: `init_on_alloc=1` on kernel cmdline

The GKI kernel turns on the compile time option by default.

For `virtio-balloon`, this security feature is not necessary. The pages
in the balloon are stored in a repository safe from other system
components and never read. When deflating, when the pages are freed back
to the system, the next allocations should have `init_on_alloc` behavior
which will provide the expected hardening behavior. Moreover, the
initialization of the pages when allocated by the balloon, uses a lot of
CPU time.

Introduce a new GFP flag that will prevent the effect of `init_on_alloc`
called `__GFP_NO_INIT_ON_ALLOC`. This allows each allocation to specify
if the zero initialization should apply to it. This also makes it easier
to audit the call sites that skip initialization.

The balloon allocation performance improves more than 100x for an
initial allocation of 6.5GB. With `init_on_alloc` enabled, balloon
driver consumed ~15 seconds, whereas, it takes <100ms for the same
allocation with `__GFP_NO_INIT_ON_ALLOC`.

Bug: 254904802
Test: Build without errors.
Change-Id: I49a0753f896c70e06d59d87244da8e98561836e5
Signed-off-by: Vaibhav Nagarnaik <vnagarnaik@google.com>