Autoupdate of 2 submodules
*build/kernel updated to 112a8c35b35272c70cc6a5250186840e772b30d4
kleaf: Add allow_ddk_unsafe_headers option.

When this is set, //common:all_headers will also include
unsafe headers.

Bug: 248351908
Test: builds

Change-Id: Idd746ca817c5308c62c50f8b955ad16455654732

*common updated to ede1fd38eeafa71983e0f578bed60f12e0b87ca1
ANDROID: kleaf: Initial list of ddk_headers (0: skeleton)

Create the skeleton rules for DDK headers, and write comments
with hands.

This CL has no real effect; all targets are declared empty
and content is filled in in follow up CLs.

Allowlist: This is a list of headers and a list of
include directories that are known to be safe
to be used by modules.

Unsafe list: This is the list of headers that are known to be used
by some device kernel build, minus allowlist.
This means, with the allowlist and unsafe list, an unspecified
device kernel build can be transitioned to DDK without any change
to the source code.

- Note that for cleaness of DDK modules, we may want to
  remove some items in the allowlist of includes and
  require device source code to #include from the
  correct directory.

Bug: 248351908

Signed-off-by: Yifan Hong <elsk@google.com>
Change-Id: I4f15d2fac703bc3b7100a72e90f289236db3e7e5