ANDROID: Sched: Add vendor hooks required for WALT
Adding vendor hooks in several areas of the scheduler code.
These are needed for implementing Qualcomm's load tracking scheme WALT
via vendor modules. WALT was originally submitted upstream, but due to
complexities it was decided that a better path forward would to be
improve on existing PELT. Link to upstream submission of WALT can be
found below.

This reverts commit a7a5cf9f192a ("Revert "ANDROID: Sched: Add vendor
hooks required for WALT""). This change effectively brings back the
original change.

Bug: 316489257
Link: https://lore.kernel.org/lkml/1477638642-17428-2-git-send-email-markivx@codeaurora.org/
Change-Id: I15894ee08bb2e38439cf67c7708e64e52e69d6e5
Signed-off-by: Ashay Jaiswal <quic_ashayj@quicinc.com>
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>
Signed-off-by: kuyo chang <kuyo.chang@mediatek.corp-partner.google.com>