ANDROID: GKI: Sort system_dlkm modules list
Clean up:
List modules in android/gki_system_dlkm_modules
in a sorted order. List of modules are not order
dependent in this file. Initially it was meant to
keep same as android/gki_aarch64_modules which has
order dependencies due to being used as MODULE_ORDER
if building using build.sh. kleaf doesn't use
android/gki_aarch64_modules and it should be deleted
as part of final switch to kleaf.

Bug: 232431411
Test: TH
Change-Id: I8a5ce0f3466d4322ed956a4cb3d2f9c8ef075a86
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>