Autoupdate of 2 submodules
*common updated to 718c482c2b6b61101180620d26712cadb042c0f5
ANDROID: kleaf: Move list of kernel modules to an extension.

This is so that it can be loaded by other packages. Some
device's may wish to not build mixed build, in which case
they'll either need to use 'auto' in module_outs (discouraged),
or load the list of GKI modules.

Test: TH
Bug: 258308697
Change-Id: Iff7778b9edae7101a2debd44a49bb5eb403bfb2f
Signed-off-by: Yifan Hong <elsk@google.com>

*common-modules/virtual-device updated to d754d756ae18e9e62a9c113adb4d5d0278a1e0ab
Merge "kleaf: Add arm and i686 target." into android14-5.15