Autoupdate of 2 submodules
*common-mainline updated to f032c7ae4502f417939c84a1ca27766f5a2ad69f
ANDROID: kleaf: Move list of kernel modules to an extension.

This is so that it can be loaded by other packages. Some
device's may wish to not build mixed build, in which case
they'll either need to use 'auto' in module_outs (discouraged),
or load the list of GKI modules.

Test: TH
Bug: 258308697
Change-Id: Iff7778b9edae7101a2debd44a49bb5eb403bfb2f
Signed-off-by: Yifan Hong <elsk@google.com>

*common-modules/virtual-device-mainline updated to 0258bf73dec88a20c022edbb3fda60001eb737eb
kleaf: Add arm and i686 target.

Unlike the arm64 / x86_64 targets, the arm and i686
target does not use mixed build. This is because
we don't have an arm / i686 GKI target. Hence, this needs
to explicitly declare the list of GKI modules in
module_outs.

In addition, the default dist directory is set to
out/{name}/dist to avoid the legacy behavior of reading
BRANCH.

Test: TH
Test: manually build it with
  bazel build //common-modules/virtual-device:virtual_device_arm_dist
  bazel build //common-modules/virtual-device:virtual_device_i686_dist
Bug: 258308697
Bug: 231753340
Change-Id: Id35048df354efb7c25b15224cad4ebe7a3a1a0a9