Autoupdate of 2 submodules
*build updated to 67eec7396100d8a3b42d9040c507564abc870bca
kleaf: Add --allow_undeclared_modules option.

This is an option to suppress the remaining_ko_files
check in kernel_build.

When it is specified, and remaining modules are found,
a warning message is produced instead of a build error.

Test: allmodconfig
Bug: 230529320
Change-Id: I3ddbac09b5f8fadf3332219a53eac8f335a36e2c

*common updated to 6bf834cb5329a892e25b7b1bfe4287b7c7f8b7cd
ANDROID: kleaf: Add build rules for allmodconfig.

These corresponds to the build.config.allmodconfig.*
build configs.

Unlike build.sh, the Kleaf variant deliberately does not copy
any outputs to the distribution directory. There's not
even a dist target.
This is because allmodconfig is a build test only.

Test: bazel build --allow_undeclared_mdoules //common:allmodconfig_{aarch64,arm,x86_64}
Bug: 258259749
Change-Id: Iefe8f69149f6391938696e90243939d42d4fda84
Signed-off-by: Yifan Hong <elsk@google.com>