Support overlapping enum prefixes in _GetConstantName.
Currently, _GetConstantName will pick an arbitrary uppercase
constant in the module that starts with the specified prefix.
This means, for example, that if called with a value of 1 and
a prefix of "IFA_" it could return IFA_F_SECONDARY instead of
IFA_ADDRESS. In the specific case of IFA_F_*, there is a
workaround that ensures that these constants are never returned,
but this does not fix the general problem.

Fix this by allowing each module to explicitly specify a list of
enum prefixes. _GetConstantName will pick the constant with the
specified prefix.

Also update iproute.py and sock_diag to use the new scheme, and
add a unit test for it.

Test: new unit test
Change-Id: Ib3e2e0a5a44ea92203bef86873eb7cabc67630c7