Better/more explicit support for nested attributes.
Currently, parsing nested attributes is super hacky, relying on
passing a negative command to the method that parses attributes.
Instead, reuse the currently-unused "nested" parameter, changing
it from an int representing the current level of nesting to a
list containing all the attributes that the current attributes
are nested under.

Test: python3 iproute.py
Change-Id: Id80150e766235e73ccaa72dce9ca4297fa43dcd3