build_utils.sh: Append modules_list with newline
modules_list file is being used as filters to grep.
If the file is not ending last line with a newline
then it is not a valid filter for grep as in POSIX
line not ending with a newline is not a line. This
strips the last module listed in a modules_list file
from generated modules.load file.

Since this file can come from the vendor during the
build, add an extra new line to make sure last module
is not getting stripped from the generated modules.load
file.

Bug: 232431411
Test: TH, manual verification of modules.order
Change-Id: I46568c2feb038684fcabf992343d8ee8272d8c64
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>