ANDROID: GKI: db845c: Update symbols list and ABI
android13-5.15 is broken on Dragonboard 845c because of
recently added symbols.

So updated the symbols list by running:
  "BUILD_CONFIG=common/build.config.db845c \
    KMI_SYMBOL_LIST_ADD_ONLY=1 build/build_abi.sh -s",

and the abi_gki_aarch64 ABI by running:
  "BUILD_CONFIG=common/build.config.gki.aarch64 \
    ABI_DEFINITION=abi_gki_aarch64.xml KMI_SYMBOL_LIST_ADD_ONLY=1 \
     build/build_abi.sh --update --print-report"

========================================================

ABI DIFFERENCES HAVE BEEN DETECTED!

2 function symbol(s) added
  'enum drm_connector_status drm_bridge_detect(struct drm_bridge *)'
  'ssize_t strscpy_pad(char *, const char *, size_t)'

========================================================

Its been a while since symbol list for db845c was updated
on android13-5.15, so this update end up adding a lot more
symbols and not just the missing symbols which caused the
boot regression on db845c.

Bug: 146449535
Fixes: 6411d0c45878 ("drm/msm/hdmi: switch to drm_bridge_connector")
Fixes: d8b1b64a070e ("rpmsg: qcom: glink: replace strncpy() with strscpy_pad()")
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Change-Id: Ida49c88d8a29ab727c61761e8be468a8791f9c62