ANDROID: remove leftover includes from revert
The android_vh_kfree_skb tracehook was reverted in commit c1e292ea7c7c
("ANDROID: remove the android_vh_kfree_skb tracehook"), so we don't need
the includes for the trace/hooks/net.h header either as this was added
in the original patch for the tracehook. This aligns with upstream.

Bug: 163716381
Signed-off-by: Carlos Llamas <cmllamas@google.com>
Change-Id: Iebb3d82bb56c6cbbe5abef3fec6e926bd2335d50