Android.bp: eliminate library targets
Target names live in a global namespace and there was a conflict with
libinput.

STG is relatively small and compiling intermediate libraries is not a
big efficiency win.

We are going to be adding more source files and a new binary soon, so
it makes sense to keep things simple until and unless we find better
organisational principles.

Change-Id: I0ec645fee2e2a0e2ad8bfc9ed8e0249c48d6f158
Signed-off-by: Giuliano Procida <gprocida@google.com>