kleaf: Rework kernel_kythe
Instead of relying on interceptor compile_commands,
rely on compile_commands from Kbuild instead.

Also fix the long-broken kythe target. Note that for
a regular kernel build, Kbuild does the following:

- cd ${OUT_DIR}
- generate intermediate files into ${OUT_DIR}
- source files are referenced via absolute paths

Hence, in the generated compile_commands.json from Kbuild:

- Relative paths refers to files in OUT_DIR
- Absolute paths appears to be outside of OUT_DIR

In this change, the output directory is re-constructed with
necessary information before running runextractor.

Note: it is not tested that the generated all.kzip actually works.
There may be some additional path sanitizing required to make
this work. However, this is needed to not break the build.

Bug: 201801372
Test: build //common:kernel_aarch64_kythe_dist

Change-Id: I15878b3d737fbacca4af0903e5fe54e2decfd6d7