ANDROID: GKI: Source GKI_BUILD_CONFIG_FRAGMENT after setting all variables
build.config.gki sources a GKI_BUILD_CONFIG_FRAGMENT before all of
the variables that are considered as part of a GKI kernel build are
declared. This reduces the effectiveness of a
GKI_BUILD_CONFIG_FRAGMENT, as it is only able to modify a subset of
the build variables.

Thus, move the logic to source GKI_BUILD_CONFIG_FRAGMENT to the end
of the GKI build config files to provide more flexibility for a
GKI_BUILD_CONFIG_FRAGMENT.

Bug: 262930113
Change-Id: I74abb45f9043acce04cb0052f54fded4340a9366
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>