kleaf: Enable full console logs
By default, the bazel output only writes to the console up to
1048576 bytes. If you hit this limit, then a message like this is
printed:

  stdout (.../bazel-out/_tmp/actions/stdout-6) 21226460 exceeds maximum
  size of --experimental_ui_max_stdouterr_bytes=1048576 bytes; skipping

and the remaining output will be redirected to the file above. This
message is (1) hard to find in the build logs and (2) there is no
message printed at the end of the build invocation to indicate some of
the logs were redirected to a file. This makes it extra hard to debug
with kleaf since you may not even know parts of the logs are hidden away
deep (10 directories!) within the bazel-out folder.

The use-case that hits this consistently is when you build with the flag
--debug_annotate_scripts which is equivalent to: `set -x; make V=1`.

This patch disables the limit so that everything gets written to the
console (stdout/stderr) without any limitation. Refer to [1] for details
on this flag.

[1] https://bazel.build/reference/command-line-reference#flag--experimental_ui_max_stdouterr_bytes

Bug: 263392927
Change-Id: Ia17636f613f846dcbdebb6d5ccd98a90ca013635
(cherry picked from commit 46ea0da28e68bf340e2870c4048c72c9125b975b)