kleaf: run modules_prepare.setup before kernel_build setup
... when building external modules.

Before this change, kernel_build[KernelEnvInfo].setup restores
$OUT_DIR/scripts/sign-file and various others as symlinks,
then modules_prepare out_dir is restored, overwriting
$OUT_DIR/scripts/sign-file as a file.

This works for the host GNU tar. However, with the prebuilt toybox tar,
it follows $OUT_DIR/scripts/sign-file to a readonly file, and
notice that the symlink goes outside of the decompress directory,
throwing an error.

Hence, to prevent this behavior, do the following in this order before
building external modules:

1. set up environment variables (kernel_env[KernelEnvInfo].setup)
2. Restore configs
   (kernel_config[KernelConfigEnvInfo].post_env_info.setup)
3. Restore modules_prepare results (modules_prepare[KernelEnvInfo.setup)
4. Restore kernel_build's outputs

Test: TH
Test: build CF manually with --config=local
Bug: 243570975

Change-Id: Ic44dba376adbada1c26f191145616a276168593f