Set default values for mtime/user/group for all tar commands.
This ensure reproducibility and deterministic
artifacts. Bazel uses content-based caching when
determining if an action needs to be rebuilt. Setting
mtime ensures that the timestamp of the output does
not affect this caching. Setting user / group ensures
that the user that executes the build command does not
affect the output.

Test: TH
Bug: 254357038
Change-Id: I9430d2335b80030719275ae2818ac8f90649b88d