Set default values for mtime/user/group for all tar commands (2nd attempt)
This ensure reproducibility and deterministic
artifacts. Bazel uses content-based caching when
determining if an action needs to be rebuilt. Setting
mtime ensures that the timestamp of the output does
not affect this caching. Setting user / group ensures
that the user that executes the build command does not
affect the output.

Comparing to the first attempt, this solution does not require the
hermetic toybox tar.

Test: TH
Bug: 254357038
Change-Id: I6c68c967c9d5a6c3c2b9a486cfb64048371171b7