Autoupdate of 2 submodules
*common updated to cae07316c96b60159503888d08ecc138d627fb2f
ANDROID: GKI: KASAN: disable INLINE

and replace with PANIC_ON_WARN. The latter is needed to align with
the standard KASAN template.

The former is needed to workaround a boot issue on cuttlefish that
hasn't been root-caused yet. Likely cause is the larger text
space needed for the INLINE config.

This issue was surfaced on cuttlefish when the BL implementation
was changed to use a builtin environment 578b47ac41fe4433108.

Bug: 218315905
Change-Id: I57486f8befa720bbf49afe89c5454af8348816a5

*common-modules/virtual-device updated to d8e1867f2148aff332ebeeea93946277f3ebf83d
ANDROID: GKI: KASAN: disable INLINE

This is needed to workaround a boot issue on cuttlefish that
hasn't been root-caused yet. Likely cause is the larger text
space needed for the INLINE config.

This issue was surfaced on cuttlefish when the BL implementation
was changed to use a builtin environment 578b47ac41fe4433108.

Bug: 218315905
Change-Id: I231c7f1b8507850c9a9a3e275ba28ce0f4f8e176