kleaf: Use cache_dir for external modules when --config=local
With --config=local, we try to cache as many output artifacts
as possible to speed up incremental builds.

For external modules, this means caching the $OUT_DIR in the
--cache_dir.

However, for regular linux-sandbox actions, the --cache_dir
is not writable.

For DDK modules, we can't use the local strategy because a DDK
module needs to restore Makefile / Kbuild files into the source
tree. If we use local strategies, this will pollute the source tree.
Hence, we use a weaker form, the processwrapper-sandbox strategy.
Because DDK modules does NOT need to symlink a lot of source files from
the core kernel (internal_exclude_kernel_build_module_srcs=False),
creating the sandbox should be fairly fast.

For regular external modules, we do not need to write into the sandbox,
so using local strategy is fine. Local strategy avoids creating
a lot of symlink from the source files of the core kernel.

Bug: 254357534
Change-Id: Ibd98479b77606b2b5198faff9ec4f516ef4ab2fc