kleaf: Add a script that analyze input changes to an action.
This script returns the list of input files for given actions,
and their hashes.

This is useful to understand why a given action needs to be re-executed
when a seemingly irrelevant dependency changes. One may do the
following:

  bazel build <target>
  inputs.py <target> > /tmp/a.txt
  # Do the incremental change
  inputs.py <target> > /tmp/b.txt
  diff /tmp/{a,b}.txt

... to see that, after the incremental change, which direct
inputs have been changed, causing a rebuild for the given target.

Implementation note: most of the code just resembles the "Input" list
when running:

  bazel aquery <query> --output=text

However, the format of the text output is not stable and not meant to
be parsed by machines. Hence, use the jsonproto format instead, and
build the list of inputs ourselves.

Test: manual
Bug: 254357038
Change-Id: Ifa45d6cc29b099637c7ca9705fe6045137866504