Autoupdate of 2 submodules
*common13-5.15 updated to 6d2ec81c11aa476b0ddcda6da7fd46022309bebc
ANDROID: GKI: KASAN: disable INLINE

Replacing with PANIC_ON_WARN to align with the standard KASAN
template.

Disabling INLINE as a workaround for a boot issue on cuttlefish.

This is the same as the change made on branch android13-5.10
(aosp/2399656).

Bug: 218315905
Change-Id: I2137932a95444b94c7e9768a7f7e2879a08dfdd0
Signed-off-by: Liz Prucka <lizprucka@google.com>

*common-modules/virtual-device13-5.15 updated to e3d4619cfda72d8331479e8bae44980d0b4c09bc
ANDROID: GKI: KASAN: disable INLINE

Workaround for a boot issue on cuttlefish.

This is the same as the change made on branch android13-5.10
(aosp/2399812).

Change-Id: Id41513b430a5154fa4da862b37bac033745a252f