kleaf: trim kernel_module -> kernel_build dependency.
It used to be that kernel_module depends on all outputs
of kernel_build and base_kernel, including vmlinux, System.map, etc.

Trim the list so that only Module.symvers and kernel.release is
depended on by kernel_module.

Because we run `make modules_install` but don't actually run depmod
when building kernel_module, provide a fake System.map file when
building the external kernel module.

Test: TH
Test: use analysis/inputs.py to check the dependency of a
  kernel_module's action
Bug: 254357038
Change-Id: Ic0bd61eb835b7e3da1ae084d5b467abd17cd77dc