Android.bp: Define suitable library defaults for linking absl libraries
When using the absl libraries as for example libabsl_strings, one would
need to also get access to the headers for inclusion. The
cc_library_headers is not visible to targets directly and should
actually also only be re-exported by the libraries themselves. Do that
by defining a cc_defaults that re-exports libabsl_headers and add this
default to all library definitions.

Additionally, libabsl_base is a fundamental dependency for all absl
libraries. Hence, also add this as dependency included as a whole to not
make users explicitly depend on it. Of course, avoid the circular
depdendency on itself by excluding it in libabsl_base itself.

Change-Id: Icc68b465d0faaf023d163cd4b28ff95fd9769c34