Autoupdate of 2 submodules
*build/bazel updated to bb8044765246704af08c9461db28eeba37168f51
pass mixed build environment variables to Bazel

Previously, the Bazel invocation of mixed builds was executed in a shell
environment that didn't include the variables pass to `m`. This CL topic
adds an allowlist in soong_injection to pass environment variables into
to the Bazel invocation.

Bug: 195029134
Test: WITH_TIDY=1 DISABLE_ARTIFACT_PATH_REQUIREMENTS=true mss tidy-packages-modules-NeuralNetworks --bazel-mode-dev
Change-Id: I7bfafa4a7aac511c5a782c9efdad650652fe3557

*build/soong updated to db23e228353ca5fa5ed4ba050a43a25ef9a1d0c6
Merge "export allowlist of environment variables to Bazel"