ANDROID: Switch QEMU riscv64 target to smode
Linux crashes when loaded and jumped to in machine mode, and since it
happens so early it's hard to debug. Using OpenSBI to start U-Boot in
supervisor mode doesn't seem to have this problem, so just do that
for now.

Note that this change requires you to pass u-boot.bin with -kernel and
not -bios. The -bios will implicitly use a distro build of OpenSBI.

Change-Id: I25341745d2bcba552ce7dfe0cc342e79fd2e1a11