kleaf: Skip symbol violation check for kasan build
Disable symbol violation check for the builds with --kasan flag.
With --kasan extra kernel configs takes effect and it changes the
required symbols for the target which triggers the symbol violation
check failures.

Considering that kasan is not a valid GKI release configuration and
we do not have any stake in restricting symbols or ABI stability for it,
disable this symbol violation check to avoid build failures.

This doesn't resolve the runtime failures if unsigned modules tries
to use missing symbol reported by this check in the symbol list; but
We expect vendors to build kernel+modules+vendor_modules together with
--kasan; and can possibly also disable the runtime symbol protection
with CONFIG_SIG_PROTECT=n if required.

Bug: 269488249
Test: TH
Test: bazel build --kasan //common-modules/virtual-device:virtual_device_aarch64
Change-Id: Ia927435ba3424739f1d316a0841ace22a7404a96
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>