kleaf: Also set trim_nonlisted_kmi_setting for kernel_env.
This ensures that progress_message_note and config_setting_tags
are set properly. In addition,

  bazel build foo_env

and

  bazel build foo

always relies on the exact same foo_env with the proper trim setting.
The second command benefits from the caching from the first command.

Note: The first command is an implementation detail and should not
be relied on. However, this is still a nice little improvement.

Test: TH
Bug: 269674780
Change-Id: Ifcf06eaf22c8c77318b105f46c157b6afa4e63eb