kleaf: Delete unused trim_nonlisted_kmi_setting. (2nd attempt)
This reverts commit b17ad4ec7f5ac61fea316dffa4d7b87428895cd6.

It is replaced by force_disable_trim now. The setting is not
used or set anywhere, so delete the setting and all transitions
associated with it.

The revert is rebased on top of the latest changes where
force_disable_trim is set to True for kasan builds.

Reason for revert: build breakage
Bug: 270670362
Bug: 269674780

Change-Id: Ie9cc2ce7f1a4f4c815cc7e06c50935c29e593580