UPSTREAM: io_uring: ensure that fsnotify is always called
[ Upstream commit f63cf5192fe3418ad5ae1a4412eba5694b145f79 ]

Ensure that we call fsnotify_modify() if we write a file, and that we
do fsnotify_access() if we read it. This enables anyone using inotify
on the file to get notified.

Ditto for fallocate, ensure that fsnotify_modify() is called.

Cc: stable@vger.kernel.org
Change-Id: Id314de62dbf7f9a11d4861257ebe427499125f40
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Sasha Levin <sashal@kernel.org>
(cherry picked from commit df1ec53252d5b5b26ea49e30438741c9a6d89857)
Bug: 268174392
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>