kleaf: Default dist dir for GKI / db845c is out/{name}/dist
... instead of out/{branch}/dist. This intentionally
deviates from the behavior of build.sh because
calculating BRANCH requires some amount of hacks to pipe
the value from build.config.constants.

The decision in build.sh done long time ago has proven to
be problematic, so this is the opportunity to revise.

If one wants to revert to the original behavior, they
may specify --dist_dir manually, e.g.

  tools/bazel run //common:kernel_aarch64_dist -- --dist_dir=out/android-mainline/dist

Bug: 231753340
Test: TH
Change-Id: Id042c62a6dd1688a738e2447a8ea5700b2f33635