UPSTREAM: io_uring/rw: remove leftover debug statement
commit 5c61795ea97c170347c5c4af0c159bd877b8af71 upstream.

This debug statement was never meant to go into the upstream release,
kill it off before it ends up in a release. It was just part of the
testing for the initial version of the patch.

Fixes: 2ec33a6c3cca ("io_uring/rw: ensure kiocb_end_write() is always called")
Change-Id: Idb52f716de1829dbdaecf601686fe9ac4d658bb7
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 4b6f8263e931d00617c63bd1e90fcf26ce70a1bb)
Bug: 268174392
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>