kleaf: Detect diff in protected exports in ABI update
In the process of updating the ABI, it is crucial to detect
any alterations in the list of protected exports, and if
such modifications are found, the update should be
terminated.

In the event of a discrepancy, the build will fail and
display an error message along with a command to update
the protected exports list. This approach is comparable to
how changes to the symbol list are currently handled during
the ABI update.

Bug: 268679215
Test: Manual as follow:
1. Remove some symbols from "android/abi_gki_protected_exports"
2. bazel run //common:kernel_aarch64_abi_update
3. ABI update should detect the change and fail.
Test: TH
Change-Id: Ia2c8056e6b933f373b195a58918d2448bb5fcb04
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>