Autoupdate of 2 submodules
*build updated to d9c171d8fe4c30769db4eb1d2729716d8f7d5519
kleaf: hermetic cc toolchain for host binaries.

If --config=hermetic_cc is set, registers hermetic CC toolchain
from prebuilts/clang for cc_binary etc., and don't use
host CC toolchain.

In this CL, only host binaries and libraries can be built.
On-device (Android) binaries and libraries are supported in
the next CL.

Note: targets specified in the `tools` attribute of a genrule
automatically applies the transition to compile against the
execution platform. We use this trick to force that sub-targets of
linux_x86_64_tests_genrule are compiled against the host, not
the target platform (which will become Android in the next CL).

Test: bazel test --config=hermetic_cc //build/kernel/kleaf/tests/cc_testing:linux_x86_64
  (examined with --toolchain_resolution_debug=//build/kernel/kleaf)

Bug: 228238975
Bug: 239688839
Change-Id: I516b01c6e09728497cab031efdca9d534a68ee8a

*prebuilts/clang/host/linux-x86 updated to 8c227ab59b49addcc9aef943948bff976c4cbf04
kleaf: hermetic cc toolchain for host binaries.

If --config=hermetic_cc is set, registers hermetic CC toolchain
from prebuilts/clang for cc_binary etc., and don't use
host CC toolchain.

In this CL, only host binaries and libraries can be built.
On-device (Android) binaries and libraries are supported in
the next CL.

Note: targets specified in the `tools` attribute of a genrule
automatically applies the transition to compile against the
execution platform. We use this trick to force that sub-targets of
linux_x86_64_tests_genrule are compiled against the host, not
the target platform (which will become Android in the next CL).

Test: bazel test --config=hermetic_cc //build/kernel/kleaf/tests/cc_testing:linux_x86_64
  (examined with --toolchain_resolution_debug=//build/kernel/kleaf)

Include-Buildifier-Warnings: enabled

Bug: 228238975
Bug: 239688839

Change-Id: I0c7ca5744393c3f108361f757772e4e5409b8dce