kleaf: Add srcs to build kernel/module to KernelCmdsInfo.
This is so that analyze_inputs can know the list
of relative paths that were used to build the kernel or
a certain module, and use that as a recognized file / header.

If an input file is not declared in Bazel but in Kbuild, it is
unrecognized. See CL for analyze_inputs for details.

Bug: 248351908
Test: TH

Change-Id: I9ecfcca43a8bcb93687d694b40f13e5f471b3687