kleaf: infer -mainline from BRANCH not CONFIG_LOCALVERSION.
Previously on mainline branch, CONFIG_LOCALVERSION=-mainline,
and BRANCH=android-mainline. The latter does not match
androidxx format, so android_release is not set.

However, because the setlocalversion script puts
the content of localversion text file
before CONFIG_LOCALVERSION, this causes the result to be
e.g. -gabcdef-mainline, not -mainline-gabcdef. We want
the string to be consistently

  MAJOR.MAJOR.MINOR-BRANCH(-KMI_GEN)?-SHA

where BRANCH could be either androidxx or mainline.

Also update the integration test to reflect the set up
on mainline branch (BRANCH is set, KMI_GENERATION is not set,
and CONIFG_LOCALVERSION is not set on all branches).

Bug: 272828027
Test: inegration test

Change-Id: Ie6388891f067e4ed7bc7f17a5a53e2d2fb4bd23f