Autoupdate of 2 submodules
*build/kernel updated to dc739c4e83ab1abc42eae96aed0f47cdfb7b5d34
kleaf: infer -mainline from BRANCH not CONFIG_LOCALVERSION.

Previously on mainline branch, CONFIG_LOCALVERSION=-mainline,
and BRANCH=android-mainline. The latter does not match
androidxx format, so android_release is not set.

However, because the setlocalversion script puts
the content of localversion text file
before CONFIG_LOCALVERSION, this causes the result to be
e.g. -gabcdef-mainline, not -mainline-gabcdef. We want
the string to be consistently

  MAJOR.MAJOR.MINOR-BRANCH(-KMI_GEN)?-SHA

where BRANCH could be either androidxx or mainline.

Also update the integration test to reflect the set up
on mainline branch (BRANCH is set, KMI_GENERATION is not set,
and CONIFG_LOCALVERSION is not set on all branches).

Bug: 272828027
Test: inegration test

Change-Id: Ie6388891f067e4ed7bc7f17a5a53e2d2fb4bd23f

*common updated to e28f3d09362a291a05f4277dc699dbca2827bad4
ANDROID: Disable AF_RXRPC for allmodconfig.

AF_RXRPC imposes a 49 char limit on UTS_RELEASE, but building with

  BUILD_NUMBER=P12345678 bazel build //common:kernel_x86_64_allmodconfig --config=stamp

requires more than 49 chars in UTS_RELEASE. Hence, this module
is disabled for allmodconfig.

Bug: 273576197
Test: ABTD
Change-Id: Ife604db109f870eda2fdbc0668d6bd1f4e4e906e
Signed-off-by: Yifan Hong <elsk@google.com>