Autoupdate of 2 submodules
*build/kernel updated to dc739c4e83ab1abc42eae96aed0f47cdfb7b5d34
kleaf: infer -mainline from BRANCH not CONFIG_LOCALVERSION.

Previously on mainline branch, CONFIG_LOCALVERSION=-mainline,
and BRANCH=android-mainline. The latter does not match
androidxx format, so android_release is not set.

However, because the setlocalversion script puts
the content of localversion text file
before CONFIG_LOCALVERSION, this causes the result to be
e.g. -gabcdef-mainline, not -mainline-gabcdef. We want
the string to be consistently

  MAJOR.MAJOR.MINOR-BRANCH(-KMI_GEN)?-SHA

where BRANCH could be either androidxx or mainline.

Also update the integration test to reflect the set up
on mainline branch (BRANCH is set, KMI_GENERATION is not set,
and CONIFG_LOCALVERSION is not set on all branches).

Bug: 272828027
Test: inegration test

Change-Id: Ie6388891f067e4ed7bc7f17a5a53e2d2fb4bd23f

*common updated to 3fbf22de81899d13fa4d11ccfdcfda3ccb419e36
ANDROID: Drop CONFIG_LOCALVERSION/KMI_GENERATION from gki_defconfig.

The text -mainline is inferred from BRANCH instead.

Test: TH (kleaf integration test)
Bug: 272828027
Change-Id: Ie5d74b69246873ac583d2abb90b1909b44c70ce0
Signed-off-by: Yifan Hong <elsk@google.com>