kleaf: Set EXT_MODULES before entering create_modules_staging
... for initramfs and vendor_dlkm, but not system_dlkm.

Change I6099930a0dfc
"build_utils.sh: retain Makefile load order for external modules"
adds a logic for create_modules_staging to rely on EXT_MODULES
to determine where to find modules.order. However, EXT_MODULES
are not required to be set for Kleaf builds.

Fix this by setting EXT_MODULES properly before diving
into create_modules_staging. The variable is calculated by
summarizing the ext_mod of `kernel_modules`
in kernel_modules_install, where ext_mod is
the directory of the Makefile (for kernel_module) or the package
name (for ddk_module).

This is not expected to cause issues downstream, but in the unfortunate
event that it does, one may set `--noset_ext_modules` temporarily
to workaround the issue.

Test: TH
Bug: 271921038
Bug: 272772812

Change-Id: Ic730fc9a7d6cc7eb6c9a648a2cf19383351950ca