kleaf: Do not use kernel_config_transition.
The only use of it is because we need to set LTO to none
when kasan is set. However, transitions on kernel_config
is also applied to dependencies, including
module_signing_key, which may be a genrule with indeterminitic
result.

This causes the following to produce surprising results:

genrule(name = "signing_key"
kernel_build(name = "foo", module_signing_key = ":signing_key")
copy_to_dist_dir(name = "foo_dist", data = [":foo", ":signing_key"])

... because foo depends on signing_key that applies the
kernel_config_transition, but foo_dist depends on the signing_key
that does not apply the transition, causing two keys to be generated.
The one used for signing is inconsistent with the one in distribution.

Also keeps the LTO value in kernel_config_settings.bzl so that a change
in LTO means a change to OUT_DIR under --cache_dir; this is to prevent
b/257288175: change in LTO values causes incremental builds to fail
with --config=local.

Test: TH
Bug: 273963531
Change-Id: I0f15dbbfac24d30a4693f2041c33ed00a0299265