kleaf: Deprecate abi.prop in ABI monitoring
* `abi.prop` file is being deprecated, its use was
recently deleted from gki/download_from_ci tool.
It is no longer used for active monitoring and it
has not been used for passive monitoring, making
it safe to remove.

* This removes Kleaf support to generate `abi.prop`
files, by deleting the rule which generates it.

* This deprecates the attribute `unstripped_modules_archive`
from `kernel_abi` rule, as it was only used for
generating `abi.prop` file.

Bug: 271560365
Change-Id: I3c8efb7ef9ed8878e4039673010cbd73e2a06ae0
Signed-off-by: Ulises Mendez Martinez <umendez@google.com>