kleaf: --config=hermetic_cc is the default.
This enables compiling cc_binary / cc_library etc. in the
kernel tree against the hermetic toolchain under
prebuilts/clang.

Also, because we want new features behind a flag so it can
be easily reverted / disabled, this change deletes the
TODO to replace --extra_toolchains with register_toolchains()
in WORKSPACE so that, if we want to revert this in the future,
we can simply delete the line that says:

  build --config=hermetic_cc

Test: TH
Bug: 239688839
Bug: 228238975
Change-Id: I112c02238827da4ce70963db5a7042523c60ec5e