ANDROID: gpiolib: of: add a quirk for legacy names in max77759_charger
The driver is using the non-standard "max77759,extbst-ctl" name for
a property describing this gpio.

While this driver and DT are under our control, Linux mysteriously
refuses to call this driver's probe function if we rename the
property in the DT to the expected name with the "-gpio" suffix.
Renaming the property is necessary in order to convert this driver to
gpiod APIs instead of the obsolete of-specific gpio APIs that are
about to be removed from Linux.

Add a temporary quirk so we can update the max77759_charger driver and
unblock merging Linux updates while investigating the probe issue.

Test: driver still loads successfully and can claim all GPIOs
Bug: 273738655
Change-Id: Ib6aa27db41c11f49acf49910a76bd3bd45f68a06
Signed-off-by: André Draszik <draszik@google.com>