kleaf: Use resolved python toolchain
The python toolchain is already registered in workspace.bzl.
Instead of using symlinks, actually check the registered toolchains
instead.

This allows us to replace the python toolchain if necessary.
Note that even if the python toolchain is replaced, we still need
the default python toolchain in the source tree under
prebuilts/build-tools for bootstarapping (for tools/bazel and
workspace_status_stamp.sh).

Bug: 274975777
Test: TH

Change-Id: I39049c48a3b8f3829a25d2fb601b0c33d2232a61