kleaf: Disable mixed build when building compile_commands.json.
This allows proper discovery on objects like init/built-in.a by
disabling KBUILD_MIXED_TREE.

This change effectively allows building compile_commands.json
for in-tree modules for actual devices, not just vmlinux
in GKI. External modules are still unsupported. A full kernel
source tree is required in KERNEL_DIR for this to work.

Test: run on a device build
Bug: 274056535
Change-Id: Ib443359d861a76e91001f004da744613423a0b5b