ANDROID: virtio-balloon: add proposed field for `wss_num_bins`
this field is being proposed in the virtio spec for balloon device and
will be soon incorporated. changes to `virtio_balloon_config` need to be
in sync between the kernel and crosvm. push it before the
`hugepage_order` so that when `wss_num_bins` field is finalized, it
doesn't cause further changes.

Bug: 276786967
Test: tested with kernel and crosvm changes in sync.

Change-Id: I367b1d875feeaee0b4f34de34cdcb373f156f788
Signed-off-by: Vaibhav Nagarnaik <vnagarnaik@google.com>