Autoupdate of 2 submodules
*build/kernel updated to 28e584690932f2ee21fe1523dd57f70cd71745a0
kleaf: Fix run kernel_config when defconfig is generated.

When the defconfig file is generated, the
arch/$SRCARCH/configs/$DEFCONFIG file does not
exist beforehand. Hence, there is no need to
record its symlink and restore it later.

In this case, simply print the path of
the new defconfig file if it exists.

If it doesn't exist, it may be the case that POST_DEFCONFIG_CMDS
already moved it away.

See microdroid for an example on how to make this properly
work with generated defconfig files.

Bug: 275468906
Bug: 276889737
Test: tools/bazel run //common-modules/virtual-device:microdroid_aarch64_config -- menuconfig

Change-Id: If0b94f577ad064c4ef06329102033717c3e8c899

*common-modules/virtual-device updated to 920328d2dc55e9fc261c6f904d54e762beda1a0a
kleaf: Fix microdroid menuconfig.

When config.sh runs, BUILD_WORKSPACE_DIRECTORY points to the
actual location of the source tree. Modify POST_DEFCONFIG_CMDS
to copy to the correct location.

Also add && true to the end of PRE_DEFCONFIG_CMDS so that
the config.sh does not exit prematurely after
eval PRE_DEFCONFIG_CMDS -- for whatever reason that I don't
understand right now.

Bug: 275468906
Test: tools/bazel run //common-modules/virtual-device:microdroid_x86_64_config -- menuconfig

Change-Id: I026dae350b7f90605be61f74258e0e84f794d1cb