kleaf: Add write_depset tooling.
The utils.write_depset function dumps a depset into
execution phase, so it does not need to be expanded
in the analysis phase (which is not parallel).

This is an effort to avoid depset.to_list()
everywhere.

A more-efficient alternative would be to add arguments to
ctx.actions.run_shell to include an Args with the depset. This
approach could save some I/O. However,
for complicated actions like kernel_build and ddk_config with multiple
steps, it can be a headache to parse the arguments
in bash at the beginning, and use the variables in individual steps.
This could make steps even less standalone. On the other hand,
the write_depset() approach in this CL is easily passed around
between the steps; all we need is to `$(cat {depset_file})` to
get its value.

Hence, we would prefer using ctx.actions.run_shell(arguments = ...)
for small actions / scripts, and utils.write_depset for big actions /
scripts.

Test: TH
Bug: 256688440

Change-Id: I7e1e07f4ebe897bd40470387a3a83b1ff84b1034