Autoupdate of 2 submodules
*build/kernel updated to b0377a072bb3f78cdacfd6d809914a9d1b0c0148
ABI: allow utf8_data_table in the symbol list

This symbol is needed to support case-folding on 6.1+ kernels.
We need to include it here since no GKI modules are using it.

Bug: 278561372
Change-Id: I1140f083957805e61f7d68e9b0f87bc2da20f430
Signed-off-by: Will McVicker <willmcvicker@google.com>

*common updated to 7e35917775b8b3e3346a87f294e334e258bf15e6
ANDROID: Add utf8_data_table for case-folding support

The upstream commit 2b3d04787012 ("unicode: Add utf8-data module") moved
the unicode nomalization functions into a module so that it could be
selectively included in the kernel based of support for case-folding.
In doing so, utf8-core.c uses symbol_request() to reference the
utf8_data_table symbol. Since GKI trims unused symbols, include
this symbol to our symbol list so that it doesn't get trimmed.

Link: https://lore.kernel.org/all/ZDnbW1qYmBLycefL@google.com/
Bug: 278561372
Change-Id: If6347087d8dffae63bb4af5d72b74f703f09b820
Signed-off-by: Will McVicker <willmcvicker@google.com>