PRE_DEFCONFIG_CMDS does not write to source tree.
Instead of writing to the source tree $ROOT_DIR/$KERNEL_DIR,
write to $OUT_DIR instead. Note that $OUT_DIR isn't defined
at this point, so it is left as a variable in the
PRE_DEFCONFIG_CMDS.

Also add && true hack for bash 5.2 bug so that menuconfig
can be used properly.

Bug: 275468906
Test: manual

Change-Id: Ic5b525d3523aa9dfc3c410f929edd7b6fd9cb471