ANDROID: GKI: Clean up: Remove protectred lists
mainline is not a GKI release branch and there is no
ABI enforcement or monitoring on it. Runtime protected
symbols and exports checks are also disabled for it.

Remove protected list of modules and exports lists
files to clean up and easy maintainance of the branch.

This also helps with module conversion starting on this
branch and cherry picks to other GKI release
branches with exports list updates.

android/gki_system_dlkm_modules: Continue updating this
list with module conversion process in the future;
so newly converted modules are available in
system_dlkm_staging_archive.tar.gz.

When a new ACK LTS spawns from the mainline
`android/gki_system_dlkm_modules` becomes the default list
for list of protected modules for all architecture and
export lists can be created with the ABI tooling.

Bug: 280352539
Test: TH
Change-Id: Id80c6d7706b4d9f18e086a86f49fd728fcdbcd02
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>