Deprecate rel_path and rel_path2.
For rel_path2, it is a relatively-new function and all
use cases should have already been cleaned up, so make
it a hard error to use.

For rel_path, Kleaf already sets up PATH to be hermetic tools
before executing _setup_env.sh, so it should be okay to use
realpath --relative-to directly. However, a lot of build configs
might still be relying on the function, hence only make this
a warning.

Test: TH
Bug: 266980402
Change-Id: Iaa25c6fffd22cc16af60bb79853cbe41b56f28d4