kleaf: kernel_build.kmi_symbol_list is configurable.
Before, we embed kmi_symbol_list in a list,
making it non-configurable. After this change,
we create a level of indirection with an extra
target that accepts select() with label as
values (not label list). This makes
kernel_build.kmi_symbol_list configurable.

We can't use alias() because alias.actual must not
be None.

Bug: 283159637
Test: TH

Change-Id: I3be4a56175cb1148796828c600a11b1832bae64d