Revert "Release Module.symvers to Android Build"
This reverts commit bfe4dc1ee49f0139d00f53b5859446c3c9e758ac.

Reason for revert: Module.symvers cannot be part of the output without some renaming because both kernel_build and kernel_module has the potential to claim it as part of the output. When one see the Module.symvers file in the dist directory, it does not know whether it belongs to the kernel_build or kernel_module's.

Change-Id: I45d85e8eefafd0411576fe7515f1911cdadf5702