kleaf: kernel_build do not depend on kernel_toolchain_info.
This change drops the dependency from kernel_build.bzl
to @kernel_toolchain_repo (specifically, for clang version).
The dependency was previously needed to get the default
toolchain_version. After the change, when toolchain_version is
set to None, Bazel's toolchain resolution will give us the fallback
toolchain that does not have
  //prebuilts/clang/host/linux-x86/kleaf:clang_version
set, which is the last one in the toolchain resolution list.
Hence, we don't need @kernel_toolchain_repo any more.

After this change, the dependency would be:

  //build/kernel/kleaf ->
  //prebuilts/clang/host/linux-x86/kleaf ->
  @kernel_toolchain_repo

(and later, it is expected that the package for rust toolchain
prebuilts will also depend on @kernel_toolchain_repo).
//build/kernel/kleaf itself won't need to depend on
@kernel_toolchain_repo.

This is one step forward towards external kernel module
builds without the GKI source tree (where @kernel_toolchain_info comes
from), which is a prerequisite for DDK v2.

Bug: 272164611
Change-Id: I1b926c581c75255bbc489f2a99af866ffde94331