kleaf: Move makefiles_test cflags tests to integration test
... in common-modules/virtual-device. We should not
test the implementation details of makefiles, but
rather test the public API. This reduces the maintenance cost
when the implementation of cflags changes.

The _makefiles_include_ordering_artifacts_test is intentionally
kept because testing the include ordering with integration tests
involves constructing a lot of confusing directories and headers.
This test is very clear because it lists all -I options explicitly.
We'll just have to change the test when cflags changes (in the upcoming
CL).

Test: TH
Bug: 282786692
Change-Id: I40eff3c8a74d77090f07b76a7dec7caf95a3a6ba