Autoupdate of 2 submodules
*build-trunk updated to 93b3c0520f0e34feed24010c23f9ac4c16252ad4
kleaf: Move makefiles_test cflags tests to integration test

... in common-modules/virtual-device. We should not
test the implementation details of makefiles, but
rather test the public API. This reduces the maintenance cost
when the implementation of cflags changes.

The _makefiles_include_ordering_artifacts_test is intentionally
kept because testing the include ordering with integration tests
involves constructing a lot of confusing directories and headers.
This test is very clear because it lists all -I options explicitly.
We'll just have to change the test when cflags changes (in the upcoming
CL).

Test: TH
Bug: 282786692
Change-Id: I40eff3c8a74d77090f07b76a7dec7caf95a3a6ba

*common-modules/virtual-device-mainline updated to 95a9375c28a5192d092a11d51c62e64deade513e
kleaf: Add integration test for ddk_module cflags.

This includes local_defines and copts. This replaces the
makefiles_test that pokes into the implementation details
of ddk_module; instead, it tests the public API only.

Test: TH
Bug: 282786692
Change-Id: I3ff8edd273a6768ac69f7732b1e6557312f6f9c7