kleaf: DDK: use a variable for all flags for a module.
When a module is built with multiple source files, these source files
share the list of the same cflags. Hence, move the cflags to
a single variable named after the module.

In addition to this change, the variables in _handle_includes /
_handle_copts are quoted properly so that $(ROOT_DIR) is not
single-quoted. It doesn't matter in practice because ROOT_DIR
is still resolved when clang is invoked, but it makes the
generated Kbuild file less confusing.

Test: TH
Bug: 282786692
Change-Id: Id1b66120efc46f939c3d3c476135e8e1c98698c8