kleaf: Add flag to ignore missing repo projects
Currently, the workspace_status_stamp.py script will perform various
operations on all projects defined in the repo manifest.

In some cases, the project directories may not be present. When this
happens, the build currently fails. Users should have the option to
change this behavior.

Add a flag, --ignore_missing_projects, which will ignore conditions
where a project is listed in the repo manifest, but does not exist in
the workspace.

Bug: 284943448
Change-Id: Icd062ecbabe7f6883733b1675e0ccf0f397bb587
Signed-off-by: John Moon <quic_johmoo@quicinc.com>