Autoupdate of 2 submodules
*build updated to 86d9360cbfdaecd18261ec6b1d5c971fdd1be498
kleaf: DDK: put cflags in file

When there are too many cflags, the build may fail with
the following error:

  make[3]: /bin/sh: Argument list too long

This is usually caused by the super long include paths
in the clang command line, though it is possible to be
caused by others. Hence, we put some of the generated flags
into a .cflags file, then provide it to clang with the @
syntax.

Caveats:
- Environment variables are not allowed in .cflags files,
  so we need to replace references to $(ROOT_DIR) manually.

Fixes: 282786692
Test: TH (see integration test in common-modules/virtual-device)
Change-Id: I1df4cc928970b5dd3e6a104c7a136957b03ba378

*common-modules/virtual-device updated to 1f9216e5e838e6020234d76fd3730332a68d8ce0
kleaf: Add test for very long cflags list.

Test: TH
Bug: 282786692
Change-Id: I84d7629ec68553cc6d4a49fd9b09c1b1efa3e3fa