ANDROID: GKI: add skb_pull_data to android/abi_gki_aarch64_virtual_device
In commit 88deda7cd8fa ("Bluetooth: btrtl: add support for the
RTL8723CS"), the symbol skb_pull_data is added to the btrtl.ko module.
This symbol is already in the normal GKI symbol list, but not in the
virtual_device list, so the build breaks for that target.  Add the
symbol to android/abi_gki_aarch64_virtual_device to fix this up.

Fixes: 88deda7cd8fa ("Bluetooth: btrtl: add support for the RTL8723CS")
Change-Id: Ie5b697563f3de2d7171c5662ee7f4b12defa3ca6
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>