kleaf: Add --no-group to all rsync calls
Currently, some rsync operations that pass the "-a" flag result in
the following error:

rsync: chgrp "<sandbox>/<file>" failed: Invalid argument (22)

Based on syscall analysis, this issue appears to be driven by the
Linux namespacing used by Bazel. Inside the sandbox, the group rsync
is trying to assign to the file does not exist, so the syscall fails.
This can occur when the user's default GID maps to some LDAP group
that is not managed locally in /etc/group.

To avoid this issue, pass "--no-group" to rsync. That way, it will
not attempt to set the group. By default, rsync will copy the file
with the user's primary GID anyway, so this change does not impact
the destination files' group assignments. This has an added benefit
of removing an extraneous syscall from all of these copy operations.

To make sure this behavior is consistent throughout Kleaf, configure
the hermetic tools to invoke rsync with --no-group.

Bug: 287321095
Change-Id: Iea49985ff465568b49d2bf082868cfa0ea95351b
Signed-off-by: John Moon <quic_johmoo@quicinc.com>