kleaf: Move --//common:use_prebuilt_gki to //build/kernel.
Use a central flag for the use_prebuilt_gki boolean
flag.

Even though the flag only affects ACK trees that
calls define_common_kernels(), moving the flag
from the ACK tree to //build/kernel avoids the dependency
bazel.py (which is under //build/kernel) to //common.
This is the last dependency from //build/kernel/... to //common
(except tests).

The //common:use_prebuilt_gki{,_set} targets are marked
deprecated. However, for backwards compatibility,
they are not removed, and they are aliased to
//build/kernel/kleaf:use_prebuilt_gki{,_is_true}.

After this change, if a device.rc specifies
--//common:use_prebuilt_gki, it may optionally
switch to a cleaner --use_prebuilt_gki that avoids
the deprecation warning.

Test: TH
Bug: 287700360
Bug: 230672545
Change-Id: I1b3c3e3fe97db7587926df03d3ff8db1b8e5d712