kleaf: Add visibility to kleaf/{impl,artifact_tests}/**/*.bzl.
This ensures that these extensions, which are implementation
details of Kleaf, doesn't get loaded outside of kleaf.

The public symbols are re-exported in kernel.bzl.

If this breaks a build, one may use
--nocheck_bzl_visibility to temporarily disable
the check.

Also fix buildifier warnings.

gki_artifacts.bzl are not included in this change due to an internal usage.

Test: TH
Bug: 283857454
Change-Id: If8139982e0e55a72729fbc058ec99965d069195e