kleaf: Prevent symlink dereference of dtstree
Currently, Kleaf will derefence any symlink to the dtstree while
forming a path relative to the kernel tree. With --config=local
builds, when the dereferenced path is outside of the kernel tree,
the dtstree becomes inaccessible because non-kernel sources are not
copied into the cache directory.

Add -s to the realpath call to prevent the symlink dereference.

Bug: 290822281
Change-Id: I3f0455caaa41c078587b10bd3906a35f6bcca4d8
Signed-off-by: John Moon <quic_johmoo@quicinc.com>