ANDROID: KVM: selfest: Add pkvm selftest
This test does not rely on the hypervisor accepting guest register
values from the untrusted host. It restructures the pKVM test
initialisation so that the register values are communicated via an
identity-mapped paged at the top of IPA space and wraps the test in a
small piece of bootstrap assembly to get the MMU up and running before
jumping to the predefined C entry point.

Bug: 428697579
Bug: 458241298
Co-developed-by: Will Deacon <willdeacon@google.com>
Change-Id: I70acf35de4e54a2b9e078037f7890a573da98b36
Signed-off-by: Will Deacon <willdeacon@google.com>
Signed-off-by: Fuad Tabba <tabba@google.com>