kleaf: abi_dump: allow empty module_outs
When the kernel_build does not specify any modules, the
abi_dump rule previously put a *.ko token in the shell
script, which is interpreted by shell as-is because
it cannot be resolved to any files.

Change it to a find subcommand so this works when
module_outs / module_implicit_outs is empty.

Bug: 284023435
Test: TH
Change-Id: I822758f0ab67074627f34a6c1a7475fc2f20a8ce