kleaf: Delete check on --kasan / --kasan_sw_tags conflict
Now that we have ensured all defconfig fragments are properly
applied, it is already checked in config_utils.bzl that
CONFIG_KASAN_OUTLINE must NOT be applied with --kasan, and
CONFIG_KASAN_OUTLINE must be applied with --kasan_sw_tags.
Hence, there's no need to explicitly check that they are specified
simultaneously.

Test: TH
Bug: 294390500
Change-Id: I0ef7cf66190a249a8595ee802441cea50408285c