kleaf: hardcode name of @platform//cpu:*
It used to be that @platform//cpu:arm and @platform//cpu:arm64
are real constraint values. In a recent bazel release, these
have been changed to aliases to other labels.

To avoid breaking in the future, use names that are prevalent
in Kleaf; that is, arm, arm64, i386, x86_64, and riscv64.

Test: builds
Bug: 295376544
Change-Id: I6b0d09d0d55f67ce5666ca0ccb959dfa66dea6f0