Autoupdate of 3 submodules
*build/kernel updated to 3c3fbc89d6575f16c1f9df0ea00e90b689d607c9
kleaf: hardcode name of @platform//cpu:*

It used to be that @platform//cpu:arm and @platform//cpu:arm64
are real constraint values. In a recent bazel release, these
have been changed to aliases to other labels.

To avoid breaking in the future, use names that are prevalent
in Kleaf; that is, arm, arm64, i386, x86_64, and riscv64.

Test: builds
Bug: 295376544
Change-Id: I6b0d09d0d55f67ce5666ca0ccb959dfa66dea6f0

*prebuilts/bazel/common updated to 78e15a6cc66a822a384e41c0094e0a0fa3cd383f
Merge "Bazel Release" into main
*prebuilts/bazel/linux-x86_64 updated to cfc546c18abe45b026a94824115cbfe3a3068a8b
Bazel Release

Bug: NA
Test: TARGET_BUILD_VARIANT=userdebug TARGET_PRODUCT=aosp_arm64 ./build/bazel/ci/mixed_droid.sh
Change-Id: I3e354cdd6decff61427a1241054cc2e857b6d6b0