ANDROID: Use DTC from PATH.
Instead of using a hard-coded path to DTC, infer its
path from PATH. This change removes the dependency on
the ${BUILDTOOLS_PREBUILT_BIN}/dtc symlink, so the
build infrastructure is free to reset PATH to a place
where the actual DTC binary exists.

This change is needed to set
--incompatible_disable_hermetic_tools_symlink_source
to true by default.

Test: bazel run //common:db845c_dist
Fixes: a71e480e82a7 ("ANDROID: build.config: Set DTC to absolute path")
Bug: 291816237
Change-Id: I75dac41802bd66a1b6c0c61d90b5b378b943b593
Signed-off-by: Yifan Hong <elsk@google.com>