kleaf: delete CONFIG_DELETED_UNSET from defconfig_must_present_test_module.
With 4eeb98f62760 ("kleaf: config_utils: properly check for unset
configs."), CONFIG_'s undeclared in Kconfig will simply
be dropped. If a ddk defconfig or kernel_build defconfig fragment
specifies `CONFIG_X is not set`, since it is correctly unset
in the final .config, don't complain if the .config doesn't contain
the line `CONFIG_X is not set`.

Bug: 295980099
Test: manual
Change-Id: I25d5b501610c77dd2364b1e4c157960c50344996