ANDROID: FF_RUMBLE support for the virtio input driver
A ff_device will be created only if the type EV_FF and the code
FF_RUMBLE is supported by the device. This is because this patchset is
meant to help with vibration support of a Gamepad.

A new custom input event type is created to help transfer rumble
effects. The new custom event and the playback event will be sent through the status queue.

Test: In test app that calls Vibration APIs, verified that the correct
input events made it to the Virtio Input Device.
Bug: 293154799
Signed-off-by: Richard Zhang <rizhang@google.com>
Change-Id: Ib129e5e2ecbae64a23a3af3c03ea7611dcf4b759