Autoupdate of 2 submodules
*common updated to b0e30c021b79d9cb9a67b12a94d1fe2f61126f14
ANDROID: arm as an option for get_gki_modules_list

If driver config depends on ARM64, driver is not
available for the ARM targets as module.

Introduce arm as an option for get_gki_modules_list()
to separate ARM64 dependent modules.

virtual_device_arm Cuttlefish target is the current
consumer of this; and it fails when there is ARM64
dependent module is introduced like OEM hypervisors.

Bug: 293529933
Test: TH
Change-Id: I462e8968faa48d58721d884688af62ff603c9a3d
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>

*common-modules/virtual-device updated to c49b1ce3e08951439dee35c9a570e622da200e0c
kleaf: virtual_device_arm: Replace arm64 modules with arm

Drivers which depends on ARM64 are not available
for virtual_device_arm but using arm64 to get list
of gki modules have them causing a build breakage.

get_gki_modules_list() now supports arm as an option;
so use that instead.

Bug: 293529933
Test: TH
Test: bazel run //common-modules/virtual-device:virtual_device_arm_dist
Change-Id: I8ffb0901f6bc6b9c338de873d93b3c5fa3ef936f
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>