ANDROID: GKI: revert all external USB/sound changes
There are a lot of USB and Sound patches added to the android15-6.1
branch due to the branch being brought over from the android14-6.1 tree.
These patches are conflicting with upstream LTS changes, AND they
consist of a set of half-backported changes that have not gotten any
updates since they were added to the android14-6.1 tree.

Revert the whole lot, starting over with a clean set of USB and sound
drivers, so that everyone can backport only the specific needed changes
going forward for the android15-6.1 tree as required.

Bug: 299122809
Change-Id: Id8557c4fc15f58fb6ab585acfd460d980c6fc9c9
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>