ANDROID: GKI: revert irc-gic changes
The irc-gic drivers have diverged in the android15-6.1 branch for no
real reason and this causes conflicts with upstream changes for them.
Because of this, revert all of the outstanding changes to allow LTS
merges to complete properly.

If they are needed in the future, they can come back in a way that works
properly, after being accepted upstream.

Bug: 180471389
Bug: 190353898
Bug: 195914333
Bug: 168730134
Cc: Elliot Berman <eberman@codeaurora.org>
Cc: Maulik Shah <mkshah@codeaurora.org>
Cc: Neeraj Upadhyay <neeraju@codeaurora.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: I318ca92ceaac910e9775dfbfeede1f83cea0c6ae