Autoupdate of 2 submodules
*common updated to cc087d337ceeb274f8aa66a6e7baa1dcc2c12d84
Remove VIDEOBUF2_DMA_SG from kernel but support building as module

This patch makes it possible to build VIDEOBUF2_DMA_SG as a module. The
reason it's not possible without this patch is that the flag is hidden
(there's no prompt), so it's automatically removed by the build script
when trying to explicitly set it.

This does introduce a divergence but it's very minor and allows us to
become GKI compatible.

Bug: b/290816514
Test: tast run $DUT arc.PlayStore.vm
Change-Id: I58da2f3e309fc56a7d2b08e3907606cdb69f60fe
Signed-off-by: Richard Fung <richardfung@google.com>

*common-modules/virtual-device updated to 4d6abce04f7a098625e0d00f8ee6680e42c26940
Build VIDEOBUF2_DMA_SG as a module

Bug: b/290816514
Test: tast run $DUT arc.PlayStore.vm
Change-Id: Ie34ec573c1155f05153e0503f0435622cdbda3d6
Signed-off-by: Richard Fung <richardfung@google.com>