ANDROID: GKI: Enable CONFIG_SWIOTLB_DYNAMIC
Newer kernels use SWIOTLB bounce buffers for DMA transactions involving
kmalloc'ed buffers that are sub-cacheline aligned. This allows the
minimum alignment for kmalloc'ed buffers to be reduced beyond the
cacheline size, as long as SWIOTLB is enabled.

Currently, a single SWIOTLB buffer is allocated at boot and is used for
bounce buffering for the lifetime of the device. This means that the
worse case SWIOTLB usage must be known and allocated up-front at boot,
which makes enabling SWIOTLB cumbersome.

Instead, CONFIG_SWIOTLB_DYNAMIC can be used to dynamically allocate
SWIOTLB buffers for bounce buffering as needed. This allows the kernel
to boot with a small sized SWIOTLB buffer, and continue to function
by allocating memory for bounce buffers as needed.

Bug: 296282294
Change-Id: If19ae55ee302f1bf97d11d8bcd0d72987939b79a
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>