ANDROID: don't glob() multiple times.
Calling glob() in macro implementation is
discouraged. However, for the sake of
simplicity in the API, the kernel_build()
macro implicitly calls glob() for all files
in the ACK tree. This expression is expensive.

To avoid multiple glob()'s when analyzing the
ACK tree, explicitly spell out srcs
for each kernel_build invocation.

Test: manual (by looking at the Bazel profile)
Bug: 305301717
Change-Id: Iabd98326674b97bc76ed037c9c560741dfd8e7ef
Signed-off-by: Yifan Hong <elsk@google.com>