i2c: switch back to use struct i2c_driver's .probe() (prepare for v6.6 kernel)
Linux commit [1] has removed the temporary struct i2c_driver::probe_new()
callback. This was a tempoorary callback to allow drivers to switch to a new
implementation of struct i2c_driver::probe() (which has the same signature
as probe_new()).

We can update this driver to use the .probe() callback easily and preempt
breaking the kernel build.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=5eb1e6e459cf
    ("i2c: Drop legacy callback .probe_new()")

Bug: 307735139
Test: TH
Change-Id: Ib56f0a0f9c7cce217b6c5a1dc02869bc833da634
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>