kleaf: artifacts test do not depend on hermetic_tools aliases.
Drop references to

  //build/kernel:hermetic-tools/...

Such aliases have been deprecated. This ensures
that running these tests won't produce warning
messages because they depend on the deprecated
aliases.

To achieve this, I defined py_binary for each
test script, then replaced py_test with hermetic_exec_test
that invokes the script with the given arguments.
in PATH before running the test. Then I modified
the tests to search for the binary in PATH.

Test: TH
Test: bazel test //common-modules/virtual-device:x86_64/goldfish_drivers/goldfish_sync_test //common:kernel_aarch64_test //common:kernel_aarch64_additional_tests --test_output=all --config=stamp
Bug: 234125794
Change-Id: Id84321227e6af04e2433516c30f8b11cbb9bcafe