kleaf: split "help kleaf" code to a different file.
The core functionality is kept in bazel.py, Move
the less-essential `bazel help kleaf` code to
kleaf_help.py to avoid clobbering bazel.py,
and keep bazel.py succinct.

The wrapper script is hard to test with (all tests
need to be an integration test), so it is crucial
to keep the file short for readability.

Test: TH
Bug: 293260835
Change-Id: I1cd3afe0684ded08b1039542b8a41b1ddb5172f4