Update prebuilt Clang to r498229b (17.0.4).
clang 17.0.4 (based on r498229b) from build 11039501.

This new respin contains the new MLGO regalloc model that is trained
locally on ARM64 AOSP. This improves app launch performance by 0.55% on
ARM64. This is on top of the 0.53% improvement from the original model.

http://go/art-benchmark?p=BootImageProfileId:36864

This also contains the inlining model from ARM64 Chromium.

Bug: http://b/293827654
Test: N/A
Change-Id: I9e7b2684fa98e1353e084a19fe776608f5917a2b