Autoupdate of 2 submodules
*system/security updated to c0e1f3589b7fa81c83324483263617cc8eb0fcce
Merge "Adding 'static bound on binder Interface implementation" into main
*system/tools/aidl updated to 461a5af6b4da20cae88cf9a7b62b7ba1c2137f75
Adding 'static bound on autogenerated code

Adding a DowncastSync trait implementations to the Binder Interface
trait required some changes to the lifetimes of the autogenerated
code. This will be used to cast a Binder Native object back to the
original object that implements the Binder Interface.

Test: CI
Bug: 278780666
Change-Id: I65158d9c425120c73bbd8e2cf75206ba3f5b651c