ANDROID: KVM: arm64: Add pKVM module register_unmask_serror
This new pKVM module op allows to catch SErrors at EL2. To do so, the
unmask and mask callbacks must return true. They are responsible for any
synchronisation needed to flush pending SErrors.

Catching SError at EL2 is meant to help debugging of such error by
crashing as early as possible to identify the culprit. The hypervisor
has no way to recover and if one is caught, a panic will occur.

Bug: 278749606
Bug: 305828248
Change-Id: Ib398a9a52bb7c708ce3f7f6b6b726aecc90c1628
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>