Revert "sched/debug: Remove the /proc/sys/kernel/sched_child_runs_first sysctl"
This reverts commit 17e7170645e34c519443ba63895264bbdee7beee.

There are loads of CTS tests that want to see this proc file present in
order to "pass".  So, fake them out and just put it back, it's not been
hooked up to anything for a long time, so it isn't hurting anything to
put it back.

Odds are no one will even notice it...

Change-Id: If70527d377ffaabaa9e796153e4c735c1cf66de1
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>