Autoupdate of 2 submodules
*external/libcxx updated to f45e673a553610dc230df3171c30ab5f8953f869
Add targets of symlinks to genrule's sources

A change is being made to sbox so that symlinks aren't resolved
before being copied into the sandbox. This requires that the targets
of the symlinks are also added to the sandbox.

Bug: 307824623
Test: m libc_musl_sysroot_libc++_headers with aosp/2850015
Change-Id: Idc46c38a2c92826b52d82add8221c69c2f1394fe

*external/libcxxabi updated to a07162d9edded852c2cec533b5b0d0b0bef2618c
Add libcxxabi_headers filegroup

libc_musl_sysroot_libc++_headers needs it.

Bug: 307824623
Test: m libc_musl_sysroot_libc++_headers with aosp/2850015
Change-Id: Ie2c8e9c22f0a60ddbb5d8374aa9c87eb046378b1