Kleaf: abi: Fix notrim transition
Currently this is broken, because the with_vmlinux transition returns a
no-op transition [0] for one scenario where no changes are needed, however
notrim always sets FORCE_DISABLE_TRIM, so the same feature cannot be
used, and the full list must be defined.

This change fixes this by returning the current values of these settings,
in which case these settings need to be part of the transition inputs too.

[0] https://bazel.build/extending/config#no-op-transitions

Fixes: a0b0cd9edaf3 ("Kleaf: Make trim only transition possible")
Bug: 283435842
Change-Id: I69215c0819b6f32c03100a021eab34313d8194bf
Signed-off-by: Ulises Mendez Martinez <umendez@google.com>