temporary bugfix: allow FUSE_PASSTHROUGH bit to work without error
Currently, code in fs/fuse/inode.c that checkes FUSE_PASSTHROUGH is
broken because the conversion of (1 << 31) to u64 results in a
mis-interpreted signed-32 integer value.

There is a different fix for this in newer kernels, and it will
take discussion to decide how to apply it - but this intermediate
fix will unblock CTS while this takes place.

Bug: 315384676
Test: Run quick CTS test

Change-Id: I2dc1c08192531c1132bd1805f0fa0b965f8f1782
Signed-off-by: Alex Gimenez <raging@google.com>