Kleaf: kernel_env: set KBUILD_BUILD_USER=kleaf
This is to make a Kleaf-built kernel binary identifyable as such for
tracking purposes and for troubleshooting.

The KBUILD_BUILD_USER has the only requirement that it is constant and
consistent across builds in different workspaces on potentially entirely
different machines. Hence, it can be set to arbitrary values as long as
the above requirement is met.

Using the KBUILD_BUILD_USER has the advantage that we do not need to
find any other location to inject that information and it is already
readily available through /proc/version at runtime and using `strings`
for static analysis against a vmlinux binary.

Bug: 303658309
Change-Id: I52ed91031967b7ff47193e39af0da4d710d985b4
Signed-off-by: Matthias Maennich <maennich@google.com>