kleaf: Move default and versioned toolchains into @kleaf_clang_toolchain. (2nd attempt)
This reverts commit 99f28c07b7b8b9c70a51fd705952cbfc31387042,
along with fixes.

Reason for revert: relanding change.

The original change incorrectly flipped the order
of default and versioned toolchains, causing tests to fail.
This change rectifies the order.

Below is the original commit message.

Simplifies toolchain registration code that needs
to be moved to MODULE.bazel.

These toolchains are prefixed with 2 and 3, repsectively,
to indicate the registration order.

Test: TH
Test: bazel test //build/kernel/kleaf/tests
Bug: 276493276
Change-Id: I3ed339be2791b446ff7aace1c670fb01f1440705