Autoupdate of 2 submodules
*build/kernel updated to b626eb43be646d8efdfb57e7d832da39fd81284a
kleaf: Fix KBUILD_MIXED_TREE for --incompatible_sandbox_hermetic_tmp.

Previously without the flag, the TreeArtifact contains symlinks to
files under execroot, which is allowed.

With the flag, the symlink points to a file in the sandbox, which
is no longer valid (the sandbox is teared down) when the TreeArtifact
is validated.

Change it to a cp so that it is straightforward to humans.

Performance: There are no obvious performance regression associated with
this change. The KbuildMixedTree action is on the critical path.
This change reduces the time for the action from 4.1s to 3.6s in a
one-time experiment.

Test: TH
Bug: 319893821
Change-Id: I63f2364e0ffaa57a7ae57a1f7730f4b510ae3a8e
Signed-off-by: Yifan Hong <elsk@google.com>

*prebuilts/kernel-build-tools updated to 53024052da2ddf71b9cce9379f905d4dd68c262d
Update Bazel to 7.0.1rc1

Acquired from https://github.com/bazelbuild/bazel/releases/tag/7.0.1rc1
and verified using update-bazel.sh.

Change-Id: Ib0023a0ca8b2c15070d2de0faa68ed43e00840cd
Bug: 319893821
Signed-off-by: Matthias Maennich <maennich@google.com>