kleaf: checkpatch: Ignore BAD_REPORTED_BY_LINK
If you include a Reported-by: tag, checkpatch will
warn with BAD_REPORTED_BY_LINK if there is not a URL link
to where the report was made.

While these links are nice, and should be included where
possible, they shouldn't be required as often reports come
in private email, and we should make sure to credit reporters
even if there isn't a public URL for that report.

Change-Id: I1a06a5c28c3dd9f7b365b475cbb446ba6d48c13c
Signed-off-by: John Stultz <jstultz@google.com>