build_utils.sh: Ensure depmod runs on all modules last
When handling charger and recovery modules lists, depmod
is run on the set of charger and recovery modules and all
modules. However, the order in which this happens is such that
depmod is run on the set of charger modules, instead of all modules
last. This means that the final modules.dep file will contain
the dependencies for the charger modules, and not the rest of the
modules.

This is not ideal, as the modules.dep file should contain the
dependencies for all modules. Thus, ensure that depmod is run
on all modules last after running depmod on the recovery and
charger modules lists.

Bug: 322408856
Bug: 323710246
Change-Id: I7e03be016877df3b528b810735efec6c55fd0f6e
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>