ANDROID: KVM: arm64: Invoke the POST callback walker on the stage2 annotation install
Commit 5abca844a17bea847b151b3f9a893d315914084c introduces stage-2
coalescing for the host pagetables but the code path is not triggered
because of a missing flag that has been removed upstream.
Add back a missing flag for the walker to make sure that we zap the
non-refocunted page table pages on our way up. The zapping of the page
table pages into blocks happens when we donate the pages from (HYP ->
HOST).

Bug: 278749606
Bug: 325040553
Change-Id: I5aea0764f07baa90f84da3a5f7e0462ba6597ca8
Signed-off-by: Sebastian Ene <sebastianene@google.com>