Autoupdate of 2 submodules
*build updated to e46a95238deb7da627c5c12adef7090eae8646f3
Kleaf: switch to the builtin JDK that comes along with the bazel release

The most recent bazel version comes bundled with a suitable JDK and
hence we can drop any special handling for running bazel with a
different JDK.

Change-Id: Id106e9e822ef471442189f398f247d08ed933bbd
Signed-off-by: Matthias Maennich <maennich@google.com>

*prebuilts/kernel-build-tools updated to f9eb77f0d666eb501ba93b3aaa0112929406f0ae
Bazel: switch to binary with builtin JDK

That spares us some custom handling to run bazel with the local JDK. The
JDK comes GPLv2 licensed (with classpath exception) and this hence needs
to be treated permissive. Yet, the jdk comes bundled and can't be used
from anything else then the build tool (bazel) itself.

Change-Id: Ib43664524fe0551c1bb22cea4e996886f909c849
Signed-off-by: Matthias Maennich <maennich@google.com>