kleaf: Disable STG restriction of types to header files for Rust builds
The Rust compiler can generate types which do not have the
DW_AT_decl_file attribute in the DWARF dump for the type. Since STG
expects all types to have this attribute when file filtering is enabled,
STG does not support file filtering for objects compiled with Rust.

This is temporary change and should be reverted once STG can ignore file
filters for Rust types.

Bug: 325423395
Change-Id: Ica9d35619c47fbc02a338ed4bc6ec823da799c73