Include branch name in the directory microdroid kernel is downloaded to
In the trunk stable world different release configurations might use
different versions of the microdroid kernel (e.g. if we are in the
process of staged release of the new microdroid kernel). This means that
Android repo will need to have multiple different prebuilts checked in
at the same time. This patch adds support for such use case by adding
the short version of the kernel branch (e.g. android15-6.1) in the path
microdroid kernel prebuilts are stored at.

Bug: 298011555
Test: ./download_from_ci --update-microdroid 11448237
Change-Id: Ic6c06952a437c0585b8064c81d3fab8d785bcebf
Signed-off-by: Nikita Ioffe <ioffe@google.com>