kleaf: kernel_toolchain_ext.bzl use hard-coded //common:build.config.constants.
Make it non-configurable to make the implementation simple.
In the case where @kleaf is used as the root module (which
will become a legacy method after DDKv2 is completed), the
constant's path is hard-coded, and a symlink is needed.
If @kleaf is used as a dependent module, it is addressed
in the next CL.

This makes the implementation much simpler.

Bug: 293260835
Change-Id: Ia50b0e6f395750b458d3733a22dd79d823e38005