kleaf: Update kernel_prebuilt_repo to support local prebuilts.
This is to prepare for DDKv2 developement, where
GKI prebuilts are downloaded via extensions. In
addition, DDKv2 supports building against local
prebuilts, so support that.

This is a complete re-write of kernel_prebuilt_repo
while supporting the original API for legacy
WORKSPACE usage.

Caveat: This change disables parallel download. However:

- The list of release blockers for 7.1.0 indicates that it may take
  some time for the official release
- There are no known users for --use_prebuilt_gki on android15 branches
- Even if there are, the downsides are small; it makes a one-time
  download take longer. (However, downloads are subject to
  bandwidth constraints in the first place).

Test: TH
Test: bazel build @gki_prebuilts//... --config=internet --use_prebuilt_gki=11508345
Bug: 291918087
Change-Id: Ib2672e36f79709a2a0e0f035c0530aba52d736c0