FROMLIST: xhci: add helper to stop endpoint and wait for completion
Expose xhci_stop_endpoint_sync() which is a synchronous variant of
xhci_queue_stop_endpoint().  This is useful for client drivers that are
using the secondary interrupters, and need to stop/clean up the current
session.  The stop endpoint command handler will also take care of cleaning
up the ring.

Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>

Bug: 264371591
Link: https://lore.kernel.org/linux-usb/20240228013619.29758-2-quic_wcheng@quicinc.com/
Change-Id: I50d5d451fe10c9221a4965e9b933da531444b054
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>