Autoupdate of 2 submodules
*build/kernel updated to 6746ab1a241ad0900bbe459635b8846f51b08c3a
kleaf: Prevent the use of :hermetic-tools/<alias>.

This prevents the direct usage of e.g. :hermetic-tools/cp
and other aliases. These aliases have been deprecated for
some time. The full hermetic toolchain must be used,
because some tools rely on other tools (for example,
tar -z relies on gzip).

If you encounter any build issues, refer to
  build/kernel/kleaf/docs/hermeticity.md
for details.

In the future, we'll completely remove these aliases.
Now, we kept these aliases for a nice error message.

Test: TH
Bug: 291816237
Change-Id: I96ee4b61ae9a53eb601e0515322f38a1af109af2

*prebuilts/build-tools updated to 8e530191a520c9000e987657c981ff535af203b2
Add bison.runfiles_dir.

This is so that we can calculate the path to common/bison
in rule implementations. The actual directory is not
depended on.

Test: TH
Bug: 291816237
Change-Id: I303ee746153d983e25980a23961b51eb05ae05be