ReadFdToString: accommodate non-shrinking reserve
The file.ReadFileToString_capacity test verifies that
base::ReadFileToString shrinks an existing string's capacity to fit the
file. Previously, std::string::reserve was responsible for this, but as
of https://wg21.link/P0966R1, it no longer shrinks. Preserve the
existing behavior explicitly, and avoid reallocating when it isn't
needed.

Bug: b/175635923
Test: libbase_test
Change-Id: I0d77cf84527aa8209828914c7b665fcbbdf0e8b2