kleaf: use /proc/self/cwd as root directory in debug info
We use "-ffile-prefix-map" option to remove build directory from debug
info and to allow debugging on any machine with source code available.
But this rule was transforming an absolute path to a relative path,
which triggers a bug in DWARF and compilers, see
https://github.com/llvm/llvm-project/issues/83174.

"/proc/self/cwd" is an absolute path that resolves to a directory where
debugger runs, so debug information remains "relocatable" and the bug is
not triggered.

Original rules were relative to a "${{KERNEL_DIR}}", but in this change
"/proc/self/cwd" corresponds to the top repo directory, because:

* Users start Bazel in top directory, it would be convenient not to
  change directory when switching to debugger.

* Generated files live outside of "${{KERNEL_DIR}}", and in non-sandbox
  build (config=local or config=fast) they are accessible from the top
  directory.

* No need to have a separate rule to rewrite
  "${{ROOT_DIR}}/${{KERNEL_DIR}}" except for unsanboxed builds, where
  "${{KERNEL_DIR}}" is a symlink to a directory outside of
  "${{ROOT_DIR}}".

Test: TH, manual builds with config=local and config=fast
Bug: 315009365
Change-Id: I7727fa93ddf39301f9f755a748d2ef3ad91007bb