Autoupdate of 3 submodules
*kernel/tests updated to 0602e120661b839b40e7785e3ab9e36e7cbf1a36
kleaf: generate test_mappings.zip with pkg_zip.

... instead of an awkward test_mappings_dist rule.
This is so that all inputs are clear.

If more TEST_MAPPING files needs to be added, they
must be explicitly added to the list.

Test: TH
Bug: 330775243
Bug: 327493370
Change-Id: Ibd5999cd5ad04e7eeb0be71170a9817d0534e324

*external/zlib updated to fed9fb3ef0fd3b0e2928797933df0c8117b276a4
Merge "kleaf: Add pkg_files for TEST_MAPPING." into main
*prebuilts/rust updated to f940ef3280feafc8efc530ee6cdd9caf75a46876
kleaf: Add pkg_files for TEST_MAPPING.

This is so that it can be included
in test_mappings.zip.

Test: TH
Bug: 330775243
Change-Id: I25c21936e1de3559cd352b7ae7d6f2e3b9a59fd3