Autoupdate of 3 submodules
*common updated to ad8cdc64a377fde493d9d0bcdb093c736ddfed0e
ANDROID: Add pkg_files for TEST_MAPPING.

This is so that it can be included
in test_mappings.zip.

Test: TH
Bug: 330775243
Change-Id: I54c92953dab027ed68bb427adf5e08029f9d7799
Signed-off-by: Yifan Hong <elsk@google.com>

*common-modules/virtual-device updated to f160da93116319d54bf71638e40b3364a82ee1f0
Use //kernel/tests/test_mappings.

test_mappings_dist is deprecated.
Use modern //kernel/tests/test_mappings.

Bug: 329305827
Bug: 330775243
Change-Id: I54633b0ac63a50cb9054c6c9fb3219e7ed2d033a

*kernel/tests updated to 0602e120661b839b40e7785e3ab9e36e7cbf1a36
kleaf: generate test_mappings.zip with pkg_zip.

... instead of an awkward test_mappings_dist rule.
This is so that all inputs are clear.

If more TEST_MAPPING files needs to be added, they
must be explicitly added to the list.

Test: TH
Bug: 330775243
Bug: 327493370
Change-Id: Ibd5999cd5ad04e7eeb0be71170a9817d0534e324