ANDROID: riscv64: Enable RISCV_EFFICIENT_UNALIGNED_ACCESS
On x86, efficient unaligned access is fundamental, while on ARM
efficient unaligned access is turned on post armv6, as this was
architecturally efficient.

The patch in the RISC-V kernel that introduces this flag notes
that several RISC-V cpus now support efficient unaligned access, and
Android would like to only support RISC-V CPUs that support this
efficient unaligned access going forward.

Bug: 271046760
Test: Verified that the kernel build still boots on Cuttlefish.
Signed-off-by: Prashanth Swaminathan <prashanthsw@google.com>
Change-Id: Ic1ec74442c670b7dfce244af3d51842df1e4e154