Autoupdate of 4 submodules
*build/kernel updated to c1eb8c642e8a0a4a2a82daa466fe8dba7211f433
kleaf: common_kernels: allow to give extra_dist.

This is to provide extra dist targets that
are not known to the define_common_kernels()
macro, e.g. the test_mapping_zip target.

Test: TH
Bug: 327493370
Change-Id: I2d38edf36e1b14889304f3d2251a129c6541f419

*common updated to ad6707146e617c49cf016e8318a87e03a7bdc4e2
ANDROID: Move test_mappings_zip from //kernel/tests/test_mappings to //common.

This is to remove the hard-coded //common package
in //kernel/tests/test_mappings, because the kernel
tree may be specified elsewhere.

Test: TH
Bug: 327493370
Change-Id: Ib792be425a96b369eab6fb3d001fb1a519ec648b
Signed-off-by: Yifan Hong <elsk@google.com>

*common-modules/virtual-device updated to 9e755bb49aa7ae051564c8a4b1a276a1f54f9285
Move test_mappings_zip from //kernel/tests/test_mappings to //common.

This is to remove the hard-coded //common package
in //kernel/tests/test_mappings, because the kernel
tree may be specified elsewhere.

Test: TH
Bug: 327493370
Change-Id: Ief42393c07799652d01c83dfeefac5a7b3a034a0

*kernel/tests updated to 2d40c22c2ecc5738d093d31a48fab1fb6779405d
test_mappings: remove //common reference.

We do not want to hardcode //common for anything
part of Kleaf tooling, because ACK could be
checked out elsewhere.

Instead, create a pkg_filegroup so it can be
the pkg_zip can refer to them in the //common
package.

Test: TH
Bug: 327493370
Change-Id: Ic8dab284353b20befade6fe5c54f9813adde3e7c