Autoupdate of 2 submodules
*build/kernel updated to c1eb8c642e8a0a4a2a82daa466fe8dba7211f433
kleaf: common_kernels: allow to give extra_dist.

This is to provide extra dist targets that
are not known to the define_common_kernels()
macro, e.g. the test_mapping_zip target.

Test: TH
Bug: 327493370
Change-Id: I2d38edf36e1b14889304f3d2251a129c6541f419

*kernel/tests updated to 2d40c22c2ecc5738d093d31a48fab1fb6779405d
test_mappings: remove //common reference.

We do not want to hardcode //common for anything
part of Kleaf tooling, because ACK could be
checked out elsewhere.

Instead, create a pkg_filegroup so it can be
the pkg_zip can refer to them in the //common
package.

Test: TH
Bug: 327493370
Change-Id: Ic8dab284353b20befade6fe5c54f9813adde3e7c