kleaf: cc_testing: delete full_static_link binaries.
The feature name is fully_static_link, not full_static_link.
However, the clang toolchain is not set up properly that
fully_static_link works, because the libunwind dependency
is not properly handled here. With fully_static_link, this
gives:

  ld.lld: error: undefined symbol: _Unwind_SetGR

The proper resolution is probably:
  I697c10818cf8 "kleaf: Add -B / -L to host cc_* rules."
  I6e78006030fe "kleaf: set --target=x86_64-unknown-linux-gnu for host cc_* rules"
However, until we have a need for fully_static_link on this branch,
let's just delete the targets to avoid confusion.

fully_static_link is supported on android15-6.6 and above.

Test: build then readelf
Bug: 325514706
Change-Id: I3698f4efcf334b293c2c3d0f4642a58e493e2bb3