ANDROID: vendor_hooks: vendor hook for context switch
Add hook to record runtime context switch information to memory, including prv
task's pid/name and next task's pid/name. When system enter dead loop, we can dump
the memory to see which task the cpu core is running. It's a good way
to find out the problem in addition to logbuf.

Bug: 335604505
Change-Id: Icbda46f52c4a18bf89adfd356a7b19405924adf5
Signed-off-by: qiusanming <qiusanming@xiaomi.com>