kleaf: Simplify kernel_build.base_kernel toolchain version checking
... by removing the toolchain version file.

Instead of passing a file that contains the
toolchain version to kernel_filegroup,
just pass the toolchain version to the platforms
that kernel_filegroup takes, similar to how we
pass toolchain version to the intenral
_kernel_build() rule. kernel_filegroup infers
the toolchain version from the resolved toolchain
like kernel_env() does.
This simplifies KernelToolchainInfo and
therefore the checking code in kernel_build.

Note: This change forces kernel_filegroup_declaration
to declare a kernel_filegroup that always uses the
versioned toolchain so it is more explicit, even
if the original kernel_build uses the default toolchain.

Test: manual
Bug: 332386858
Change-Id: I1b0928015b2d1bf045fe76cc7e5b20254cd3d0a2