Only commit microdroid kernel prebuilt onces
Currently, both arm64 and x86_64 kernel prebuilts belong to the same git
project. This means that the commit message for the second path has
confusing empty gitlog inside (see aosp/3018344 and aosp/3018343 for
example).

This patch fixes the issue by committing the prebuilt only once. You can
check aosp/3053322 and an example output of the script.

Bug: 291071984
Test: ./build/kernel/gki/download_from_gki --update-microdroid
Change-Id: Iad1e37d9061fba259e28e264c3cdd4788c41f07f