Update config to build with updated boringssl
This CL fixes some configuration errors that arose with the updated
boringssl. This includes some linkage configuration errors, and some
temporary fixes required while `static inline` is still unsupported
in `rust-bindgen`.
We have
added the `unsupported_inline_wrapped` `cfg` to the `bssl-sys`
Android.bp file, to include the functions gated by this flag.

This CL also modifies the linker configuration to properly point to the
generated bindings file.

This is the AOSP changes required; there are some upstream boringssl
changes that add wrapper functions for some functions that are `OPENSSL_INLINE`.
These boringssl changes need to be pulled in for this to work as intended.

Test: m checkbuild
Change-Id: I0dbcfbbc03db2e1c3257a29ed71bfe3bc53179df