Implement multiport support in virtio-console
Port 0 remains as a `UCLASS_SERIAL udevice`, but when the driver
negotiates multiport support, the port 0 udevice additionally becomes a
bus for additional `UCLASS_SERIAL` port devices.

The control queues are stored in the port 0 `priv` structure, while
port-specific data is stored in a `plat` structure per port. Not sure if
this is an abuse of the `plat` system.

`drivers/virtio/virtio_console.h` is copied from the Linux kernel.

Also adds a `puts` implementation to write blocks of data rather than
individual characters.

Bug: 342507671
Change-Id: I07334123d29a4a3542abc087bf434001e0509391
Test: run with and without aosp/3101843