kleaf: log stgdiff report into a separate line
Users often copy the generated stgdiff report into their commit messages
when submitting updates to the stg files. Since the report is prepended
with a log level (e.g. ERROR, INFO) sometimes this gets unintentionally
transferred to the commit message too.

In order to make it easier for users to copy only the relevant bits from
the report, log the '{short_report}' summary into a separate line. This
makes the start of the report evident such as in the following sample:

  ERROR: ABI DIFFERENCES HAVE BEEN DETECTED!
  ERROR: Report summary:

  variable symbol changed [...]

Test: tools/bazel run //common:kernel_aarch64_abi_dist
Change-Id: Ic6f78631c5de50b0bbbe2aad73e4e541447873e1
Signed-off-by: Carlos Llamas <cmllamas@google.com>