kleaf: abi_update add --print_git_commands & deprecate --commit
The
  bazel run //common:kernel_aarch64_abi_update -- --commit
always prefers hermetic tools, but also falls back to
host git. If there are git hooks that invokes the hermetic
tools with unknown flags, the --commit feature will break.

To mitigate this, deprecate the --commit flag. Instead,
a --print_git_commands flag is added to print the Git
commands that the user may choose to run to create the
Git commit.

Bug: 341805657
Change-Id: I4b9991824249b58a3e0b5899c06cf7881e028b61