Properly enable 16k page size.
This allows us to check that the configs in the
corresponding fragments are actually enabled properly.

Other architectures will be fixed in other CLs. In
case of any issues, this allows for a smaller
revert.

Due to b/345544754, CONFIG_ARCH_MMAP_RND_COMPAT_BITS
and CONFIG_ARCH_MMAP_RND_BITS needs to be explicitly set.

This change also changes `16k_gki.fragment` to
the native kernel_build.page_size=16k. A notable
effect of this change is that CONFIG_LOCALVERSION
changes from 4k to empty, which is expected
because 16k_gki.fragment is outdated.

Test: TH
Test: bazel build //common-modules/virtual-device:virtual_device_aarch64_16k_config
  then examine .config before and after the change. After
  the change, the CONFIG_LOCALVERSION is correctly dropped.
Test: examine vermagic of output modules; there are no -4k suffix.
Bug: 347036722
Bug: 345544754
Bug: 346730877
Change-Id: Ie182c63c17e3692d5f4ac9dfc3c373bec4bae451