server use defconfig fragments.
This allows us to check that the configs in the
corresponding fragments are actually enabled properly.

This change also drops CONFIG_V4L2_MEM2MEM_DEV
from linux_distro.fragment. It was no-op because
dependencies aren't met.
v4l2-mem2mem.ko has not been in module_outs either,
indicating that the config is not enabled.

Other architectures will be fixed in other CLs. In
case of any issues, this allows for a smaller
revert.

Test: bazel build :server_x86_64_dist :server_aarch64_dist
Change-Id: I92ba8339ae4dc92dcf03de421a8024733a884307