Autoupdate of 2 submodules
*build/kernel updated to 1af4fcb58ec4723d34fe50199c3f37dc7c5e69e0
kleaf: clean up abs paths in .config. (android14)

With a patch that allows relative paths of
UNUSED_KSYMS_WHITELIST to be interpreted against
the output directory, this change is possible to
remove absolute paths from .config, thereby removing
ir-reproducibility.

UNUSED_KSYMS_WHITELIST is stored in $OUT_DIR
before kernel_config is archived by
kernel_filegroup_declaration().

MODULE_SIG_KEY and SYSTEM_TRUSTED_KEY are keys, so
they are not distributed; they are restored when
re-setting up the environment for kernel_config,
and skipped for kernel_filegroup.

Test: TH
Bug: 333769605
Bug: 342390208
Change-Id: I289563675e3d2ec182cf88801ee831ceda2ba3a9

*common updated to acacba8bfa1cadff3a692aa575f5bdd33dada24f
BACKPORT: FROMGIT: module: allow UNUSED_KSYMS_WHITELIST to be relative against objtree.

If UNUSED_KSYMS_WHITELIST is a file generated
before Kbuild runs, and the source tree is in
a read-only filesystem, the developer must put
the file somewhere and specify an absolute
path to UNUSED_KSYMS_WHITELIST. This worked,
but if IKCONFIG=y, an absolute path is embedded
into .config and eventually into vmlinux, causing
the build to be less reproducible when building
on a different machine.

This patch makes the handling of
UNUSED_KSYMS_WHITELIST to be similar to
MODULE_SIG_KEY.

First, check if UNUSED_KSYMS_WHITELIST is an
absolute path, just as before this patch. If so,
use the path as is.

If it is a relative path, use wildcard to check
the existence of the file below objtree first.
If it does not exist, fall back to the original
behavior of adding $(srctree)/ before the value.

After this patch, the developer can put the generated
file in objtree, then use a relative path against
objtree in .config, eradicating any absolute paths
that may be evaluated differently on different machines.

Signed-off-by: Yifan Hong <elsk@google.com>
Reviewed-by: Elliot Berman <quic_eberman@quicinc.com>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>

(cherry picked from commit a2e3c811938b4902725e259c03b2d6c539613992
 https://git.kernel.org/pub/scm/linux/kernel/git/mcgrof/linux.git modules-next)
Bug: 333769605

Change-Id: I0696ac8f686329795034ada5a4587af4ecbb774f
[elsk: apply change to gen_autoksyms.sh instead because
 CONFIG_UNUSED_KSYMS_WHITELIST is parsed there. Revert change
 to Makefile.modpost. Edit init/Kconfig instead of kernel/module/Kconfig
 because UNUSED_KSYMS_WHITELIST is defined there.]
Bug: 342390208
Signed-off-by: Yifan Hong <elsk@google.com>