Autoupdate of 2 submodules
*build/kernel updated to a9545d296a2bf68a3aa048f4761d565f0c0046b0
kleaf: Option to clean up abs paths in .config.

With a patch that allows relative paths of
UNUSED_KSYMS_WHITELIST to be interpreted against
the output directory, this change is possible to
remove absolute paths from .config, thereby removing
ir-reproducibility.

UNUSED_KSYMS_WHITELIST is stored in $OUT_DIR
before kernel_config is archived by
kernel_filegroup_declaration().

MODULE_SIG_KEY and SYSTEM_TRUSTED_KEY are keys, so
they are not distributed; they are restored when
re-setting up the environment for kernel_config,
and skipped for kernel_filegroup.

Because the patch is not universally applied on
the frozen braches android14-5.15 and android14-6.1,
this is made an attribute,
kernel_build.rewrite_absolute_paths_in_config,
so non-frozen branches with the patch can opt in.

Test: TH
Bug: 333769605
Bug: 342390208
Change-Id: Ic2b65b3e6867a661fa306b19125787a98badf228

*common updated to 3c19f7015ee303dd2171755eb764a33e63970f07
ANDROID: set rewrite_absolute_paths_in_config for GKI aarch64.

For GKI targets with kmi_symbol_list set,
also set rewrite_absolute_paths_in_config to True
so that CONFIG_UNUSED_KSYMS_WHITELIST is not an
absolute path any more, increasing reproducibility
on different machines.

This is made possible with c6427490abb1
"Revert^2 "BACKPORT: FROMGIT: module: allow UNUSED_KSYMS_WHITELIST to be relative against objtree.""
so that Kbuild recognizes relative paths by searching
$(objtree) first.

This does not change x86 builds because it
doesn't have kmi_symbol_list set.

This change does not affect device builds like
db845c, rockpi and fips140. Adding the attribute
to these targets would be okay, but it is not
covered by this change.

Bug: 333769605
Bug: 342390208
Change-Id: I844c251979bae5277474837eaa055f1346b0afa2
Signed-off-by: Yifan Hong <elsk@google.com>