ANDROID: GKI: init vendor data in __init_rwsem
Some semaphores are created before the VendorHook driver is
registered, so the value cannot be reliably initialized in the
trace_android_vh_rwsem_init() call. Inorder to use
android_vendor_data1 as a pointer, which point to our own
struct, we have to init android_vendor_data1 as 0. So we can
determine whether this android_vendor_data1 has been inited.
And in this way we will not have to add another
android_vendor_data.

Bug: 345653057

Change-Id: I5b61a30aad2dbff7e5feecd4782e50cf690175a6
Signed-off-by: seanwang1 <seanwang1@lenovo.com>