ANDROID: add kthread_stop_put to the symbol list for virtual targets
In commit a9da6ddaef9e ("kthread: add kthread_stop_put"), the macro
kthread_stop_put() was removed from the usb-ip driver as it was added
for other parts of the kernel to use.  But that requires that we add the
symbol to the symbol list in order for the usbip driver to properly
build.

Fixes: a9da6ddaef9e ("kthread: add kthread_stop_put")
Change-Id: Ic5a100159dc438a82a59d3f3394d1900b08912c1
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>