ANDROID: abi fixups for HCI_AMP removal changes
In commit 5af2e235b0d5 ("Bluetooth: HCI: Remove HCI_AMP support"), the
HCI_AMP support was removed from the kernel in order to address
CVE-2024-38620, but that broke a bunch of bluetooth api checks as
structures were made smaller.  Put those "empty" structures back in
order to preserve the ABI, but yet the logic that drove the removal is
still in place (i.e. the feature is gone.)

Bug: 161946584
Fixes: 5af2e235b0d5 ("Bluetooth: HCI: Remove HCI_AMP support")
Change-Id: If611dc20380848269a230bf8d5d2e1ee5a165d42
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>