ANDROID: Update .xml file du eto sruct clk_core change
In commit 4af115f1a20a ("clk: Get runtime PM before walking tree during
disable_unused"), the internal struct clk_core structure gets a new
field.  That disturbs the abi checker, as many apis use a pointer to
struct clk_core.  But because all callers treat this as an opaque
structure, and it is internal to the kernel core only, it is safe to
update the definition of it as there is no real "abi break" here at all.

type 'struct clk_core' changed
  byte size changed from 248 to 264
  member 'struct hlist_node rpm_node' was added
  30 members ('struct device_node * of_node' .. 'struct kref ref') changed
    offset changed by 128

Bug: 161946584
Fixes: 4af115f1a20a ("clk: Get runtime PM before walking tree during disable_unused")
Change-Id: Ic57abb142c040b62a006c5391f4c041c6e293c34
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>