ANDROID: virtio-dxgkrnl: fix mangled task_struct ptr
The dxgprocess struct currently keeps a pointer to the task_struct for
the thread that originally opened the device for that process. However,
a dxgprocess can outlive any particular thread, and if the timing is
right, a new process thread trying to open /dev/dxg can end up occupying
the same task_struct pointer memory of a previously closed thread. In
this case, the new process will end up re-using a dxgprocess for an
existing process instead of creating a new one.

This removes the reference to the task_struct entirely and just keeps
track of the tgid and the original pid.

Test: tested opening apps in an order that would repro this problem
Signed-off-by: Colin Downs-Razouk <colindr@google.com>
Change-Id: If1fe549392f775752fece7e64123cd849da35760