Autoupdate of 2 submodules
*build/blueprint updated to 134bf53915e4833ce915e31fd8fdadde6c5939ae
Support variable bindings in selects

This allows us to recreate soong config value variables in selects.

This adds a new "any" pattern to selects, which is the same as "default"
except that it doesn't match undefined variables, and it's (currently)
the only pattern that can accept a binding.

The syntax looks like:
```
select(soong_config_variable("my_namespace", "my_variable"), {
    any @ my_binding: "foo" + my_binding,
    default: "other value",
})
```

Bug: 323382414
Test: m nothing --no-skip-soong-tests
Change-Id: I4feb4073172d8797dee5472f43f9c248a76c3f1f

*build/soong updated to ab352c516a787dc397b5446b3292f51f0890f12d
Merge "Add tests for selects with bindings" into main