virtio-console: move `plat` data to `priv`
The original multiport console implementation stored device-specific
data in the `plat` field. I don't fully understand the implications of
using the `plat` field, but I now believe it is device-tree related. On
x86_64 the device tree is not used so it was relatively safe to store
arbitrary data in this field, but generating a device tree on arm64
possibly interacts with the `plat` data.

I don't have hard evidence of this, but it seems like a plausible
explanation. Whatever the true cause, this change fixes the multiport
console functionality on arm64 cuttlefish runs.

Bug: b/351052783
Test: multiport arm64 qemu cf
Test: single-port arm64 qemu cf
Test: multiport x86_64 qemu cf
Test: single-port x86_64 qemu cf
Test: single-port x86_64 crosvm cf
Change-Id: I94a603e67d3e51b591424b0e6bff1b95449262fa