kleaf: Allow generated headers in exported hdrs.
For ddk_module/ddk_submodule/ddk_headers, if hdrs contains generated
headers, include_dirs are automatically adjusted to also include
<file_root>/<workspace_root>/<package>/<include_dir>.

Implementation details:

This change uses file.root instead of genfiles_dir for each individual
files. Consider the following case:

ddk_header A
  depends on genrule G0
  depends on wrapper W1 (transition T1)
    depends on genrule G1
  depends on wrapper W2 (transition T2)
    depends on genrule G2

When running _ddk_header_impl for A, ctx.genfiles_dir applies to G0, but
it is incorrect to be applied on files from G1 and G2 because the
transition modifies the root of the file.

Bug: 349420132
Change-Id: Iae004b54e4ce9868fa6273f64a5e6b1c6f23455c