Fix Darwin Rust prebuilt builds
The error we were seeing where 'prebuilt_libstd' was relying on
a missing dependency 'prebuilt_libstd' for darwin-x86 was because Soong
produces an intervariant dependency between the arm64 and x86 targets
for Darwin so they can use the same underlying "universal binary" build
outputs. The result is that darwin-x86 modules have a dependency on
darwin arm64 modules.

The filtering added in aosp/3178079 means setting these prebuilts to
HostSupportedNoCross will avoid this dependency if HostCross is not
supported, thus removing the dependency on the  missing darwin arm64
variant.

Bug: 349612831
Bug: 351760165
Test: m libstd # on a darwin host
Change-Id: I4bd2a0f1640fef479522911a44d56a66fac3e07d