Autoupdate of 2 submodules
*build/kernel updated to 0590a755d6c865369822977ae028853ee632291d
kleaf: Add sysroot_dir.

This is a filegroup containing a File
which is a directory so we can infer its
path.

Bug: 228099181
Change-Id: Ia014f3e44876dfb0519ae03d0c3d4aced03d7367

*prebuilts/clang/host/linux-x86 updated to 89c6a0fd1d0c9e166225dca809ea21c15519cab1
kleaf: sysroot_path -> sysroot_dir.

This is so that we infer a path from File.path instead of
getting the error-prone string at analysis phase. This is
similar to how bin_dirs/lib_dirs are handled.

Bug: 228099181
Change-Id: I80db813e06ba6ff0b4be508a18122a9cad9bcdf3