kleaf: Add support for single_module builds
Allow the device kernel_build() macro to list the in-tree kernel modules
to build as part of the `make_goals` and skip compiling the GKI modules
that are defined in the gki_defconfig by removing `modules` from the
`make_goals`.

Due to issues in kbuild (even upstream), we need to build the in-tree
kernel modules with KBUILD_MODPOST_WARN=1 to avoid modpost errors when
in-tree modules depend on GKI modules. We need to follow-up on lkml. For
external modules, we can avoid using KBUILD_MODPOST_WARN=1 by stitching
together the base_kernel's Module.symvers with the device kernel's
Module.symvers, which allows the external modules to link with both GKI
and the device kernel's in-tree modules.

Bug: 323942195
Change-Id: I67ec05566970f37c4af69015ba89312d89e53299
Signed-off-by: Will McVicker <willmcvicker@google.com>