kleaf: platform_transition respects --host_platform.
If platform_transition sees attr.target_platform as None,
it will not apply any transitions. However, we are already in the
exec configuration from hermetic_tools (cfg="exec"), so the target
platform is already set to the execution platform.

Except for RBE and with register_execution_platforms, the only execution
platform is the host platform, aka the value of --host_platform.

By default, the value of --host_platform is @platforms//host.
@platforms//host is always linux x86_64 ATTOW.

With this change, platform_transition effectively uses --host_platform
(under cfg="exec" from hermetic tools) instead of the hard-coded
@platforms//host. This allows us to override --host_platform later.

Bug: 325514706
Change-Id: Id08d508d183eacfcee62ec3cff6a910bbbd2e64f