ANDROID: Convert build_config to a class in create-tracefile.py
create-tracefile.py supports --repo-dir followed by the root directory
of kernel source. By default, the script assumes the root directory is
its fifth parent directory. This argument lets the user run the script
outside of the source tree. To validate the paths under --repo-dir,
build_config is converted to a class. It checks the paths when they are
needed.

Test: ./create-tracefile.py -t logs --repo-dir src
Bug: 355559072
Change-Id: If9032f9706081bbf6e5898b913d9c2713a63659d
Signed-off-by: Hsin-Yi Chen <hsinyichen@google.com>