kleaf: create native_binary_with_arg
Delete hermetic_tools.extra_args and instead, let individual binaries
invoke with the default arguments properly.

This is done by introducing the native_binary_with_arg rule, which
is similar to embedded_exec, but with args provided to the
native_binary_with_arg target instead. It also has a much cleaner
implementation than embedded_exec.

The native_binary_with_arg wraps any binary so that it is invoked with
a given list of arguments. It doesn't require bash, but it does require
a functional CC toolchain.

Benefits of this change:
- Cleaner hermetic_tools implementation. We no longer handle extra_args
  in it.
- Allow us to apply transition on the whole tool (tar in this case).
  Otherwise, if arg_wrapper and tar_args.txt are in different
  configurations, arg_wrapper aren't able to find tar_args.txt.
- This is a good replacement of embedded_exec, so that copy_to_dist_dir
  can start to not depend on embedded_exec.

Bug: 325514706
Test: unit tests
Change-Id: Ie8793d6d065e73510724fef8368fccbb4953b5b5