kleaf: Add support to build flattened vendor_dlkm
Currently only the system_dlkm image can be built with the flattened
structure, i.e. without `uname -r` in the path. This feature simplifies
the process of loading the modules by the ueventd and thus should be
also implemented for vendor_dlkm. This solution is based on the existing
introduced for system_dlkm.

Test: Build and deploy {vendor,system}_dlkm.img
Bug: 355275989

Change-Id: Ibb2e3ba6a044929015045eea393bb43146d53be3