kleaf: Use @platforms//host in some places.
@platforms//host is strictly equal to HOST_CONSTRAINTS. Use
@platforms//host for inheritance instead of relying on the raw values.

For all intents and purposes, @platforms//host is linux x86_64. In
kernel_build(), just use @platforms//host instead of hard-coding the
values.

Note that @platforms//host does not change with --host_platform; in
fact, it is the reverse. --host_platform is by default @platforms//host
and may be changed to something else.

Bug: 325514706
Change-Id: I1c90e5cb0566185949a7caacae3fb2e1846f3627