Autoupdate of 2 submodules
*build updated to 272d52e8ee3245aa7d71307298a4f33fdce5ee13
kleaf: disable kleaf-host-cc for kernel_build().

The cc toolchain is now using kleaf-host-cc to govern all features
for C++ building, so we use that instead.

This allows us to disable the C++ features for both musl and non-musl.

Bug: 325514706
Change-Id: Ie59c302c8dc5f4e8eb98bebaadaf9898538c1471

*prebuilts/clang/host/linux-x86 updated to afbcbb58bfeb8ce914431ac16e89c1b41fc9b4cc
kleaf: refactor linux.bzl and add kleaf-host-cc feature.

This is a global feature flag that enables building C++ code.
This is needed to properly disable the feature when Kbuild
is used.

Bug: 325514706
Change-Id: If829b5a0eec87fa72f774a5b82e41622eebd7298