Autoupdate of 2 submodules
*build updated to 400debf41c89a66c3137131206430838835d65e1
kleaf: Add --musl_kbuild that kbuild builds musl host binaries.

When the flag is set, kbuild builds host binaries (fixdep, objtool,
etc.) against musl libc.

This is added to --config=musl and --config=canary to test that
GKI can be built with this flag.

Bug: 325514706
Change-Id: I75d3184d04fafcdca447b12e75c72356af0a5b58

*prebuilts-master/clang/host/linux-x86 updated to 9aff34cf939ab4a3a76ace923d80389168cf9681
Registers musl toolchain.

The toolchain allows building CC binaries and libraries against
musl libc. The toolchain is selected if the target platform has
//build/kernel/kleaf/platforms/libc:musl constraint value, which is
not the default at the time when this patch is submitted.

Note: the support for cc_* rules are incomplete by this CL. This change
only allows building kernel with the musl toolchain.

Bug: 325514706
Change-Id: Ibcbb74d13e0ccf35f4d57ddaf5ac68cd169bba19