Autoupdate of 2 submodules
*prebuilts/kernel-build-tools updated to 9ad5771a67a7140be423e981ebdef52b49844601
kleaf: expose libc_musl_file

So that the cc toolchain can use it in dynamic_runtime_lib

Bug: 325514706
Change-Id: I7cd9e9b80d69f55c1153c7effc3a3e3d8823ec10

*prebuilts/clang/host/linux-x86 updated to 9aff34cf939ab4a3a76ace923d80389168cf9681
Registers musl toolchain.

The toolchain allows building CC binaries and libraries against
musl libc. The toolchain is selected if the target platform has
//build/kernel/kleaf/platforms/libc:musl constraint value, which is
not the default at the time when this patch is submitted.

Note: the support for cc_* rules are incomplete by this CL. This change
only allows building kernel with the musl toolchain.

Bug: 325514706
Change-Id: Ibcbb74d13e0ccf35f4d57ddaf5ac68cd169bba19