ANDROID: arm64: dts: mediatek: mt8186: Fix eDP vs ps8640
The current Mediatek support in this tree seems to be a hodgepodge of
downstream dts forward-ported from an older kernel tree and stuff
that's landed upstream. As such, eDP panels on mt8186 devices using
ps8640 weren't working. This was because:

* The regulator node name for "ldo_vcn33" didn't match the regulator
  driver in this tree. Make it match, though it's still different than
  what's upstream.
* The panel hasn't been updated to "edp-panel" even though we have the
  patch that correctly moves "auo,b116xw03" to be a non-eDP panel. See
  commit ad3e33fe071d ("drm/panel: Move AUX B116XW03 out of panel-edp
  back to panel-simple")

Let's fix these two things which seems to make the panel come up on
tentacruel.

NOTE: This is just piling up another patch on the current house of
cards. The right solution for corsola support in this tree is to
delete downstream corsola support in this tree, re-backport it from
upstream, and then add any extra patches for functionality that's
missing (hopefully nothing or almost nothing).

Bug: 350646423
TEST=Boot tentacruel

Fixes: 5d0ea1539ead ("ANDROID: mt8186: enable auto-configuration of eDP panel")
Change-Id: I6520e6f3e9ee36412713398b9390d63630c02f42
Signed-off-by: Douglas Anderson <dianders@google.com>