kleaf: swig_wrapper uses runfiles.
Relies on runfiles_dir to find the runfiles when the executable is
executed at `bazel run`, e.g. when `bazel run xxx_config`. This is the
recommended way.

With this change, `bazel run //build/kernel:swig` works, though this is
not a public API and is not how swig is supposed to be invoked.

Bug: 325514706
Change-Id: Ife8feade309a2d405a4bf97db207d8b1c82b7acb