kleaf: Pass RAMDISK_COMPRESS/RAMDISK_EXT to build_boot_images.
It was previously incorrectly missed. If the build config
does NOT specify RAMDISK_COMPRESS/RAMDISK_EXT, building boot images
may break.

This is hit when
- kernel_images(build_initramfs = True)
- INITRAMFS_VENDOR_RAMDISK_FRAGMENT_NAME is empty
in which case build_boot_images will try to build
ramdisk.{ramdisk_compress}; but it may be unable to do so when
RAMDISK_COMPRESS/RAMDISK_DECOMPRESS/RAMDISK_EXT is not set in the build
config but only in BUILD files.

Bug: 361733833
Change-Id: If1fd54a9197404b2e4112e0b2d8a169caf2742af