ANDROID: arm64: mt8186: dts: Remove clocks from USB power domain node
The upstream version of DTS for mt8186 has introduced clock entry
inside power domain nodes. This is causing issues with XHCI during
suspend.

Note:
 This patch is a temporary revert of a00d4a98af44. The final fix
 will be done on upstream and then integrated with the AOSP.

Both domain and XHCI driver activate clocks which results in
clock_refcount to be 2. During suspend only XHCI is disabled
and only XHCI releases clocks. Power domain is still active
because it is a parent of wakeup-source (XHCI).
The ATF prevents the SoC going to S3 state as it check that clock
were not released (because its clock_refcount == 1). EC detects
sleep transition timeout and resets the board.

Bug: 351926349
Test: run powerd_dbus_suspend
Change-Id: I904810bb324956f812d24cec5308fe919b1eb943
Signed-off-by: Wojciech Macek <wmacek@google.com>