dist: warn for providing options through cmdline
With this change, most options set through the command line will trigger
a deprecation warning. Options specified in copy_to_dist_dir() in BUILD
file do not trigger the warning. This is a step towards having a similar
command line interface as pkg_install(), eventually allowing us to move
to pkg_install().

This excludes options --destdir and -q, which are provided in
pkg_install().

Bug: 331730853
Change-Id: I351a11514064750153b5f5fbed1b7f5ed70c3c85