ANDROID: Make build-kernel.sh script idempotent
In current state of the art, running build-kernel.sh twice on arm64
results in a broken Image file - in one of the steps, we replace kernel
image with a FIT image, which extends the kernel with devices trees.
Running the script twice creates a nested FIT FIT image, which does not
boot. We already had a regression because of this on corsola devices,
which resulted in P0 issue (b/363260874)

Bug: 364242800
Test: Ran the script a few times in a row, compiled the image (m) and
booted on Kyogre and Chinchou devices.

Change-Id: I6fd1f375a724c4e1d8bb0c92b395af6eca6f8f8c
Signed-off-by: Mateusz Urbańczyk <urbanczyk@google.com>