kleaf: ddk_module.linux_includes are useful with submodules
If the top-level ddk_module defines linux_includes, or if its
dependencies (kernel_build, hdrs, deps, textual_hdrs) specifies
linux_includes, it was the case that these linux_includes are not
applied to individual submodules. This change fixes the issue.

In particular, for common ddk_headers inherited from kernel_build,
this allows submodules to also see the ddk_headers.

Bug: 363143263
Change-Id: Ibde3da4d5391a9c880f05f13d5a0770d6f4515bb