ANDROID: rust_binder: use AllocationView for copy_into
The use of `copy_into` that copies data between objects uses `alloc`
directly, rather than using the `view` which provides bounds checks for
accesses to the data part of the transaction data. Introduce a
`copy_into` method for the view type and change the conversion code to
use it for bounds checks.

This fixes an equivalent to commit 4df153652cc4 ("binder: fix UAF caused
by offsets overwrite") in Rust Binder.

Change-Id: Ic7a78c95f16967fbfc2e95fc7dd741151a5de6ae
Signed-off-by: Alice Ryhl <aliceryhl@google.com>