Create priority inversion benchmarks
Create .ditto files that are able to replicate the priority inversion
phenomenon.
These benchmarks create 3 threads with different priorities.
The high priority thread shares a mutex with the low priority thread.
The mid priority thread is supposed to preempt the low priority thread,
that is holding a resource that the high priority thread is blocked on,
thus indirectly preventing the high priority thread from running.

Test: cmake tests
Bug: 365781383
Change-Id: Ib956ae12be2b6f5b1af4d70aa79ad38a3b6a41ab
Signed-off-by: Alessio Balsini <balsini@google.com>