what_commits_per_release: add script to find the subset of commits for a merge
For LTS merges, it is nice to see only the commits that a specific git
range would be relevent.  This script prints out the commits when
comparing it to a list of files used to build a kernel.

Taken from a very old python2 script used for Pixel LTS merges, but
rewritten in sh to make it simpler and actually work on modern systems.

The output format might be tweaked over time, it is not stable right now

Bug: 364985605
Change-Id: I48dce04d962add1abc024d162f181367463a636e
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>