CPU absolute duration workload
Ditto already provides instructions to generate CPU workloads either
with a fixed amount of iterations (the total duration of which depends
on other factors, such as the CPU architecture and frequency), and a
fixed CPU bandwidth given a period.
There are cases where it's handy to force a CPU workload that lasts
exactly an absolute amount of time. This is what this patch introduces.

Bug: 365781383
Test: Manual verification with Perfetto
Change-Id: Id83497c056b8d92eed44aec3df52ff0dfc01eb17
Signed-off-by: Alessio Balsini <balsini@google.com>