kleaf: Fix KERNEL_DIR deprecation warning
KERNEL_DIR was not set at this point because build config is not sourced
yet.

Delete the old warning. Instead, emit an error if the value of
kernel_build.makefile is inconsistent with KERNEL_DIR.

Also add a deprecation warning if kernel_build.makefile is NOT set to
encourage people to set it explicitly.

Bug: 236012223
Change-Id: I2b0ac75bb38ea611b6c733bf7811130324b7d051