ANDROID: GKI: Add to task_struct size via cmdline
To reduce the size of vendor data allocated in the task_struct, from 512
bytes to a significantly lower 48 bytes, the move to a dynamically sized
task_struct is being made.
As part of this effort, provide means for vendors to pass a size value
via kernel cmdline. Use the passed value to dynamically add to the
task_struct size to accommodate vendor data.
The cmdline parameter to be used is 'android_task_struct_vendor_size'.
For eg., vendors can add the following to the bootargs section of their
devicetree to add an extra 512 bytes to the task_struct:
"android_task_struct_vendor_size=512"
To access this additional memory, use the android_task_vendor_data
function provided.

Bug: 233921394
Change-Id: I6d5ab92080b82f29bbe9735d40f7d0b1e5bb5913
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>