kleaf: _setup_env.sh no longer override PATH.
Kleaf already provides the correct PATH to _setup_env.sh,
so there is no need for _setup_env.sh to do this.

As a side-effect, menuconfig no longer needs run_additional_setup
to re-instate PATH, because it is not overridden.

As a side-effect, HERMETIC_TOOLCHAIN variable no longer has any
effect.

As a side-effect, we no longer need the trick of
KLEAF_INTERNAL_BUILDTOOLS_PREBUILT_BIN.

Bug: 364337887
Change-Id: I118f4e0fd5c1d90912b31140547f8389e5650c70