ANDROID: arm64: call vendor hook in reworked handers
commit cd5ceadc2b37 ("arm64: rework FPAC exception handling") and commit
474385adcd84 ("arm64: split EL0/EL1 UNDEF handlers") have seperated EL0
and EL1 handler. Then, vendor hooks are no longer being called from EL1
handlers.

Call the vendor hooks from EL1 handler to achieve the purpose of them.

It helps the bootloader produce the summary information when
kernel panic or watchdog reset occurs. It can be obtained through
Ramdump, etc., and this information can be used to quickly classify
and respond to problems.

Bug: 222638752
Bug: 369039214
Change-Id: Ia7c0b007f4548cb0263a3b65ad766e636e3f68dd
Signed-off-by: Sangmoon Kim <sangmoon.kim@samsung.com>