ANDROID: Use DTC from PATH. (2nd attempt)
This is a re-apply of 2a2cffe9e206
("ANDROID: Use DTC from PATH.") that adds DTC from PATH. Because
checkpatch.sh and checkpatch_presubmit.sh no longer exists, and it is
not allowed to source _setup_env.sh directly, this change can be
re-applied.

The warning message has been there for a year now, so it is okay to
remove it so that, if dtc is not found in PATH, $DTC is left empty.

Below is the original message.

ANDROID: Use DTC from PATH.

Instead of using a hard-coded path to DTC, infer its
path from PATH. This change removes the dependency on
the ${BUILDTOOLS_PREBUILT_BIN}/dtc symlink, so the
build infrastructure is free to reset PATH to a place
where the actual DTC binary exists.

This change is needed to set
--incompatible_disable_hermetic_tools_symlink_source
to true by default.

Test: bazel run //common:db845c_dist
Fixes: a71e480e82a7 ("ANDROID: build.config: Set DTC to absolute path")
Bug: 291816237
Change-Id: I4f6051dc079a33c8e5e2632aa7ed040049300601
Signed-off-by: HONG Yifan <elsk@google.com>