ANDROID: Allow MODVERSIONS + Rust with no symbol export
We have two patches [1] [2] which provide proper MODVERSIONS support,
with no current objections. Since they're not in yet though, this is the
lowest impact way to allow the new `ashmem` driver to be tested, as it
does not need module support. Once these are merged upstream, we can
back this out.

[1] https://lore.kernel.org/all/20240925233854.90072-1-mmaurer@google.com/
[2] https://lore.kernel.org/all/20241008183823.36676-21-samitolvanen@google.com/

Bug: 347787665
Change-Id: I979def75f67d1e140253795c0fc706133d23e506
Signed-off-by: Matthew Maurer <mmaurer@google.com>