Suggest an alternative for tags that are not on any branch
The release tags of most Google-owned repositories under github/google
do not belong to any branch on the repository and may belong to a fork
outside of the repository. This breaks external_updater because of lack
forge committer permission error. In these cases, we want to merge with
the parent of the tag. These tags are identical to their parents.

Bug: 370077034
Test: TreeHugger
Change-Id: Ic1307efd64a4fbb0ddb2955ceb76eac52e7d7914