kleaf: modules_prepare hack to make fixdep work
fixdep in the modules_prepare archive is built from the GKI source tree
and has runpath to be:

    $ORIGIN/../../../../prebuilts/kernel-bulid-tools/...

This does not work when fixdep is executed in the DDK workspace where
@kleaf is not the root module.

To work around this, temporarily employ a hack that symlinks
prebuilts/kernel-bulid-tools to <execroot>/<workspace_name> (i.e. PWD).
This is so that the above runpath still works.

Bug: 372807147
Change-Id: Id2ac3233b5b8646f6b983b59aedbe9538074d080