Autoupdate of 2 submodules
*build updated to d07fe34b97ebebb2f786749d7155bdd1b1b56957
kleaf: set ARCH if not already set.

This is so that setting it in build.config is no longer necessary.

Bug: 236012223
Change-Id: Ibf7da3756d4f6cea9e4118725e138b610059ea4a

*common updated to 2e0cceb846cfd43b4752e3df1c727866f5d6bd40
ANDROID: Delete ARCH from build.config.

We have enforced kernel_build.arch to be the same as ARCH in
build configs for quite a while, so there is no need to set
it in build configs, but let Kleaf infer from kernel_build.arch
instead.

Bug: 236012223
Change-Id: I40a832a175528ed1472811a712073557543ff2d4
Signed-off-by: HONG Yifan <elsk@google.com>