Expand runpath for kbuild Rust host tools
When using the LLVM from clang, `generate_rust_target` ends up with a
dependency on `libunwind.so`. For a normal kernel build, this is a
non-issue, but the Kleaf hermetic build prevents this library from being
found.

This makes it so that the Rust prebuilt libdir is added to the runpath
for all Rust host tools generated by kbuild, causing `libunwind.so` to
resolve.

Bug: 373554783
Change-Id: I13c3fb15e45e4a929909adecdab59eb846938785
Signed-off-by: Matthew Maurer <mmaurer@google.com>