ltp: syscalls: pipe2: Update pipe size check
If the pipe size is set to a value less than the page size. The
kernel rounds the size up to 1 page.

In the case of Android's x86_64 16KiB emulator this is rounded up
to the 4KiB kernel page size.

Update the test to handle this case. This is a no-op on all other
targets.

Bug: 366098040
Test: atest vts_ltp_test_x86_64:syscalls.pipe2_04_64bit#syscalls.pipe2_04_64bit
Change-Id: I78fce55fef53c0f628fc5aa5789fddd7f16fbefb
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>