kleaf: Fix DDKv2 with separate kernel source tree
With DDKv2 set up, i.e. @kleaf is a dependent module, if a
a kernel_build() sets KERNEL_DIR to a separate kernel source tree to
build kernel modules or dtbs, it should be allowed. Before this change,
it was incorrectly assumed that the KERNEL_DIR must be below @kleaf.

Implementation details: This change makes it so that we ONLY fix up the
value of KERNEL_DIR from downloaded prebuilt scripts when using
kernel_filegroup(). When using kernel_build(), we assume that the user
has provided the correct KERNEL_DIR, so we don't change the value.

Test: integration test
Bug: 375647893
Fixes: 6af2b8a169e7
    ("kleaf: Re-declare KERNEL_DIR under KLEAF_REPO_WORKSPACE_ROOT (2nd attempt)")
Change-Id: I565c0e95afbde412fb664458023260c31eedff56