kleaf: Don't use hardcoded path for GKI boot.img certification key
Currently, the path to the GKI boot.img certification key is hardcoded
in build_utils.sh. However, when building with DDKv2 with kleaf as
a dependent module residing in external/kleaf, the path to the key
is not the same as what is hardcoded in build_utils.sh.

Therefore, set a variable that can be read to get the path to the
GKI boot.img certification key.

Bug: 369906147
Bug: 373908709
Bug: 375675093
Change-Id: Ic4f480c3c8b008483a51846c19e629f9d02e5859
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>
(cherry picked from commit d471cabb355e7c6382e547326d5c454913abc86c)
[isaacmanjarres: Android 13 builds don't always use kleaf, so use
a fallback path for the certification key.]