ANDROID: rust_binder: fix type confusion when unmounting binderfs
The binderfs directory holds two kinds of files:
* the binder-control file
* contexts (e.g., binder, hwbinder, vndbinder)
When binderfs is unmounted, the binderfs_evict_inode function is used to
clean up both kinds of files. However, binderfs_evict_inode always uses
rust_binder_remove_device to clean up the private data of the file,
which only makes sense for the contexts. This results in type confusion
when unmounting binderfs.

Revert back to the same approach as C binderfs which uses a `struct
binder_device` for both kinds of files. Furthermore, the functions that
Rust exports to deal with the context are renamed to "context" to
emphasize that they are used only in the context case. The
binder-control file will hold a null pointer for the context.

An alternate approach would be to get rid of `struct binder_device` and
use a null pointer for the private data of binder-control. However, this
makes it hard to avoid leaking its minor number in binderfs_minors when
unmounting binderfs.

This bug is not present in C binderfs.

Reported-by: Domen Puncer Kugler <domen.puncerkugler@nccgroup.com>
Fixes: 0d512d37b077 ("ANDROID: rust_binder: add binderfs support to Rust binder")
Change-Id: I0084e37525eaf1412cb8af18fc7abdc7ba1016b9
Signed-off-by: Alice Ryhl <aliceryhl@google.com>