kleaf: kernel_env run_env is from regular env
This change is so that run_env is identical to the regular env setup
script. This avoids adding a separate file for `bazel run` environment,
causing a lot of places to be updated.

The toolchain is not properly set up in the run env before and after
this change.

Because run_env is identical to the regular KernelEnvInfo, delete it
everywhere.

Bug: 370450569
Change-Id: I9487bd51ca5fdf4caa27ef4f339f4bf93b64cc46