kleaf: hermetic_tools run_setup is deprecated.
Merge hermetic_tools.run_setup with setup, and deprecate run_setup.
Instead, at execution phase, use BUILD_WORKSPACE_DIRECTORY and
BAZEL_TEST to determine whether short_path or path should be used.

This is so that we don't have to fork the command line in the analysis
phase for different environments.

Bug: 370450569
Change-Id: I7b69d8cfd25a35a5bc30802c5141f9e6a28932e9