kleaf: ddk_headers: Set allow_files = True for hdrs and deprecate textual_hdrs
In preparation for adding --user_ddk_unsafe_headers, allow (empty)
filegroups to be added to the hdrs input list. When adding an empty
filegroup, bazel complains that the filegroup doesn't provide any .h
files.

This issue is discussed in:
https://github.com/bazelbuild/bazel/issues/7516#issuecomment-558288702

The recommendation is that "allow_files" is errorprone itself and we
should not use it in this case. If really restricting to .h files is
important, we can add the restriction in the rule implementation and
properly cover the filegroup.

textual_hdrs has identical functionality as hdrs now. Deprecate
textual_hdrs.

Bug: 376318144
Signed-off-by: Elliot Berman <quic_eberman@quicinc.com>

[elsk: partial cherry-pick of the original change because textual_hdrs
 does not exist on main-kernel-build-2023.]
Change-Id: I1fdf421dafee6fdde82d7d37cd4fb40385f12e7f
Signed-off-by: HONG Yifan <elsk@google.com>