kleaf: optimize ddk_config time when defconfig/kconfig is not set.
In the common case when defconfig/kconfig is not set, it is previously
attempted to skip `make olddefconfig`. However, the `test -s` check is
not effective, because `utils.write_depset()` always writes at least
a new line to the file. Hence, use a better check.

This saves a 4 seconds on my machine for a ddk_config that does not
have defconfig / kconfig.

Bug: 370681291
Change-Id: I959925187403a55d2c65a9f2b2920fdf25e0402b