Autoupdate of 2 submodules
*build/kernel updated to a4ee864c7a67db82071025e55313f9ec7681db7d
kleaf: build proc_macros against glibc.

Even with --config=rust_platform or with --musl_kbuild,
libmacros.so needs to be built against glibc because the
rust toolchain itself is built against glibc. Hence, also
get the flags to build host stuff against glibc, then
set PROCMACROLDFLAGS.

Bug: 373900675
Change-Id: Ia9c4003da9f41676c67b31a5de342c4c55737ef5

*common updated to 5ff498e7bc1e5f1427ee4e86c34d58f38baafa43
FROMLIST: rust: add PROCMACROLDFLAGS

These are additional flags to be passed when linking proc macros for the
Rust toolchain. If unset, it defaults to $(HOSTLDFLAGS).

This is needed because the list of flags to link hostprogs is not
necessarily the same as the list of flags used to link libmacros.so.
When we build proc macros, we need the latter, not the
former. To distinguish between the two, introduce this new variable
to stand out from HOSTLDFLAGS used to link other host progs.

Signed-off-by: HONG Yifan <elsk@google.com>

Bug: 373900675
Link: https://lore.kernel.org/r/20241017210430.2401398-2-elsk@google.com
Change-Id: Ifc678501880633e1074fea24bf1c4ec563b07a7e
Signed-off-by: Alice Ryhl <aliceryhl@google.com>