kleaf: Add support for extracting prebuilt signed GKI artifacts
The signed prebuilt boot.img is currently stored in the prebuilts
directory under the signed/ directory as part of a tarball. However, it
is desirable to have the signed boot.img readily available without
having to decompress that tarball.

Therefore, add support to extract the signed GKI artifacts so
that they can be referenced.

Bug: 335043058
Bug: 373908709
Change-Id: I9e62be9d8fb0b0a4aefcddf60ca13bce0acb8a30
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>