kleaf: DdkWorkspaceSetupTest allow interactive mode.
The test creates an unshare mount namespace with directory structure
remounted. It is usually very difficult to picture what's going on
within these directories. This change adds a -i option to the
integration test. When specified, DdkWorkspaceSetupTest starts
up an interactive shell after setting up the workspace but before
running any Bazel commands to build things.

To start a DDKv2 setup with common source tree in @kleaf, run

    bazel run //build/kernel/kleaf/tests/integration_test -- \
        -i DdkWorkspaceSetupTest.test_kleaf_module_below_ddk_workspace

To start a DDKv2 setup with prebuilts, run

    bazel run //build/kernel/kleaf/tests/integration_test -- \
        -i DdkWorkspaceSetupTest.test_setup_with_downloaded_prebuilts

To start a DDKv2 setup with both prebuilts and a kernel source tree, run

    bazel run //build/kernel/kleaf/tests/integration_test -- \
        -i DdkWorkspaceSetupTest.test_setup_with_local_prebuilts

Test: bazel run //build/kernel/kleaf/tests/integration_test -- \
    -i DdkWorkspaceSetupTest.test_kleaf_module_below_ddk_workspace
Change-Id: I7a502105b1c163cc2f53999c568c76ac115a6cf9