Autoupdate of 2 submodules
*build/kernel updated to 38b3fa294ebf9660353ef2848cdc3c61d5f54f67
kleaf: Add test on kernel_aarch64_dist with DDKv2 setup.

Check that the full GKI dist can be built properly.

Bug: 377799264
Bug: 377799294
Change-Id: I67ae21237a5c05d36d78e9e33e0fd3c07231fed8

*common updated to 939e9e4aa18e81276e862e3109d460a09af1ece1
ANDROID: Fix platforms config_setting values.

The values must be wrapped in a str(package_relative_label()), otherwise
they are not resolved to the correct label especially when @kleaf is not
the root module.

Bug: 377799294
Change-Id: I100929bace62a3268af197d6fab462c89cf5fe5f
Signed-off-by: HONG Yifan <elsk@google.com>