Autoupdate of 2 submodules
*build/kernel updated to 074f1df1b8a2c6aa38dff7d0ab017f277f9940ec
kleaf: Add test on kernel_aarch64_dist with DDKv2 setup.

Check that the full GKI dist can be built properly.

Bug: 377799264
Bug: 377799294
Change-Id: I67ae21237a5c05d36d78e9e33e0fd3c07231fed8

*common updated to 5c5911ff0ad7522dad1b446eef05090d14ae7ef8
ANDROID: Fix platforms config_setting values.

The values must be wrapped in a str(package_relative_label()), otherwise
they are not resolved to the correct label especially when @kleaf is not
the root module.

Bug: 377799294
Change-Id: I100929bace62a3268af197d6fab462c89cf5fe5f
Signed-off-by: HONG Yifan <elsk@google.com>