Merge 88ff021e1fea ("f2fs: Require FMODE_WRITE for atomic write ioctls") into android14-5.15-lts
Steps on the way to 5.15.168

Resolves merge conflicts in:
	fs/f2fs/file.c

by just merging commit 88ff021e1fea ("f2fs: Require FMODE_WRITE for
atomic write ioctls") out of the way entirely.  If this needs to come
back in the future, it can do so in a merge-safe way, the f2fs codebase
has diverged so far from upstream in this branch, that merges here are
almost impossible to make anymore.

Change-Id: I000851e6e5899799e8d01e344f5b2f086690d2cc
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>