BACKPORT: usb: dwc3: core: Set force_gen1 bit for all applicable SuperSpeed ports
Currently if the maximum-speed is set to Super Speed for a 3.1 Gen2
capable controller, the FORCE_GEN1 bit of LLUCTL register is set only
for one SuperSpeed port (or the first port) present. Modify the logic
to set the FORCE_GEN1 bit for all ports if speed is being limited to
Gen-1.

Suggested-by: Thinh Nguyen <Thinh.Nguyen@synopsys.com>
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>
Acked-by: Thinh Nguyen <Thinh.Nguyen@synopsys.com>
Link: https://lore.kernel.org/r/20241112182018.199392-1-quic_kriskura@quicinc.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(cherry picked from commit ce25e2a8d82de7f8bcbedd18973e5b8030749d45
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git master)

Bug: 379220296
Change-Id: I4b2b12ea343aa1f8a97c812815b9d39d8b2b836f
[quic_kuruva: Minor correction, 'num_usb3_ports' is part of the
DWC structure in upstream. However, it has been acknowledged that
it has moved to dwc3_vendor to avoid an ABI break,
replacing 'dwc' with 'vdwc' in the change.]
Signed-off-by: Rajashekar kuruva <quic_kuruva@quicinc.com>