ANDROID: GKI: x86: Enable X86 userspace shadow stack
Shadow stack protection is a hardware feature that detects function
return address corruption. This helps mitigate ROP attacks.
Applications must be enabled to use it, and old userspace does not
get protection "for free".

A shadow stack is a secondary stack allocated from memory which cannot
be directly modified by applications. When executing a CALL instruction,
the processor pushes the return address to both the normal stack and the
shadow stack. Upon function return, the processor pops the shadow stack
copy and compares it to the normal stack copy. If the two differ,
the processor raises a control-protection fault.

NOTE: It is safe to enable this security feature since it doeas not
      change anything by itself. It's not active unless additional
      consitions are met [1] and can be explicitly disabled at runtime
      via kernel boot args [2].

[1] To use userspace shadow stack you need HW that supports it, a kernel
    configured with it and userspace libraries compiled with it.

[2] The kernel Kconfig option is X86_USER_SHADOW_STACK. When compiled in,
    shadow stacks can be disabled at runtime with the kernel parameter:
    nousershstk.

At run time, /proc/cpuinfo shows CET features if the processor supports
CET. "user_shstk" means that userspace shadow stack is supported on the
current kernel and HW.

NOTE: The x86 ABI is under construction and it is expected to have
incompatible updates on the android15-6.6-desktop branch.

Bug: 391769036
Change-Id: I8a512fc1d1a2f8fddaf3c61239a18da026724e81
Signed-off-by: Slawomir Rosek <srosek@google.com>
Signed-off-by: Jeff Xu <jeffxu@google.com>