ANDROID: virtio-dxgkrnl: let dxg_map_iospace propagate the error to the caller
When dxg_map_iospace fails with either vma merged or failed to grab the
mm write lock, it's not fatal, and the caller should be allowed to retry
in those cases. This CL propagates the error code to the caller so that
in the future, if we see the caller fails in non-fatal cases, one of the
callers in the stack can retry.

Change-Id: Iedf7a4878ceacf10162bdbc28da09d97a5e607f8
Signed-off-by: Kaiyi Li <kaiyili@google.com>