ANDROID: rust_binder: don't use mmap write lock
The mmap write lock was dropped from the mm series. Update the
page_range abstractions to use the read lock instead and introduce a
mutex for the cases that require exclusive synchronization.

The mmap lock is very very contended, so removing a use of the write
lock is a good idea anyway.

Change-Id: I383cce17fa0c89320ad224091c964ed76f4a20bc
Signed-off-by: Alice Ryhl <aliceryhl@google.com>