ANDROID: iommu/arm-smmu-v3-kvm: Support command queue batching
Similar to the kernel driver, we can batch commands at EL2 to avoid
writing to MMIO space, this is quite noticable if the SMMU doesn't
support range invalidation so it has to invalidate page per page.

Doing a smoke test on Qemu with various device (MMC, virtio-net..)
shows number of prod writes dropped by 50% while cons reads remains
the same, this because of the batching of sync commands as Qemu supports
RIL, for HW without that reduction would be much greater.

Bug: 357781595
Bug: 384432312

[@smostafa port to android17-6.18]
Change-Id: I8bc4ad1448add1d43c6320cc79531929c129ded2
Signed-off-by: Mostafa Saleh <smostafa@google.com>