rust: Add rustc rpath to PROCMACROLDFLAGS
Proc macros are currently built without an explicit runpath. They have a
dependency on `libunwind.so` which is not on the default search path for
kleaf. `rustc-1.81` depended on `libunwind.so` directly, so it was
preloaded when it went to `dlopen` proc macros. The `rustc-1.82` binary
doesn't have this dependency anymore, so we need to adjust the runpath
to ensure that proc macros can find `libunwind.so`.

Bug: 384734194
Test: With 1.82, tools/bazel build //common:kernel_x86_64 --rust
Change-Id: I1665e65fb998df1370fa836cf380f7bbfc22a318