Add a flag to control extracting the symbols
Since extracting the symbols requires building the kernel with notrim
and with_vmlinux options, it requires building twice. That results in
>6m build times. Hide this behind a flag that can be enabled in CI.

We need to first enable this new flag by default to not break the
existing CI testing. Once the CI is setup with the flag, a follow up
change will set the flag to False by default.

Bug: 385170658
Change-Id: I0dc87aefece84d936864041882a0bd4824279b5f
Signed-off-by: Will McVicker <willmcvicker@google.com>