Autoupdate of 2 submodules
*build/kernel updated to bada068157e1f74f54a127f23662102b64b7c3be
rust: Add rustc rpath to PROCMACROLDFLAGS

Proc macros are currently built without an explicit runpath. They have a
dependency on `libunwind.so` which is not on the default search path for
kleaf. `rustc-1.81` depended on `libunwind.so` directly, so it was
preloaded when it went to `dlopen` proc macros. The `rustc-1.82` binary
doesn't have this dependency anymore, so we need to adjust the runpath
to ensure that proc macros can find `libunwind.so`.

Bug: 384734194
Test: With 1.82, tools/bazel build //common:kernel_x86_64 --rust
Change-Id: I1665e65fb998df1370fa836cf380f7bbfc22a318

*common updated to 10ce996d75e57c64554365122588d1aba9608cf2
ANDROID: Update Rust compiler to 1.82.0

We're doing this here directly rather than backporting from mainline
because mainline still doesn't have PROCMACROLDFLAGS, which is needed to
be able to take this update.

Bug: 369422063
Bug: 359429865
Test: With 1.82, tools/bazel build //common:kernel_x86_64 --rust
Change-Id: I519d2055f100845c83569f7610327438f3dad7cf
Signed-off-by: Matthew Maurer <mmaurer@google.com>