kleaf: Optimize out rsync OUT_DIR/include in ddk_module_config main action.
Optimized out the duplicated rsync out_dir/include's in
a ddk_module_config that inherits from a parent.

Before:

    parent:
    0.7s kernel_build out_dir/include -> $OUT_DIR/include
    0.8s common -> parent kconfig_ext
    0.7s $OUT_DIR/include -> <parent out_dir>/include

    child:
    0.7s kernel_build out_dir/include -> $OUT_DIR/include
    0.6s parent/kconfig_ext child/kconfig_ext
    0.8s parent/out_dir/include -> $OUT_DIR/include
    0.9s $OUT_DIR/include -> child out_dir/include

After:

    parent:
    0.8s common -> parent kconfig_ext
    0.7s $OUT_DIR/include -> <parent out_dir>/include

    child:
    0.6s parent/kconfig_ext child/kconfig_ext
    0.7s parent/out_dir/inclde -> child out_dir/include

This reduces the build time of parent ddk_config + child ddk_module_config
from 7+5=12s to 6+3=9s. In addition, assuming that there will be a
single parent but a lot of children ddk_module_configs, this greatly
reduces the contention of sandbox I/Os, so the second saving grows
linearly with the number of ddk_module's sharing the same ddk_config.

As a side-effect, this change skips replacing the fake $ROOT_DIR
variable in auto.conf.cmd from kernel_build before using it in ddk_config.
This is fine because the variable is only used as a check against
$(srctree) to determine if auto.conf.cmd needs to be rebuilt or not.
However, we always rebuild auto.conf.cmd; see b/263415662 and
647a2e4b38e9 ("kleaf: Improve hack for fixdep."). So there are no real
effect because of this change. The reproducibility issue of
auto.conf.cmd is tracked in b/386268853.

Bug: 370681291
Bug: 386268853
Change-Id: Ia361ec6d2455cf33697193677ed7dd4410042ea3