ANDROID: Move BRANCH/KMI_GENERATION from build.config.common to .constants.
We are deprecating build.config.common so we have
one less build config for build-config-less builds. Even though the
value of `BRANCH` and `KMI_GENERATION` is technically not needed in
Bazel's analysis phase,
we still move it back to build.config.constants so we have a central
place to put branch-specific variables.

Bug: 276493276
Bug: 236012223
Change-Id: I0d63399af44e2a4da379fa224f4e53e25052207a
Signed-off-by: HONG Yifan <elsk@google.com>