UPSTREAM: Revert "PCI: Use preserve_config in place of pci_flags"
This reverts commit f858b0fab28d8bc2d0f0e8cd4afc3216f347cfcc which is
commit 7246a4520b4bf1494d7d030166a11b5226f6d508 upstream.

This patch causes a regression in cuttlefish/crossvm boot on arm64.

The patch was part of a series that when applied will not cause a regression
but this patch was backported to the 6.1 branch by itself.

The other patches do not apply cleanly to the 6.1 branch.

Change-Id: Ic43827b09479ce94cba23f9621f1afefa743cf91
Signed-off-by: Terry Tritton <terry.tritton@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 479a42eedbaf777e7cc1118448c3835f6e606bd8)
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>