ANDROID: add binder.impl kernel parameter
This introduces a new binder.impl parameter for the kernel commandline.
The parameter controls whether the device should use the C or Rust
implementation of Binder.

This patch does not contain the Rust portion of this logic as Rust
Binder will be a DDK module for this kernel version. binder_use_rust is
exported so that the module can access it when loaded to check whether
the driver should be enabled. Note that Rust Binder must be a
vendor-boot module for the init sequence to work right.

The parameter has already been hooked up to Android userspace. This
means that the commands

	adb shell setprop kcmdline.binder c
	adb shell setprop kcmdline.binder rust

will cause Android to set binder.impl on the kernel commandline next
time the phone is rebooted.

Since it's infeasible to change the Binder driver used by a running
system, this parameter cannot be changed once a Binder driver is loaded.

Bug: 388786466
Change-Id: I2ea11cd4110ee86f88ca606c55bb9cc2924d576a
Signed-off-by: Alice Ryhl <aliceryhl@google.com>
Signed-off-by: Carlos Llamas <cmllamas@google.com>