ANDROID: README.md: document FROMLIST not being allowed for rejected changes
Recently, multiple patch series have been submitted to Android kernels
with the FROMLIST: tag, despite them containing changes that have been
rejected by the upstream community, sometimes many years ago.

FROMLIST is to be used for things that have a chance to be accepted
upstream, or for work-in-progress changes, NOT for rejected submissions,
as using this tag should not hide this rejection state.

Bug: 391973735
Change-Id: I2e73d43dd780c01e9cfd6d1f249c1c9eebcf1a3d
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>