ANDROID: KVM: Fix missing_donations for pKVM heap alloc
Because having to top-up the hypervisor heap alloc with more than 255
pages is a rare endeavour and to space a bit of space, the per-CPU
missing donation variable is capped to 8-bits.

Sadly a broken macro lead to that variable being truncated after 8-bits
and not limited to... which means 256 missing donations would end-up
being 0 for the host.

Bug: 357781595
Bug: 273748186
Fixes: f618b4931ab3 ("ANDROID: KVM: arm64: Allow to read missing donations for pKVM heap alloc")
Change-Id: I170f78c4ede2b0f3251d3077768d2a9a6ffddeb4
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>