ANDROID: GKI: Enable USB4/Thunderbolt as GKI module
When initializing the ACPI bus (which is a built-in part of a GKI)
the OSC_SB_NATIVE_USB4_SUPPORT is negotiated only if CONFIG_USB4
is enabled. This means that USB4 support needs to be enabled as part
of gki_defconfig (built-in or module) to ensure proper negotation
of OSC capabilities.

The general problem with IS_ENABLED() was already described here
b/259510298 and the recommendations are to either convert modules
to be GKI modules or refactor the code to avoid IS_ENABLED() usage.
Since USB4/Thunderbolt is commonly used platform independent module
enabling it as a GKI module is reasonable.

Bug: 394566441
Change-Id: Iae5bb836f0574bac35274c67282f8efd439c884c
Signed-off-by: Slawomir Rosek <srosek@google.com>